### 2025-04-06 #8001 @editorialbot
[PRE REVIEW]: NeuroCAPs: A Python Package for Performing Co-Activation Patterns Analyses on Resting-State and Task-Based fMRI Data
**Submitting author:** @donishadsmith (<a href="http://orcid.org/0000-0001-7019-3520">Donisha Smith</a>)
**Repository:** https://github.com/donishadsmith/neurocaps
**Branch with paper.md** (empty if default branch):
**Version:** 0.23.8.post1
