### 2025-04-06 #8001 @donishadsmith
Hello!

If possible, I was wondering if the latest version (0.24.7) could be considered for review? The software article that was already automatically generated includes recently added features.  In addition, that version includes the latest refactorings and documentation updates that hopefully allows the review process to go more smoothly. However, 0.23.8.post1 is tagged on my repo if it is not possible to change the version for review.

Thank you!
