+++ paper/paper.bib
@@ -188,7 +188,7 @@
     license = {BSD-4-Clause},
     title = {{nilearn}},
     url = {https://github.com/nilearn/nilearn},
-    doi = {https://doi.org/10.5281/zenodo.8397156}
+    doi = {10.5281/zenodo.8397156}
 }
 
 @article{scikit-learn,
