### 2025-04-06 #8001 @donishadsmith
Suggested reviewers for this submission considering their expertise in fMRI methodology, data denoising, functional connectivity, and Python-based neuroimaging pipelines:

- Dr. Hao-Ting Wang (htwangtw)
- Dr. Stephan Heunis (jsheunis)
- Dr. Azeez Adebimpe (a3sha2)
- Dr. Camille Maumet (cmaumet)
