### 2025-04-07 4f42914
Add tqdm for concatenation
