### 2025-04-07 8c177e8
Add tqdm for concatenation
