+++ tests/test_TimeseriesExtractor.py
@@ -2,6 +2,7 @@ import copy, glob, os, re, shutil, sys
 import joblib, pytest, numpy as np, pandas as pd
 
 from neurocaps.extraction import TimeseriesExtractor
+from neurocaps._utils import _standardize
 
 from .utils import (
     Parcellation,
@@ -666,9 +667,9 @@ def test_condition_extraction(get_vars):
     active_condition = copy.deepcopy(extractor.subject_timeseries["01"]["run-001"])
 
     scan_list = get_scans(bids_dir, "rest")
-    assert np.array_equal(timeseries[scan_list, :], rest_condition)
+    assert np.allclose(_standardize(timeseries[scan_list, :]), rest_condition, atol=0.00001)
     scan_list = get_scans(bids_dir, "active")
-    assert np.array_equal(timeseries[scan_list, :], active_condition)
+    assert np.allclose(_standardize(timeseries[scan_list, :]), active_condition, atol=0.00001)
 
 
 @pytest.mark.parametrize("confound_type", [None, "testing_confounds"])
@@ -929,8 +930,15 @@ def test_fd_censoring(get_vars):
     assert extractor_fd_float.qc["01"]["run-001"]["mean_high_motion_length"] == 1
     assert extractor_fd_float.qc["01"]["run-001"]["std_high_motion_length"] == 0
 
-    assert np.array_equal(
-        extractor_fd_dict.subject_timeseries["01"]["run-001"], extractor_fd_float.subject_timeseries["01"]["run-001"]
+    # Get non censored
+    extractor_no_censor = TimeseriesExtractor()
+    extractor_no_censor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
+
+    assert np.allclose(
+        extractor_fd_dict.subject_timeseries["01"]["run-001"],
+        extractor_fd_float.subject_timeseries["01"]["run-001"],
+        _standardize(np.delete(extractor_no_censor.subject_timeseries["01"]["run-001"], 39, axis=0)),
+        atol=0.00001,
     )
     # Test Conditions only "rest" condition should have scan censored
     no_condition_timeseries = copy.deepcopy(extractor_fd_dict.subject_timeseries["01"]["run-001"])
@@ -941,7 +949,11 @@ def test_fd_censoring(get_vars):
     assert extractor_fd_dict.qc["01"]["run-001"]["frames_interpolated"] == 0
     assert extractor_fd_dict.qc["01"]["run-001"]["mean_high_motion_length"] == 1
     assert extractor_fd_dict.qc["01"]["run-001"]["std_high_motion_length"] == 0
-    assert np.array_equal(no_condition_timeseries[scan_list, :], extractor_fd_dict.subject_timeseries["01"]["run-001"])
+    assert np.allclose(
+        _standardize(no_condition_timeseries[scan_list, :]),
+        extractor_fd_dict.subject_timeseries["01"]["run-001"],
+        atol=0.00001,
+    )
 
     scan_list = get_scans(bids_dir, "active", fd=True)
     extractor_fd_dict.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2, condition="active")
@@ -950,7 +962,11 @@ def test_fd_censoring(get_vars):
     assert extractor_fd_dict.qc["01"]["run-001"]["frames_interpolated"] == 0
     assert extractor_fd_dict.qc["01"]["run-001"]["mean_high_motion_length"] == 0
     assert extractor_fd_dict.qc["01"]["run-001"]["std_high_motion_length"] == 0
[…]
