+++ docs/examples/analysis.rst
@@ -45,8 +45,9 @@ in the ``CAP`` class and this information is used by all methods in the class. T
 
     .. code-block:: none
 
+        Concatenating Subjects [GROUP: A]: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:01<00:00, 668.15it/s]
         Clustering [GROUP: All Subjects]: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 9/9 [00:00<00:00, 20.38it/s]
-        2025-03-09 08:20:26,819 neurocaps.analysis.cap [INFO] [GROUP: All Subjects | METHOD: elbow] Optimal cluster size is 5.
+        2025-04-07 18:15:07,367 neurocaps.analysis.cap [INFO] [GROUP: All Subjects | METHOD: elbow] Optimal cluster size is 5.
 
 .. image:: embed/All_Subjects_elbow.png
     :width: 600
@@ -95,8 +96,15 @@ Performing CAPs on Groups
 
     .. code-block:: none
 
+        Concatenating Subjects [GROUP: A]: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 4/4 [00:01<00:00, 582.04it/s]
+        Concatenating Subjects [GROUP: A]: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 6/6 [00:01<00:00, 706.37it/s]
+
+.. rst-class:: sphx-glr-script-out
+
+    .. code-block:: none
+
         Clustering [GROUP: A]: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 19/19 [00:01<00:00, 18.71it/s]
-        2025-01-31 13:29:54,234 neurocaps.analysis.cap [INFO] [GROUP: A | METHOD: silhouette] Optimal cluster size is 2.
+        2025-04-07 18:15:53,981 neurocaps.analysis.cap [INFO] [GROUP: A | METHOD: silhouette] Optimal cluster size is 2.
 
 .. image:: embed/A_silhouette.png
     :width: 600
@@ -106,7 +114,7 @@ Performing CAPs on Groups
     .. code-block:: none
 
         Clustering [GROUP: B]: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 19/19 [00:01<00:00, 12.48it/s]
-        2025-01-31 13:29:57,757 neurocaps.analysis.cap [INFO] [GROUP: B | METHOD: silhouette] Optimal cluster size is 2.
+        2025-04-07 18:15:55,236 neurocaps.analysis.cap [INFO] [GROUP: B | METHOD: silhouette] Optimal cluster size is 2.
 
 .. image:: embed/B_silhouette.png
     :width: 600

[…]
