+++ paper/paper.bib
@@ -4,8 +4,8 @@
   month        = apr,
   year         = {2025},
   publisher    = {Zenodo},
-  doi          = {10.5281/zenodo.15127766},
-  url          = {https://doi.org/10.5281/zenodo.15127766}
+  doi          = {10.5281/zenodo.15171184},
+  url          = {https://doi.org/10.5281/zenodo.15171184}
 }
 
 @article{Jiang2022,
