### 2025-04-08 6048822
Minor code cleaning
