+++ README.md
@@ -203,7 +203,7 @@ extractor = TimeseriesExtractor(
     parcel_approach=parcel_approach,
     standardize=True,
     use_confounds=True,
-    detrend=True,
+    detrend=False,
     low_pass=0.1,
     high_pass=None,
     n_acompcor_separate=2,
