### 2025-04-09 fe2fcdf
Change detrending and plotting default
