### 2025-04-09 f98b267
Remove directives under 0.25.0
