+++ neurocaps/extraction/timeseriesextractor.py
@@ -92,9 +92,6 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
             - Confound names follow fMRIPrep's naming scheme (versions >= 1.2.0).
             - Wildcards are supported: e.g., "cosine*" matches all confounds starting with "cosine".
 
-        .. versionchanged:: 0.23.0 Changed default from ``None`` to ``"basic"``. The ``"basic"`` option provides the
-           same functionality that ``None`` did in previous versions.
-
     fd_threshold: :obj:`float`, :obj:`dict[str, float | int]`, or :obj:`None`, default=None
         Threshold for volume censoring based on framewise displacement (FD). Computed only after dummy volumes are
         removed.
@@ -136,8 +133,6 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
                 censored high motion volumes and "1" indicates non-censored, low motion volumes, only the volumes at\
                 index 3, 5, 6, 7, and 9 would be interpolated.
 
-            .. versionadded:: 0.22.3 "interpolate" key added.
-
         .. important::
             - A column named "framewise_displacement" must be available in the confounds file.
             - ``use_confounds`` must be set to True.
@@ -178,8 +173,6 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
             be present in the confounds tsv file.
             - "min" and "max" keys only apply when "auto" is True.
 
-        .. versionadded:: 0.24.7 Can now use "auto" instead of ``{"auto": True}`` for automatic dummy volume detection.
-
     dtype: :obj:`str` or "auto", default=None
         The NumPy dtype to convert NIfTI images to.
 
@@ -218,8 +211,6 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
         ::
 
             {"subjectID": {"run-ID": {"frames_scrubbed": int, "frames_interpolated": int, "mean_high_motion_length": float, "std_high_motion_length": float}}}
-
-        .. versionadded:: 0.24.3
 
     See Also
     --------
@@ -434,8 +425,6 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
             about how this offset affects the calculation of task conditions, see the "Extraction of Task Conditions"
             section below.
 
-            .. versionadded:: 0.20.0
-
         tr: :obj:`int`, :obj:`float` or :obj:`None`, default=None
             Repetition time (TR), in seconds, for the specified task. If not provided, the TR will be automatically
             extracted from the first BOLD metadata file found for the task, searching first in the pipeline directory,
@@ -446,25 +435,17 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
             to adjust for slice time correction when ``condition`` is not None. Values can range from 0 to 1. For more
             details, see the "Extraction of Task Conditions" section below.
 
-            .. versionadded:: 0.21.0
-
         run_subjects: :obj:`str`, :obj:`list[str]` or :obj:`None`, default=None
             A string (if single subject) or list of subject IDs to process (e.g. ``run_subjects=["01", "02"]``).
             Processes all subjects if None.
 
-            .. versionchanged:: 0.24.0 A string can be used if specifying a single subject.
-
         exclude_subjects: :obj:`str`, :obj:`list[str]` or :obj:`None`, default=None
             A string (if single subject) or list of subject IDs to exclude (e.g. ``exclude_subjects=["01", "02"]``).
-
-            .. versionchanged:: 0.24.0 A string can be used if specifying a single subject.
 
         exclude_niftis: :obj:`str`, :obj:`list[str]` or :obj:`None`, default=None
             A string (if single file) or List of the specific preprocessed NIfTI files to exclude, preventing their
[…]
