### 2025-04-09 f6f7b99
Update tests and docs
- Set detrend True for certain tests due to minor floating point differences. Done to not reduce atol.
