+++ tests/test_TimeseriesExtractor.py
@@ -647,7 +647,7 @@ def test_condition_extraction(get_vars):
     """
     bids_dir, pipeline_name = get_vars
 
-    extractor = TimeseriesExtractor()
+    extractor = TimeseriesExtractor(detrend=True)
 
     extractor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
     timeseries = copy.deepcopy(extractor.subject_timeseries["01"]["run-001"])
@@ -874,7 +874,7 @@ def test_fd_censoring(get_vars):
     bids_dir, pipeline_name = get_vars
 
     # Should censor; use_confounds is True by default
-    extractor_censor = TimeseriesExtractor(fd_threshold=0.35)
+    extractor_censor = TimeseriesExtractor(fd_threshold=0.35, detrend=True)
 
     extractor_censor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
     assert extractor_censor.subject_timeseries["01"]["run-001"].shape == (39, 400)
@@ -885,7 +885,7 @@ def test_fd_censoring(get_vars):
 
     # Check "outlier_percentage"
     extractor_low_outlier_threshold = TimeseriesExtractor(
-        fd_threshold={"threshold": 0.35, "outlier_percentage": 0.0001},
+        fd_threshold={"threshold": 0.35, "outlier_percentage": 0.0001}, detrend=True
     )
 
     extractor_low_outlier_threshold.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
@@ -911,9 +911,7 @@ def test_fd_censoring(get_vars):
     assert extractor_low_outlier_threshold.qc == {}
 
     # Test that dictionary fd_threshold and float are the same
-    extractor_fd_dict = TimeseriesExtractor(
-        fd_threshold={"threshold": 0.35, "outlier_percentage": 0.30},
-    )
+    extractor_fd_dict = TimeseriesExtractor(fd_threshold={"threshold": 0.35, "outlier_percentage": 0.30}, detrend=True)
 
     extractor_fd_dict.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
     assert extractor_fd_dict.subject_timeseries["01"]["run-001"].shape == (39, 400)
@@ -922,7 +920,7 @@ def test_fd_censoring(get_vars):
     assert extractor_fd_dict.qc["01"]["run-001"]["mean_high_motion_length"] == 1
     assert extractor_fd_dict.qc["01"]["run-001"]["std_high_motion_length"] == 0
 
-    extractor_fd_float = TimeseriesExtractor(fd_threshold=0.35)
+    extractor_fd_float = TimeseriesExtractor(fd_threshold=0.35, detrend=True)
 
     extractor_fd_float.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
     assert extractor_fd_float.qc["01"]["run-001"]["frames_scrubbed"] == 1
@@ -931,7 +929,7 @@ def test_fd_censoring(get_vars):
     assert extractor_fd_float.qc["01"]["run-001"]["std_high_motion_length"] == 0
 
     # Get non censored
-    extractor_no_censor = TimeseriesExtractor()
+    extractor_no_censor = TimeseriesExtractor(detrend=True)
     extractor_no_censor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
 
     assert np.allclose(
@@ -1028,13 +1026,13 @@ def test_censoring_with_sample_mask(get_vars):
 
     # First get no condition data
[…]
