### 2025-04-09 778d7b4
Remove mask file in data
