### 2025-04-09 0d6ee59
Add mean and std fd and place them first in qc report
