### 2025-04-10 0c58136
Account for additional edge case
