### 2025-04-10 d8db1e2
[skip ci] Minor doc update to properties and remove already addressed colab warning
