### 2025-04-11 7713026
Add colon in log message and update test
- Update old test comments
