+++ tests/test_TimeseriesExtractor.py
@@ -287,7 +287,7 @@ def test_default_confounds():
     assert extractor.signal_clean_info["confound_names"] is None
 
 
-def test_delete_property(get_vars):
+def test_delete_property(get_vars, caplog):
     """
     Ensures deleter property works.
     """
@@ -424,9 +424,10 @@ def test_check_raise_error():
         (Parcellation.get_custom("parcellation"), False, "Custom"),
     ],
 )
-@pytest.mark.enable_logs
-def test_basic_extraction(get_vars, parcel_approach, use_confounds, name):
+def test_basic_extraction(get_vars, parcel_approach, use_confounds, name, caplog):
     """Tests basic extraction/basic use case."""
+    import logging
+
     bids_dir, pipeline_name = get_vars
 
     shape_dict = {"Schaefer": 100, "AAL": 116, "Custom": 426}
@@ -434,7 +435,7 @@ def test_basic_extraction(get_vars, parcel_approach, use_confounds, name):
 
     extractor = TimeseriesExtractor(
         parcel_approach=parcel_approach,
-        standardize="zscore_sample",
+        standardize=True,
         use_confounds=use_confounds,
         detrend=True,
         low_pass=0.15,
@@ -446,6 +447,21 @@ def test_basic_extraction(get_vars, parcel_approach, use_confounds, name):
     print(extractor)
 
     extractor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
+
+    with caplog.at_level(logging.INFO):
+        extractor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
+
+    assert "Preparing for Timeseries Extraction using [FILE:" in caplog.text
+
+    msg = (
+        "The following confounds will be used for nuisance regression: cosine_00, cosine_01, cosine_02, cosine_03, "
+        "cosine_04, cosine_05, cosine_06, rot_x, rot_y, rot_z."
+    )
+
+    if use_confounds:
+        assert msg in caplog.text
+    else:
+        assert not msg in caplog.text
 
     # No error; Testing __str__
     print(extractor)
@@ -738,11 +754,13 @@ def test_confounds(get_vars, confound_type):
         all(i in returned_confounds for i in correct_confounds)
 
 
-def test_acompcor_separate(get_vars):
+def test_acompcor_separate(get_vars, caplog):
     """
-    Ensures acompcor componenets specified in `confound_names` are removed due to `n_acompcor_separate`
+    Ensures acompcor components specified in `confound_names` are removed due to `n_acompcor_separate`
     being specified.
     """
+    import logging
+
     bids_dir, pipeline_name = get_vars
 
     confounds = ["cosine*", "rot*", "a_comp_cor_01", "a_comp_cor_02", "a_comp_cor*"]
@@ -750,10 +768,27 @@ def test_acompcor_separate(get_vars):
     extractor_with_confounds = TimeseriesExtractor(
         use_confounds=True,
         confound_names=confounds,
-        n_acompcor_separate=3,
+        n_acompcor_separate=1,
     )
 
-    extractor_with_confounds.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
+    with caplog.at_level(logging.INFO):
+        extractor_with_confounds.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
+
+    # Possibly change language when only 1 component used
+    msg1 = (
+        "Since `n_acompcor_separate` has been specified, acompcor components in `confound_names` will be "
[…]
