### 2025-04-11 8caf211
Optimize performance for concatenating data
