+++ neurocaps/analysis/cap.py
@@ -1,6 +1,7 @@
 """Contains the CAP class for performing co-activation patterns analyses"""
 
-import collections, copy, itertools, os, re, sys, tempfile
+import collections, itertools, os, re, sys, tempfile
+from operator import itemgetter
 from typing import Callable, Literal, Optional, Union
 
 # Conditional import based on major and minor version of Python
@@ -418,10 +419,6 @@ class CAP(_CAPGetter):
     def _process_subject_timeseries(subject_timeseries):
         if isinstance(subject_timeseries, str) and subject_timeseries.endswith(".pkl"):
             subject_timeseries = _convert_pickle_to_dict(subject_timeseries)
-        elif isinstance(subject_timeseries, dict) and len(list(subject_timeseries)) == 1:
-            # Potential mutability issue if only a single subject in dictionary potentially due to the variable
-            # for the concatenated data being set to the subject timeseries if concatenated data is empty.
-            subject_timeseries = copy.deepcopy(subject_timeseries)
 
         return subject_timeseries
 
@@ -444,24 +441,21 @@ class CAP(_CAPGetter):
         if not self._groups:
             self._groups = {"All Subjects": list(subject_timeseries)}
 
-        # Sort Ids lexicographically (also done in `TimeseriesExtractor`)
+        # Sort IDs lexicographically (also done in `TimeseriesExtractor`)
         self._groups = {group: sorted(self._groups[group]) for group in self._groups}
 
-        concatenated_timeseries = {group: None for group in self._groups}
-
         self._generate_lookup_table()
-
-        self._mean_vec = {group: None for group in self._groups}
-        self._stdev_vec = {group: None for group in self._groups}
 
         # Intersect subjects in subjects table and the groups for tqdm
         group_map = {
             group: sorted(set(self._subject_table).intersection(self._groups[group])) for group in self._groups
         }
 
+        # Collect timeseries data in lists
+        group_arrays = {group: [] for group in self._groups}
         for group in group_map:
             for subj_id in tqdm(
-                group_map[group], desc=f"Concatenating Subjects [GROUP: {group}]", disable=not progress_bar
+                group_map[group], desc=f"Collecting Subject Timeseries Data [GROUP: {group}]", disable=not progress_bar
             ):
                 subject_runs, miss_runs = self._get_runs(runs, list(subject_timeseries[subj_id]))
 
@@ -472,15 +466,24 @@ class CAP(_CAPGetter):
                     LG.warning(f"[SUBJECT: {subj_id}] Excluded from the concatenated timeseries due to having no runs.")
                     continue
 
-                for curr_run in subject_runs:
-                    if concatenated_timeseries[group] is None:
-                        concatenated_timeseries[group] = subject_timeseries[subj_id][curr_run]
-                    else:
-                        concatenated_timeseries[group] = np.vstack(
-                            [concatenated_timeseries[group], subject_timeseries[subj_id][curr_run]]
-                        )
+                subj_arrays = itemgetter(*subject_runs)(subject_timeseries[subj_id])
+
+                if len(subject_runs) == 1:
+                    subj_arrays = [subj_arrays]
+
+                group_arrays[group].extend(subj_arrays)
+
+        # Only stack once per group; avoid bottleneck due to repeated calls on large data
+        concatenated_timeseries = {group: None for group in self._groups}
+        for group in tqdm(group_arrays, desc="Concatenating Timeseries Data Per Group", disable=not progress_bar):
+            concatenated_timeseries[group] = np.vstack(group_arrays[group])
+
+        del group_arrays
 
         if self._standardize:
+            self._mean_vec = {group: None for group in self._groups}
+            self._stdev_vec = {group: None for group in self._groups}
+
[…]
