+++ CHANGELOG.md
@@ -13,6 +13,11 @@ noted in the changelog (e.g., new functions or parameters, changes in parameter 
 improvements/enhancements. All fixes and modifications are backwards compatible.
 - *.postN* : Consists of documentation changes or metadata-related updates, such as modifications to type hints.
 
+## [0.26.1] - 2025-04-11
+### 🐛 Fixes
+- Fixed performance bottleneck when stacking large timeseries by only calling NumPy's `vstack` once per group
+instead of once per subject and run pair.
+
 ## [0.26.0] - 2025-04-10
 - Full source repository now archived on Zenodo, instead of just the pure source code
 ### 🚀 New/Added
