### 2025-04-11 1eb0287
Refactor to validate first before assignment
