### 2025-04-11 780aa1b
Add explicit sorted to test
- Also add additional test assert
