+++ paper/paper.bib
@@ -4,8 +4,8 @@
   month        = apr,
   year         = {2025},
   publisher    = {Zenodo},
-  doi          = {10.5281/zenodo.15186615},
-  url          = {https://doi.org/10.5281/zenodo.15186615}
+  doi          = {10.5281/zenodo.15200608},
+  url          = {https://doi.org/10.5281/zenodo.15200608}
 }
 
 @article{Jiang2022,
