### 2025-04-12 e127542
Updates for upcoming nilearn changes & Minimum dependency fix
- Decoding of the Schaefer label no longer needs to be done for versions
  greater than 0.11.1
  so try-except block added to handle this
- Removes the added "Background" label from the AAL and Schaefer
- Uses clean_args parameter for NiftiLabelsMasker if version greater
  than 0.11.1
- Changes made to be flexible to accomodate above and below 0.11.1
- copy_header parameter in resample_to_img was added in 0.11.1
- Increase minimum dependencies
- Fix tests to run with minimum and latest dependencies
