+++ paper/paper.bib
@@ -4,8 +4,8 @@
   month        = apr,
   year         = {2025},
   publisher    = {Zenodo},
-  doi          = {10.5281/zenodo.15200608},
-  url          = {https://doi.org/10.5281/zenodo.15200608}
+  doi          = {10.5281/zenodo.15205287},
+  url          = {https://doi.org/10.5281/zenodo.15205287}
 }
 
 @article{Jiang2022,
