+++ CHANGELOG.md
@@ -13,6 +13,10 @@ noted in the changelog (e.g., new functions or parameters, changes in parameter 
 improvements/enhancements. All fixes and modifications are backwards compatible.
 - *.postN* : Consists of documentation changes or metadata-related updates, such as modifications to type hints.
 
+## [0.26.4] - 2025-04-13
+### 🐛 Fixes
+- Fix for upcoming use of clean_args in ``NiftiLabelsMasker``.
+
 ## [0.26.3] - 2025-04-13
 - Dependency fixes
 ### 🐛 Fixes
