### 2025-04-13 92e129c
Expand supported file extensions and allow pickling of figures
