+++ paper/paper.bib
@@ -4,8 +4,8 @@
   month        = apr,
   year         = {2025},
   publisher    = {Zenodo},
-  doi          = {10.5281/zenodo.15205697},
-  url          = {https://doi.org/10.5281/zenodo.15205697}
+  doi          = {10.5281/zenodo.15208111},
+  url          = {https://doi.org/10.5281/zenodo.15208111}
 }
 
 @article{Jiang2022,
