+++ neurocaps/analysis/change_dtype.py
@@ -5,8 +5,8 @@ from typing import Union, Optional
 
 import numpy as np
 
-from .._utils import _convert_pickle_to_dict, _dicts_to_pickles, _logger
 from ..typing import SubjectTimeseries
+from .._utils import _convert_pickle_to_dict, _dicts_to_pickles, _logger
 
 LG = _logger(__name__)
 
@@ -72,7 +72,7 @@ def change_dtype(
     changed_dtype_dicts = {}
 
     for indx, curr_dict in enumerate(subject_timeseries_list):
-        if isinstance(curr_dict, str) and curr_dict.endswith(".pkl"):
+        if isinstance(curr_dict, str):
             curr_dict = _convert_pickle_to_dict(curr_dict)
         else:
             curr_dict = copy.deepcopy(curr_dict)
