+++ neurocaps/analysis/cap.py
@@ -32,6 +32,7 @@ from .._utils import (
     _check_parcel_approach,
     _collapse_aal_node_names,
     _logger,
+    _pickle_object,
     _run_kmeans,
     _save_contents,
     _standardize,
@@ -244,6 +245,7 @@ class CAP(_CAPGetter):
         show_figs: bool = False,
         output_dir: Optional[str] = None,
         progress_bar: bool = False,
+        as_pickle: bool = False,
         **kwargs,
     ) -> Self:
         """
@@ -308,6 +310,12 @@ class CAP(_CAPGetter):
 
         progress_bar: :obj:`bool`, default=False
             If True and ``cluster_selection_method`` is not None, displays a progress bar.
+
+        as_pickle: :obj:`bool`, default=False
+            When ``output_dir`` and ``cluster_selection_method`` is specified, plots are saved as pickle filess, which
+            can be further modified, instead of png images.
+
+            .. versionadded:: 0.26.5
 
         **kwargs:
             Additional keyword arguments when ``cluster_selection_method`` is specified:
@@ -401,7 +409,7 @@ class CAP(_CAPGetter):
         self._concatenated_timeseries = self._concatenate_timeseries(subject_timeseries, runs, progress_bar)
 
         if self._cluster_selection_method:
-            self._select_optimal_clusters(configs, show_figs, output_dir, progress_bar, **kwargs)
+            self._select_optimal_clusters(configs, show_figs, output_dir, progress_bar, as_pickle, **kwargs)
         else:
             self._kmeans = {}
             for group in self._groups:
@@ -417,7 +425,7 @@ class CAP(_CAPGetter):
 
     @staticmethod
     def _process_subject_timeseries(subject_timeseries):
-        if isinstance(subject_timeseries, str) and subject_timeseries.endswith(".pkl"):
+        if isinstance(subject_timeseries, str):
             subject_timeseries = _convert_pickle_to_dict(subject_timeseries)
 
         return subject_timeseries
@@ -506,7 +514,7 @@ class CAP(_CAPGetter):
 
         return runs, miss_runs
 
-    def _select_optimal_clusters(self, configs, show_figs, output_dir, progress_bar, **kwargs):
+    def _select_optimal_clusters(self, configs, show_figs, output_dir, progress_bar, as_pickle, **kwargs):
         self._cluster_scores = {}
         self._optimal_n_clusters = {}
         self._kmeans = {}
@@ -575,12 +583,12 @@ class CAP(_CAPGetter):
             LG.info(f"[GROUP: {group} | METHOD: {method}] Optimal cluster size is {self._optimal_n_clusters[group]}.")
 
             if show_figs or output_dir is not None:
-                self._plot_method(method, performance_dict, group, show_figs, output_dir, **kwargs)
+                self._plot_method(method, performance_dict, group, show_figs, output_dir, as_pickle, **kwargs)
 
         self._cluster_scores = {"Cluster_Selection_Method": method}
         self._cluster_scores.update({"Scores": performance_dict})
 
-    def _plot_method(self, method, performance_dict, group, show_figs, output_dir, **kwargs):
+    def _plot_method(self, method, performance_dict, group, show_figs, output_dir, as_pickle, **kwargs):
         y_titles = {
             "elbow": "Inertia",
             "davies_bouldin": "Davies Bouldin Score",
@@ -612,7 +620,13 @@ class CAP(_CAPGetter):
         if output_dir:
             self._makedir(output_dir)
             save_name = f"{group.replace(' ', '_')}_{self._cluster_selection_method}.png"
[…]
