### 2025-04-14 45af4b0
Update test and remove debug print in test
