+++ tests/test_transition_matrix.py
@@ -1,4 +1,3 @@
-import glob, os
 import numpy as np, pytest
 
 from neurocaps.analysis import CAP, transition_matrix
