### 2025-04-14 de2f334
Fix silent log message & add docs to __str__
- Improve docs
