+++ neurocaps/_utils/extraction/extract_timeseries.py
@@ -256,6 +256,7 @@ def _extract_timeseries(
         # Continue extraction if the continue keyword isn't hit when assessing the fd threshold or events
         # May still be skipped prior to extraction if there are more confound regressors than timeseries row
         timeseries, skip_run = _perform_extraction(data, LG)
+
         if skip_run:
             continue
 
@@ -265,6 +266,7 @@ def _extract_timeseries(
             )
         else:
             subject_timeseries[subj_id].update({run_id: timeseries})
+
             # Report framewise displacement quality control
             if data.produce_qc_report:
                 qc[subj_id].update({run_id: _report_qc(data)})
@@ -567,7 +569,7 @@ def filter_censored_scan_indices(data):
     return sorted(list(scans)), len(all_censored_indxs), n_interpolated_scans
 
 
-def _validate_scan_bounds(data, arr_len, LG=None, warn=False):
+def _validate_scan_bounds(data, arr_len, LG=None):
     """
     Checks if indices in ``data.scans`` (which are the indices for the event condition) are within the timeseries
     boundaries.
@@ -575,9 +577,9 @@ def _validate_scan_bounds(data, arr_len, LG=None, warn=False):
     if max(data.scans) <= arr_len - 1:
         return data.scans
 
-    # Warn is false in instances were shift is used to prevent unnecessary logging, the warning is specifically
-    # for instances related to misalignment such as potentially incorrect onsets and durations
-    if warn:
+    # Warn if tr_shift was not done, the warning is specifically for instances related to misalignment such as
+    # potentially incorrect onsets and durations
+    if not data.tr_shift:
         LG.warning(
             f"{data.head}" + f"[CONDITION: {data.condition}] Max scan index exceeds timeseries max index. "
             f"Max condition index is {max(data.scans)}, while max timeseries index is {arr_len - 1}. Timing may "
@@ -608,6 +610,12 @@ def _perform_extraction(data, LG):
     """
     # Extract confound information of interest and ensure confound file does not contain NAs
     confounds = _process_confounds(data, LG)
+
+    if data.use_confounds and (data.confound_names and confounds is None):
+        LG.warning(
+            f"{data.head}" + "None of the requested confounds were found. Nuisance regression will not be done but "
+            "timeseries extraction will continue."
+        )
 
     # Ensure number of confounds not greater than timeseries shape
     if skip_run := _check_dimensions(data, confounds, LG):
@@ -646,9 +654,9 @@ def _perform_extraction(data, LG):
     timeseries = _process_timeseries(timeseries, data) if data.censored_frames else timeseries
 
     if data.condition:
-        # Ensure condition indices don't exceed timeseries due to tr shift; filter_censored_scan_indices only called
-        # when censoring is done
-        data.scans = _validate_scan_bounds(data, timeseries.shape[0], warn=False)
+        # Ensure condition indices don't exceed timeseries due to tr shift; filter_censored_scan_indices only being
+        # called when censoring is done
+        data.scans = _validate_scan_bounds(data, timeseries.shape[0], LG)
 
         if data.verbose:
             LG.info(data.head + f"Nuisance regression completed; extracting [CONDITION: {data.condition}].")
@@ -665,22 +673,22 @@ def _process_confounds(data, LG):
     requested, then this function obtains a filtered version of the confound dataframe containing columns that will be
     used for nuisance regression.
     """
-    if not data.use_confounds:
+    if not data.use_confounds or (data.use_confounds and data.confound_df.empty):
         return None
-    else:
-        confound_names = copy.deepcopy(data.signal_clean_info["confound_names"]) if data.confound_names else []
 
[…]
