+++ tests/test_TimeseriesExtractor.py
@@ -288,7 +288,7 @@ def test_default_confounds():
     assert extractor.signal_clean_info["confound_names"] is None
 
 
-def test_delete_property(get_vars, caplog):
+def test_delete_property(get_vars):
     """
     Ensures deleter property works.
     """
@@ -447,8 +447,6 @@ def test_basic_extraction(get_vars, parcel_approach, use_confounds, name, caplog
     # No error; Testing __str__
     print(extractor)
 
-    extractor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
-
     with caplog.at_level(logging.INFO):
         extractor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
 
@@ -486,6 +484,38 @@ def test_basic_extraction(get_vars, parcel_approach, use_confounds, name, caplog
         condition="active",
     )
     assert extractor.subject_timeseries["01"]["run-001"].shape == (20, shape)
+
+
+def test_missing_confound_messages(get_vars, caplog):
+    """Test warnings issued for all or some confounds missing."""
+    import logging
+
+    bids_dir, pipeline_name = get_vars
+
+    with caplog.at_level(logging.WARNING):
+        # `use_confounds` default is True
+        extractor = TimeseriesExtractor(confound_names=["placeholder"])
+
+        extractor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name)
+
+        msg = (
+            "None of the requested confounds were found. Nuisance regression will not be done but timeseries "
+            "extraction will continue."
+        )
+        assert msg in caplog.text
+        caplog.clear()
+
+        # Should not produce warning
+        extractor = TimeseriesExtractor(use_confounds=False, confound_names=["placeholder"])
+        assert not msg in caplog.text
+        caplog.clear()
+
+        # Only one invalid confound case
+        extractor = TimeseriesExtractor(confound_names=["placeholder", "cosine*"])
+        extractor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name)
+
+        assert "The following confounds were not found: placeholder." in caplog.text
+        assert not msg in caplog.text
 
 
 def test_report_qc(tmp_dir, get_vars):
@@ -849,7 +879,7 @@ def test_n_acompcor_separate_error():
     """
 
     msg = (
-        "`n_acompcor_separate` specified `use_confounds` is not True, so separate WM and CSF components "
+        "`n_acompcor_separate` specified and `use_confounds` is not True, so separate WM and CSF components "
         "cannot be regressed out since confounds tsv file generated by fMRIPrep is needed."
     )
     with pytest.raises(ValueError, match=re.escape(msg)):
@@ -1553,30 +1583,30 @@ def test_skip_when_underdetermined_or_saturated(get_vars, caplog):
     with caplog.at_level(logging.WARNING):
         extractor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name)
 
-    assert "(n=40) is equal to or greater than the timeseries shape (n=40)." in caplog.text
-    assert not extractor.subject_timeseries
+        assert "(n=40) is equal to or greater than the timeseries shape (n=40)." in caplog.text
+        assert not extractor.subject_timeseries
 
-    reset_non_steady(bids_dir, pipeline_name)
+        reset_non_steady(bids_dir, pipeline_name)
 
-    # Should pass since censorind done after regression
-    add_non_steady(bids_dir, pipeline_name, 39)
-    extractor = TimeseriesExtractor(confound_names=["non_steady*"], fd_threshold=0.35)
[…]
