+++ docs/_templates/classmethod.rst
@@ -1,0 +1,6 @@
+{{ objname }}
+{{ underline }}====================
+
+.. currentmodule:: {{ module }}
+
+.. automethod:: {{ objname }}
