+++ docs/api/exceptions.rst
@@ -1,0 +1,14 @@
+:mod:`neurocaps.exceptions`
+---------------------------
+Module containing custom exceptions.
+
+.. currentmodule:: neurocaps.exceptions
+
+.. autosummary::
+   :template: exceptions.rst
+   :nosignatures:
+   :toctree: generated/
+
+   BIDSQueryError
+   NoElbowDetectedError
+   UnsupportedFileExtensionError
