+++ docs/api/extraction.rst
@@ -1,0 +1,14 @@
+:mod:`neurocaps.extraction`
+---------------------------
+Module containing the ``TimeseriesExtractor`` class for extracting timeseries data from preprocessed BIDS datasets
+(using pipelines such as fMRIPrep), reporting quality control metrics (number of frames scrubbed or interpolated) as
+well as for pickling (serializing) and visualizing the extracted data.
+
+.. currentmodule:: neurocaps.extraction
+
+.. autosummary::
+   :template: class.rst
+   :nosignatures:
+   :toctree: generated/
+
+   TimeseriesExtractor
