+++ docs/api/api.rst
@@ -1,0 +1,33 @@
+API
+===
+
+Publicly available classes and functions within NeuroCAPs.
+
+.. toctree::
+   :hidden:
+   :maxdepth: 1
+
+   exceptions
+   extraction
+   analysis
+   typing
+
+.. autosummary::
+   :nosignatures:
+
+   neurocaps.exceptions.BIDSQueryError
+   neurocaps.exceptions.NoElbowDetectedError
+   neurocaps.extraction.TimeseriesExtractor
+   neurocaps.analysis.CAP
+   neurocaps.analysis.change_dtype
+   neurocaps.analysis.merge_dicts
+   neurocaps.analysis.standardize
+   neurocaps.analysis.transition_matrix
+   neurocaps.typing.SubjectTimeseries
+   neurocaps.typing.ParcelConfig
+   neurocaps.typing.SchaeferParcelConfig
+   neurocaps.typing.AALParcelConfig
+   neurocaps.typing.ParcelApproach
+   neurocaps.typing.SchaeferParcelApproach
+   neurocaps.typing.AALParcelApproach
+   neurocaps.typing.CustomParcelApproach
