+++ docs/api/typing.rst
@@ -1,0 +1,20 @@
+:mod:`neurocaps.typing`
+-----------------------
+Module containing custom types.
+
+.. currentmodule:: neurocaps.typing
+
+.. autosummary::
+   :template: types.rst
+   :nosignatures:
+   :toctree: generated/
+
+   SubjectTimeseries
+   ParcelConfig
+   SchaeferParcelConfig
+   AALParcelConfig
+   ParcelApproach
+   SchaeferParcelApproach
+   AALParcelApproach
+   CustomParcelApproach
+   CustomRegionHemispheres
