### 2025-04-16 be76e42
Add py.typed & Update and format demos
