+++ paper/paper.bib
@@ -4,8 +4,8 @@
   month        = apr,
   year         = {2025},
   publisher    = {Zenodo},
-  doi          = {10.5281/zenodo.15221628},
-  url          = {https://doi.org/10.5281/zenodo.15221628}
+  doi          = {10.5281/zenodo.15232915},
+  url          = {https://doi.org/10.5281/zenodo.15232915}
 }
 
 @article{Jiang2022,
