### 2025-04-18 8373569
Message no longer mandatory for exception
