+++ paper/paper.bib
@@ -1,13 +1,3 @@
-@software{zenodo,
-  author       = {Smith, Donisha},
-  title        = {NeuroCAPs: A Python Package for Performing Co-Activation Patterns Analyses on Resting-State and Task-Based fMRI Data: Archive},
-  month        = apr,
-  year         = {2025},
-  publisher    = {Zenodo},
-  doi          = {10.5281/zenodo.15232915},
-  url          = {https://doi.org/10.5281/zenodo.15232915}
-}
-
 @article{Jiang2022,
     title={Time-varying dynamic network model for dynamic resting state functional connectivity in fMRI and MEG imaging},
     volume={254}, ISSN={10538119},
