+++ CHANGELOG.md
@@ -13,7 +13,7 @@ noted in the changelog (e.g., new functions or parameters, changes in parameter 
 improvements/enhancements. All fixes and modifications are backwards compatible.
 - *.postN* : Consists of documentation changes or metadata-related updates, such as modifications to type hints.
 
-## [0.26.7] - 2025-04-18
+## [0.26.8] - 2025-04-18
 ### ♻ Changed
 - Exceptions no longer require message
 ### 🐛 Fixes
