### 2025-04-19 da34301
Increment version
- Improvements to merge_dicts:
	- Reduce IO operations when str provided
	- Reduce calls to np.vstack
	- Fix issue with reduced dictionaries when str provided that only recognize pkl
	  files and were somewhat fragile since `'pkl' in` was used
	  instead of `.endswith('.pkl')`. Now several different file
	  extensions recognized
	- Remmoved dead code
