+++ paper/paper.md
@@ -81,7 +81,7 @@ Labeling [@Tzourio-Mazoyer2002], and Human Connectome Project extended [@Huang20
 dimensionality reduction.
 - censors high-motion volumes using fMRIPrep-derived framewise displacement values and stores the extracted timeseries
 information in a dictionary mapping subject IDs to run IDs and their associated timeseries.
-- reports quality control information for framewise displacement.
+- reports quality control information related to framewise displacement and dummy volumes.
 - saves extracted timeseries data as a pickle file.
 - visualizes timeseries data for a specific subject's run.
 
