+++ .github/workflows/testing.yaml
@@ -74,6 +74,6 @@ jobs:
 
       - name: Upload coverage reports to Codecov for Ubuntu
         if: ${{ matrix.os == 'ubuntu-latest' && matrix.python-ver == '3.12' }}
-        uses: codecov/codecov-action@v5.4.0
+        uses: codecov/codecov-action@v5.4.2
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
