### 2025-04-21 4da8937
Add some additional redundant unit tests
