+++ tests/test_TimeseriesExtractor.py
@@ -1,6 +1,7 @@
 import copy, glob, math, os, re, shutil, sys
 import joblib, pytest, numpy as np, pandas as pd
 
+import neurocaps._utils.extraction.extract_timeseries as et
 from neurocaps.extraction import TimeseriesExtractor
 from neurocaps._utils import _standardize
 
@@ -813,11 +814,10 @@ def test_confounds(get_vars, confound_type):
         )
 
     if confound_type == "testing_confounds":
-        from neurocaps._utils.extraction.extract_timeseries import _Data, _extract_valid_confounds
 
         confound_file = get_confound_data(bids_dir, pipeline_name)
 
-        data = _Data(files={"confound": confound_file}, verbose=False)
+        data = et._Data(files={"confound": confound_file}, verbose=False)
 
         correct_confounds = [
             "cosine_00",
@@ -832,7 +832,7 @@ def test_confounds(get_vars, confound_type):
             "rot_z",
         ]
 
-        returned_confounds = _extract_valid_confounds(data, confound_names, None)
+        returned_confounds = et._extract_valid_confounds(data, confound_names, None)
         assert isinstance(returned_confounds, pd.DataFrame)
         assert not returned_confounds.empty
         assert len(returned_confounds.columns) == len(correct_confounds)
@@ -888,17 +888,15 @@ def test_get_acompcor_separate(get_vars, n):
     "acompcor" components are extracted for nuisance regression when `n_acompcor_separate` are specified and not
     specified.
     """
-    from neurocaps._utils.extraction.extract_timeseries import _Data, _process_confounds
-
     bids_dir, pipeline_name = get_vars
     confound_file = get_confound_data(bids_dir, pipeline_name)
     confound_json = confound_file.replace(".tsv", ".json")
     correct_confounds = ["a_comp_cor_00", "a_comp_cor_01", "a_comp_cor_02", "a_comp_cor_03"]
 
-    data = _Data(files={"confound": confound_file, "confound_meta": confound_json}, verbose=False)
+    data = et._Data(files={"confound": confound_file, "confound_meta": confound_json}, verbose=False)
     data.signal_clean_info = {"n_acompcor_separate": n, "use_confounds": True, "confound_names": None}
 
-    confounds = _process_confounds(data, None)
+    confounds = et._process_confounds(data, None)
     assert isinstance(confounds, pd.DataFrame)
     assert len(correct_confounds) == len(confounds.columns) == 4
     assert confounds.shape == (40, 4)
@@ -906,14 +904,14 @@ def test_get_acompcor_separate(get_vars, n):
 
     data.signal_clean_info["confound_names"] = ["rot_x"]
     correct_confounds += ["rot_x"]
-    confounds = _process_confounds(data, None)
+    confounds = et._process_confounds(data, None)
     assert isinstance(confounds, pd.DataFrame)
     assert len(correct_confounds) == len(confounds.columns) == 5
     assert confounds.shape == (40, 5)
     assert all(i in correct_confounds for i in confounds.columns)
 
     data.signal_clean_info["use_confounds"] = False
-    confounds = _process_confounds(data, None)
+    confounds = et._process_confounds(data, None)
     assert not confounds
 
 
@@ -969,6 +967,22 @@ def test_non_censor_error():
             use_confounds=False,
             fd_threshold=0.35,
         )
+
+
+def test_filter_censored_scan_indices_unit():
+    """
[…]
