### 2025-04-22 9aafcdf
Minor code cleaning, remove redundancy, add test
