+++ tests/test_exceptions.py
@@ -1,4 +1,5 @@
 import glob, os, re
+
 import pytest
 
 from neurocaps.extraction import TimeseriesExtractor
