### 2025-04-22 00e14e0
[skip ci] Fix type hint for _Data
