### 2025-04-23 b7cec19
[skip ci] Add notes to simulate event data docstring
