### 2025-04-23 eb49eee
Refactor calculate_metrics and caps2radar
