+++ neurocaps/analysis/cap.py
@@ -462,7 +462,6 @@ class CAP(_CAPGetter):
         group_map = {
             group: sorted(set(self._subject_table).intersection(self._groups[group])) for group in self._groups
         }
-
         # Collect timeseries data in lists
         group_arrays = {group: [] for group in self._groups}
         for group in group_map:
@@ -951,105 +950,47 @@ class CAP(_CAPGetter):
         # Assign each subject's frame to a CAP
         predicted_subject_timeseries = self._build_prediction_dict(subject_timeseries, runs, continuous_runs)
 
-        cap_names, cap_numbers, group_cap_counts = self._get_caps_info()
+        cap_names, max_cap, group_cap_counts = self._get_caps_info()
 
         # Get combination of transitions in addition to building the base dataframe dictionary
-        if "transition_probability" in metrics:
-            cap_pairs = self._get_pairs()
-            df_dict, temp_dict = self._build_df(metrics, cap_names, cap_pairs)
-        else:
-            df_dict = self._build_df(metrics, cap_names)
+        cap_pairs = self._get_pairs() if "transition_probability" in metrics else None
+        df_dict = self._build_df(metrics, cap_names, cap_pairs)
+
+        # Create function map
+        metric_map = {
+            "temporal_fraction": self._compute_temporal_fraction,
+            "counts": self._compute_counts,
+            "persistence": self._compute_persistence,
+            "transition_frequency": self._compute_transition_frequency,
+            "transition_probability": self._compute_transition_probability,
+        }
 
         # Generate list for iteration
         distributed_dict = self._distribute(predicted_subject_timeseries)
 
         for subj_id in tqdm(distributed_dict, desc="Computing Metrics for Subjects", disable=not progress_bar):
             for group, curr_run in distributed_dict[subj_id]:
-                if "temporal_fraction" in metrics:
-                    # Get frequency
-                    frequency_dict = {
-                        key: np.where(predicted_subject_timeseries[subj_id][curr_run] == key, 1, 0).sum()
-                        for key in range(1, group_cap_counts[group] + 1)
-                    }
+                for metric in metrics:
+                    sub_info = [subj_id, group, curr_run]
 
-                    self._update_dict(cap_numbers, group_cap_counts[group], frequency_dict)
+                    df = df_dict[metric] if metric != "transition_probability" else df_dict[metric][group]
 
-                    if "temporal_fraction" in metrics:
-                        proportion_dict = {
-                            key: value / (len(predicted_subject_timeseries[subj_id][curr_run]))
-                            for key, value in frequency_dict.items()
-                        }
+                    # Common base arguments for all functions
+                    args = [predicted_subject_timeseries[subj_id][curr_run], sub_info, df]
 
-                        new_row = [subj_id, group, curr_run] + [items for items in proportion_dict.values()]
-                        df_dict["temporal_fraction"].loc[len(df_dict["temporal_fraction"])] = new_row
+                    if metric in ["temporal_fraction", "counts", "persistence"]:
+                        args.extend([group_cap_counts[group], max_cap])
+                        args += [tr] if metric == "persistence" else []
 
-                if "counts" in metrics:
-                    count_dict = {}
-                    for target in cap_numbers:
-                        # + 1 is always added to segments to handle the + 1 needed to account for transitions and to avoid
-                        # a NaN for persistence. This ensures counts is 0 if target not present
-                        if target in predicted_subject_timeseries[subj_id][curr_run]:
[…]
