### 2025-04-23 1662fc2
Move order of function and update log
