+++ .pre-commit-config.yaml
@@ -32,7 +32,7 @@ repos:
         - black
 
 -   repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.11.4
+    rev: v0.11.6
     hooks:
         - id: ruff
           # Ensure no print statements are in the codebase; only logging allowed
