### 2025-04-24 f5ddf58
Remove data, prep submodule
