+++ tests/data/nilearn_data/README.md
@@ -1,6 +1,0 @@
-# Nilearn data folder
-
-This directory is used by Nilearn to store datasets
-and atlases downloaded from the internet.
-It can be safely deleted.
-If you delete it, previously downloaded data will be downloaded again.
