### 2025-04-24 b5a1766
Convert "tests/data" to submodule; update yaml and demos
