+++ .github/workflows/minimum_dependencies.yml
@@ -19,6 +19,8 @@ jobs:
     name: Python ${{ matrix.python-ver }} check on ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v4
+        with:
+          submodules: 'recursive'
       - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-ver }}
