+++ .github/workflows/benchmark.yml
@@ -16,6 +16,7 @@ jobs:
     steps:
       - uses: actions/checkout@v4
         with:
+          submodules: 'recursive'
           fetch-depth: 0
       - uses: actions/setup-python@v5
         with:
