+++ .github/workflows/docker-deploy.yml
@@ -11,7 +11,8 @@ jobs:
 
     steps:
       - uses: actions/checkout@v4
-
+        with:
+          submodules: 'recursive'
       - name: Log in to Docker Hub
         uses: docker/login-action@v3
         with:
