### 2025-04-24 da1f55c
[skip ci] Update installation guide
