+++ docs/user_guide/installation.rst
@@ -45,6 +45,8 @@ To install the latest development version from the source, there are two options
     git clone --depth 1 https://github.com/donishadsmith/neurocaps/
     cd neurocaps
     pip install -e .
+    # Clone with submodules to include test dataset ~140 MB
+    git submodule update --init
 
 Windows Users
 ^^^^^^^^^^^^^
@@ -55,6 +57,8 @@ To include PyBIDS when installing the development version on Windows, use:
     git clone --depth 1 https://github.com/donishadsmith/neurocaps/
     cd neurocaps
     pip install -e .[windows]
+    # Clone with submodules to include test dataset ~140 MB
+    git submodule update --init
 
 Docker
 ------
