+++ README.md
@@ -68,6 +68,8 @@ pip install git+https://github.com/donishadsmith/neurocaps.git
 git clone --depth 1 https://github.com/donishadsmith/neurocaps/
 cd neurocaps
 pip install -e .
+# Clone with submodules to include test dataset ~140 MB
+git submodule update --init
 
 ```
 **Windows Users**
@@ -79,6 +81,8 @@ To include PyBIDS when installing the development version on Windows, use:
 git clone --depth 1 https://github.com/donishadsmith/neurocaps/
 cd neurocaps
 pip install -e .[windows]
+# Clone with submodules to include test dataset ~140 MB
+git submodule update --init
 ```
 
 ## Docker
