+++ .github/workflows/python-publish.yml
@@ -24,7 +24,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install build setuptools wheel twine
-      shell: bash
 
     - name: Build and publish
       env:
@@ -33,4 +32,3 @@ jobs:
       run: |
         python -m build
         twine upload dist/*
-      shell: bash

[…]
