### 2025-04-25 55c619b
Minor refactor
