### 2025-04-26 ceb37cb
Refactor and expand file support
- "plotting_defaults.py" changed to "plotting_utils.py"
- "heatmap_utils.py" deleted and functions are now stored in
  _MatrixVisualizer class in "plotting_utils.py"
- Shared plotting function now stored in PlotFuncs to increase
  reusability
- "pickle_utils.py" changed to "io.py"
- Created IO class to store IO operations
- Removed some internal methods in ``CAP`` that are either generalized
  IO operations or generalized plotting functions
- Support compressed serialized files
- Logged IO warnings come from "io.py" specifically instead of being
  scattered acr
[…]
