+++ neurocaps/exceptions.py
@@ -40,8 +40,8 @@ class UnsupportedFileExtensionError(Exception):
     """
     Unsupported File Extension Exception.
 
-    This exception occurs when attempting to unpickle a file that does not have a supported extension (i.e. ".pkl",
-    ".pickle", and ".joblib").
+    This exception occurs when attempting to unpickle a file that does not have a supported extension (e.g ".pkl",
+    ".pickle", ".joblib", the preceding extensions with ".gz" if version >= 0.28.2).
 
     .. versionadded:: 0.26.5
     """
