+++ neurocaps/_utils/analysis/__init__.py
@@ -1,4 +1,3 @@
 from .capgetter import _CAPGetter
 from .cap2statmap import _build_tree, _cap2statmap, _get_target_indices
 from .run_kmeans import _run_kmeans
-from .heatmap_utils import _create_display, _save_contents

[…]
