### 2025-04-27 7c9c22d
Ensure all plots closed when ``show_figs`` is False
