### 2025-05-01 5bf2b5c
Improve event timing conversion
