### 2025-05-02 #8001 @juliaferraioli
Hi @htwangtw @EtienneCmb @cmaumet would you be interested in reviewing this submission for JOSS?
