### 2025-05-03 385e4bd
Reduce line length from 120 to 100 characters
- Minor doc string cleaning
- Minor refactoring
- Clean some comments
