+++ neurocaps/analysis/cap.py
@@ -38,96 +38,104 @@ class CAP(_CAPGetter):
     """
     Co-Activation Patterns (CAPs) Class.
 
-    Performs k-means clustering for CAP identification, computes temporal dynamics metrics, and provides visualization
-    tools for analyzing brain activation patterns.
+    Performs k-means clustering for CAP identification, computes temporal dynamics metrics, and
+    provides visualization tools for analyzing brain activation patterns.
 
     Parameters
     ----------
     parcel_approach: :obj:`ParcelConfig`, :obj:`ParcelApproach`, or :obj:`str`, default=None
-        Specifies the parcellation approach to use. Options are "Schaefer", "AAL", or "Custom". Can be initialized with
-        parameters, as a nested dictionary, or loaded from a pickle file. For detailed documentation on the expected
-        structure, see the type definitions for ``ParcelConfig`` and ``ParcelApproach`` in the "See Also" section.
+        Specifies the parcellation approach to use. Options are "Schaefer", "AAL", or "Custom". Can
+        be initialized with parameters, as a nested dictionary, or loaded from a pickle file. For
+        detailed documentation on the expected structure, see the type definitions for ``ParcelConfig``
+        and ``ParcelApproach`` in the "See Also" section.
 
     groups: :obj:`dict[str, list[str]]` or :obj:`None`, default=None
-        Optional mapping of group names to lists of subject IDs for group-specific analyses. If None, on the first call
-        of ``self.get_caps()``, "All Subjects" will be set as the default group name and be populated with the subject
-        IDs in ``subject_timeseries``. Groups remain fixed for the entire instance of the class.
+        Optional mapping of group names to lists of subject IDs for group-specific analyses. If
+        None, on the first call of ``self.get_caps()``, "All Subjects" will be set as the default
+        group name and be populated with the subject IDs in ``subject_timeseries``. Groups remain
+        fixed for the entire instance of the class.
 
 
     Properties
     ----------
     parcel_approach: :obj:`ParcelApproach`
-        Parcellation information with "maps" (path to parcellation file), "nodes" (labels), and "regions"
-        (anatomical regions or networks). This property is also settable (accepts a dictionary or pickle file).
-        Returns a deep copy.
+        Parcellation information with "maps" (path to parcellation file), "nodes" (labels), and
+        "regions" (anatomical regions or networks). This property is also settable (accepts a
+        dictionary or pickle file). Returns a deep copy.
 
     groups: :obj:`dict[str, list[str]]` or :obj:`None`:
         Mapping of groups names to lists of subject IDs. Returns a deep copy.
 
     subject_table: :obj:`dict[str, str]` or :obj:`None`
-        Lookup table mapping subject IDs to their groups. Derived from ``self.groups`` each time ``self.get_caps()``
-        is ran. While this property can be modified using its setter, any changes will be overwritten based on
-        ``self.groups`` on the subsequent call to ``self.get_caps()``. Returns a deep copy.
+        Lookup table mapping subject IDs to their groups. Derived from ``self.groups`` each time
+        ``self.get_caps()`` is ran. While this property can be modified using its setter, any
+        changes will be overwritten based on ``self.groups`` on the subsequent call to
+        ``self.get_caps()``. Returns a deep copy.
 
     n_clusters: :obj:`int`, :obj:`list[int]`, or :obj:`None`
         An integer or list of integers representing the number of clusters used for k-means.
         Defined after running ``self.get_caps()``.
 
     cluster_selection_method: :obj:`str` or :obj:`None`:
[…]
