+++ docs/introduction.rst
@@ -51,42 +51,62 @@ Citing
 
 Usage
 -----
-This package contains two main classes: ``TimeseriesExtractor`` for extracting the timeseries, and ``CAP`` for performing the CAPs analysis.
+This package contains two main classes: ``TimeseriesExtractor`` for extracting the timeseries, and
+``CAP`` for performing the CAPs analysis.
 
 Main features for ``TimeseriesExtractor`` includes:
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-- **Timeseries Extraction:** Extract timeseries for resting-state or task data using Schaefer, AAL, or a lateralized Custom parcellations (which can be manually defined) for spatial dimensionality reduction.
-- **Parallel Processing:** Parallelize at the subject-level (one subject per CPU core) to speed up timeseries extraction.
-- **Saving Timeseries:** Save the nested dictionary containing timeseries (mapping subject id -> run id -> timeseries data) as a pickle file.
-- **Reporting Quality Control:** Reports statistics related to framewise displacement and dummy volumes per-subject.
-- **Visualization:** Visualize the timeseries at the region or node level of the parcellation for a given subject and run.
+- **Timeseries Extraction:** Extract timeseries for resting-state or task data using Schaefer, AAL,
+or a lateralized Custom parcellations (which can be manually defined) for spatial dimensionality
+reduction.
+- **Parallel Processing:** Parallelize at the subject-level (one subject per CPU core) to speed up
+timeseries extraction.
+- **Saving Timeseries:** Save the nested dictionary containing timeseries (mapping subject id
+-> run id -> timeseries data) as a pickle file.
+- **Reporting Quality Control:** Reports statistics related to framewise displacement and dummy
+volumes per-subject.
+- **Visualization:** Visualize the timeseries at the region or node level of the parcellation for a
+given subject and run.
 
 Main features for ``CAP`` includes:
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 - **Grouping:** Perform CAPs analysis for entire sample or groups of subject IDs
-- **Optimal Cluster Size Identification:** Perform the Davies Bouldin, Silhouette, Elbow, or Variance Ratio criterions to identify the optimal cluster size and automatically save the optimal model as an attribute.
+- **Optimal Cluster Size Identification:** Perform the Davies Bouldin, Silhouette, Elbow, or
+Variance Ratio criterions to identify the optimal cluster size and automatically save the optimal
+model as an attribute.
 - **Parallel Processing:** Use parallel processing to speed up optimal cluster size identification.
-- **CAPs Visualization:** Visualize the CAPs as outer products or heatmaps at either the region or node level of the parcellation.
+- **CAPs Visualization:** Visualize the CAPs as outer products or heatmaps at either the region or
+node level of the parcellation.
 - **Save CAPs as NifTIs:** Convert the atlas used for parcellation to a statistical NifTI image.
 - **Surface Plot Visualization:** Project CAPs onto a surface plot.
 - **Correlation Matrix Creation:** Create a correlation matrix from CAPs.
-- **CAPs Metrics Calculation:** Calculate several CAPs metrics as described in `Liu et al., 2018 <https://doi.org/10.1016/j.neuroimage.2018.01.041>`_ [1]_ and `Yang et al., 2021 <https://doi.org/10.1016/j.neuroimage.2021.118193>`_ [2]_:
-    - *Temporal Fraction:* The proportion of total volumes spent in a single CAP over all volumes in a run.
[…]
