### 2025-05-06 1c2703a
Add function to clear variable remove private class
