+++ neurocaps/_utils/io.py
@@ -1,5 +1,5 @@
 """
-Internal IO class and functions for creating directories, issuing warnings, serializing (with
+Internal IO functions for creating directories, issuing warnings, serializing (with
 Joblib or pickle), unserializing (with Joblib), and additional utilities.
 """
 
@@ -14,114 +14,112 @@ from ..exceptions import UnsupportedFileExtensionError
 LG = _logger(__name__)
 
 
-class _IO:
-    @staticmethod
-    def makedir(output_dir: str) -> None:
-        """Creates non-existent directory."""
-        if output_dir and not os.path.exists(output_dir):
-            os.makedirs(output_dir)
+def _makedir(output_dir: str) -> None:
+    """Creates non-existent directory."""
+    if output_dir and not os.path.exists(output_dir):
+        os.makedirs(output_dir)
 
-    @staticmethod
-    def issue_file_warning(param_name: str, param: str, output_dir: str) -> None:
-        """Issues warning."""
-        if param is not None and not output_dir:
-            LG.warning(
-                f"`{param_name}` supplied but no `output_dir` specified. Files will not be saved."
-            )
 
-    @staticmethod
-    def filename(
-        base_name: str, add_name: Union[str, None], pos: str, ext: Union[None, str] = None
-    ) -> str:
-        """
-        Adds the file extension and the prefix or suffix to the file name depending on if ``pos``
-        is "prefix" or "suffix".
-        """
-        if not add_name:
-            return f"{base_name}.{ext}" if ext else base_name
+def _issue_file_warning(param_name: str, param: str, output_dir: str) -> None:
+    """Issues warning."""
+    if param is not None and not output_dir:
+        LG.warning(
+            f"`{param_name}` supplied but no `output_dir` specified. Files will not be saved."
+        )
 
-        if pos == "prefix":
-            # Use - since prefix is currently for ``CAP.calculate_metrics``
-            add_name = os.path.splitext(add_name.rstrip())[0].rstrip()
-            filename = f"{add_name}-{base_name}"
-        else:
-            filename = f"{base_name}_{add_name.rstrip().replace(' ', '_')}"
 
-        filename = filename.replace(" ", "_")
+def _filename(
+    base_name: str, add_name: Union[str, None], pos: str, ext: Union[None, str] = None
+) -> str:
+    """
+    Adds the file extension and the prefix or suffix to the file name depending on if ``pos``
+    is "prefix" or "suffix".
+    """
+    if not add_name:
+        return f"{base_name}.{ext}" if ext else base_name
 
-        return f"{filename}.{ext}" if ext else filename
+    if pos == "prefix":
+        # Use - since prefix is currently for ``CAP.calculate_metrics``
+        add_name = os.path.splitext(add_name.rstrip())[0].rstrip()
+        filename = f"{add_name}-{base_name}"
+    else:
+        filename = f"{base_name}_{add_name.rstrip().replace(' ', '_')}"
 
-    @staticmethod
-    def serialize(obj: Any, output_dir: str, save_filename: str, use_joblib=False) -> None:
-        """
-        Serialize an object with standard pickle or Joblib. Internally, plots use pickle for
-        compatibility with pickle or Joblib.
-        """
-        with open(os.path.join(output_dir, save_filename), "wb") as f:
-            joblib.dump(obj, f) if use_joblib else pickle.dump(obj, f)
+    filename = filename.replace(" ", "_")
 
-    @staticmethod
-    def get_obj(obj: Any, needs_deepcopy=True) -> Any:
-        """Determines if object needs deepcopy or needs to be unserialized."""
-        if not isinstance(obj, str):
[…]
