### 2025-05-06 #8001 @juliaferraioli
Great suggestions, @donishadsmith!

@djmannion and @behinger, would either of you be able to review this submission for JOSS?
