### 2025-05-08 #8001 @donishadsmith
Hi @juliaferraioli,

I apologize for pinging you again so soon. Following up on djmannion's response, and keeping in mind that htwangtw mentioned being available after May 20th, I wanted to suggest another potential reviewer in case one is still needed by then.

From the [reviewer website](https://reviewers.joss.theoj.org/reviewers/1750), I've identified Dr. Roberto Guidotti (robbisg), who has been developing a package called [sekupy](https://github.com/robbisg/sekupy), which includes a [module](https://sekupy.readthedocs.io/en/latest/states/base.html) for assessing dynamic brain states with clustering algorithms and computing temporal dynamics metrics. This shares a strong methodological overlap with core functionalities from my submission.

Thank you!
