### 2025-05-08 #8001 @juliaferraioli
Another great suggestion -- one to which I'll add. @robbisg and @snastase, would either of you you be able to review this paper for JOSS?
