### 2025-05-08 #8001 @danielskatz
@juliaferraioli - FYI, you should be able to use the same command @Kevin-Mattheus-Moerman did, or even just `@editorialbot assign me as editor`
