+++ neurocaps/analysis/cap.py
@@ -685,7 +685,8 @@ class CAP(_CAPGetter):
         }
 
         # Create plot dictionary
-        plot_dict = _check_kwargs(_PlotDefaults.get_caps(), **kwargs)
+        filtered_kwargs = {k: v for k, v in kwargs.items() if k not in ["S", "max_nbytes"]}
+        plot_dict = _check_kwargs(_PlotDefaults.get_caps(), **filtered_kwargs)
 
         plt.figure(figsize=plot_dict["figsize"])
 
@@ -779,17 +780,10 @@ class CAP(_CAPGetter):
                 "transition_frequency",
                 "transition_probability",
             ],
-            list[
-                Literal[
-                    "temporal_fraction",
-                    "persistence",
-                    "counts",
-                    "transition_frequency",
-                    "transition_probability",
-                ]
-            ],
+            list[str],
+            tuple[str],
             None,
-        ] = None,
+        ] = ("temporal_fraction", "persistence", "counts", "transition_frequency"),
         return_df: bool = True,
         output_dir: Optional[str] = None,
         prefix_filename: Optional[str] = None,
@@ -916,18 +910,18 @@ class CAP(_CAPGetter):
                 - If only a single run available for a subject, the original run ID (as opposed to
                   "run-continuous") will be used.
 
-        metrics: {"temporal_fraction", "persistence", "counts", "transition_frequency", "transition_probability"}, \
-                 :obj:`list["temporal_fraction", "persistence", "counts", "transition_frequency", "transition_probability"]`, \
-                 or :obj:`None`, default=None
+        metrics: {"temporal_fraction", "persistence", "counts", "transition_frequency", "transition_probability"},\
+                 :obj:`list[str]`, :obj:`tuple[str]`, or :obj:`None`,\
+                 default=("temporal_fraction", "persistence", "counts", "transition_frequency")
             The metrics to calculate. Available options include "temporal_fraction", "persistence",
             "counts", "transition_frequency", and "transition_probability". Defaults to
-            ``["temporal_fraction", "persistence", "counts", "transition_frequency"]`` if None.
+            ``("temporal_fraction", "persistence", "counts", "transition_frequency")`` if None.
 
-            .. versionchanged:: 0.28.0 Default changed to None; however, the default metrics\
-            remains the same and is backwards compatible.
+            .. versionchanged:: 0.28.6 Default changed to tuple to provide better clarity; however,\
+            the default metrics remains the same and is backwards compatible.
 
         return_df: :obj:`str`, default=True
-            If True, returns ``pandas.DataFrame`` inside a dictionary`, mapping each dataframe to
+            If True, returns ``pandas.DataFrame`` inside a dictionary, mapping each dataframe to
             their metric.
 
         output_dir: :obj:`str` or :obj:`None`, default=None
@@ -1124,7 +1118,7 @@ class CAP(_CAPGetter):
     @staticmethod
     def _filter_metrics(metrics):
         metrics = (
-            ["temporal_fraction", "persistence", "counts", "transition_frequency"]
+            ("temporal_fraction", "persistence", "counts", "transition_frequency")
             if metrics is None
             else metrics
         )
@@ -1178,12 +1172,14 @@ class CAP(_CAPGetter):
 
             if miss_runs:
                 LG.warning(
-                    f"[SUBJECT: {subj_id}] Does not have the requested runs: {', '.join(miss_runs)}."
+                    f"[SUBJECT: {subj_id}] Does not have the requested runs: "
+                    f"{', '.join(miss_runs)}."
                 )
 
             if not subject_runs:
                 LG.warning(
[…]
