### 2025-05-09 6534a23
Update test for future nilearn version
