### 2025-05-12 #8001 @donishadsmith
Hey @juliaferraioli, snastaseq made an excellent suggestion! Similar to robbisg, yibeichen also has [experience](https://github.com/yibeichan/prettymouth) with dynamic functional connectivity techniques and metrics related to temporal dynamics.
