### 2025-05-12 #8001 @juliaferraioli
Thanks @danielskatz; I though there were permissions associated with the bot :-)

@yibeichen would you be able to review this paper for JOSS?
