### 2025-05-12 cf3b625
Add docs and types to internal functions
- Also simplify return for ``_initialize_outer_product_subplot`` to make
  return hint easier.
