+++ neurocaps/extraction/timeseriesextractor.py
@@ -2,11 +2,12 @@
 
 import json, os, re
 from functools import lru_cache
-from typing import Callable, Literal, Optional, Union
+from typing import Any, Callable, Literal, Optional, Union
 from typing_extensions import Self
 
 import matplotlib.pyplot as plt, numpy as np
 from joblib import Parallel, delayed
+from numpy.typing import NDArray
 from pandas import DataFrame
 from tqdm.auto import tqdm
 
@@ -327,8 +328,8 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
         }
 
     @staticmethod
-    def _validate_init_params(param, struct):
-
+    def _validate_init_params(param: str, struct: Any) -> None:
+        """Validates structure for ``dummy_scans`` and ``fd_threshold``."""
         mandatory_keys = {
             "dummy_scans": {"auto": (bool,)},
             "fd_threshold": {"threshold": (float, int)},
@@ -729,7 +730,8 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
 
         return self
 
-    def _validate_get_bold_params(self):
+    def _validate_get_bold_params(self) -> None:
+        """Validates ``condition_tr_shift`` and ``slice_time_ref``."""
         if self._task_info["condition_tr_shift"] != 0:
             if (
                 not isinstance(self._task_info["condition_tr_shift"], int)
@@ -753,7 +755,20 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
 
     @staticmethod
     @lru_cache(maxsize=4)
-    def _call_layout(bids_dir, pipeline_name):
+    def _call_layout(bids_dir: str, pipeline_name: Union[str]):
+        """
+        Returns ``BIDSLayout``.
+
+        Raises
+        ------
+        ModuleNotFoundError
+            The pybids package is not installed.
+
+        Note
+        ----
+        The type hint for the returned (``BIDSLayout`)` is not added to allow certain functions in
+        ``TimeseriesExtractor`` to be used on Windows machines that do not have pybids installed.
+        """
         try:
             import bids
         except ModuleNotFoundError:
@@ -780,8 +795,18 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
 
         return layout
 
-    # Get valid subjects to iterate through
-    def _setup_extraction(self, layout, subj_ids, exclude_niftis, verbose):
+    def _setup_extraction(
+        self, layout, subj_ids: list[str], exclude_niftis: list, verbose: bool
+    ) -> None:
+        """
+        Get valid subjects (stored in ``self._subject_info``) and all required information needed
+        to extract timeseries.
+
+        Note
+        ----
+        The type hint for ``layout`` is not added to allow certain in ``TimeseriesExtractor`` to be
+        used on Windows machines that do not have pybids installed.
+        """
         base_dict = {"layout": layout, "subj_id": None}
 
         for subj_id in subj_ids:
@@ -857,14 +882,22 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
     def _get_files(
         self,
         layout,
-        extension,
-        subj_id,
-        scope="derivatives",
-        suffix=None,
-        desc=None,
-        event=False,
-        space="attr",
+        extension: str,
+        subj_id: str,
+        scope: str = "derivatives",
+        suffix: Union[str, None] = None,
+        desc: Union[str, None] = None,
+        event: bool = False,
+        space: str = "attr",
     ):
+        """
+        Queries specific files (sorted lexicographically) using ``BidsLayout``.
+
+        Note
+        ----
+        The type hint for ``layout`` is not added to allow certain in ``TimeseriesExtractor`` to be
+        used on Windows machines that do not have pybids installed.
[…]
