### 2025-05-13 #8196 @donishadsmith
Hi, @yibeichan and @robbisg! If it helps with the review, I have a pull request that adds type hints and some basic doc strings to the internal functions for the public classes. However, since the review started, I wanted to ask permission first before merging anything. It does pass all tests. The specific pull request is at: donishadsmith/neurocaps#25.
