### 2025-05-13 b81d727
Simplify conditional statement & minor typing update
