### 2025-05-16 44f92f7
Make private function public to obtain CAP assignments
- Temporarily restrict plotly version
