### 2025-05-16 1d2b564
Add docs and types to internal functions & new function (#25)
* Add docs and types to internal functions

- Also simplify return for ``_initialize_outer_product_subplot`` to make
  return hint easier.

* Simplify conditional statement & minor typing update

* Type hint update

* Make private function public to obtain CAP assignments

- Temporarily restrict plotly version
