+++ neurocaps/analysis/cap.py
@@ -11,6 +11,7 @@ from joblib import Parallel, delayed
 from nilearn.plotting.cm import _cmap_d
 from neuromaps.transforms import mni152_to_fslr, fslr_to_fslr
 from neuromaps.datasets import fetch_fslr
+from numpy.typing import NDArray
 from scipy.stats import pearsonr
 from tqdm.auto import tqdm
 
@@ -488,7 +489,10 @@ class CAP(_CAPGetter):
         """
         self._groups = None
 
-    def _generate_lookup_table(self):
+    def _generate_lookup_table(self) -> None:
+        """
+        Creates dictionary mapping subject IDs to their associated group. This function is
+        called whenever ``self._concatenate_timeseries`` is called."""
         self._subject_table = {}
 
         for group in self._groups:
@@ -501,7 +505,13 @@ class CAP(_CAPGetter):
                 else:
                     self._subject_table.update({subj_id: group})
 
-    def _concatenate_timeseries(self, subject_timeseries, runs, progress_bar):
+    def _concatenate_timeseries(
+        self, subject_timeseries: SubjectTimeseries, runs: Union[list, None], progress_bar: bool
+    ) -> dict[str, NDArray]:
+        """
+        Concatenates the timeseries data of all subjects into a single numpy array if ``groups`` are
+        None or group-specific numpy arrays.
+        """
         # Create dictionary for "All Subjects" if no groups are specified
         if not self._groups:
             LG.info(
@@ -563,7 +573,8 @@ class CAP(_CAPGetter):
 
         return concatenated_timeseries
 
-    def _scale(self, concatenated_timeseries):
+    def _scale(self, concatenated_timeseries: NDArray) -> NDArray:
+        """Scales the concatenated timeseries data."""
         for group in self._groups:
             concatenated_timeseries[group], self._mean_vec[group], self._stdev_vec[group] = (
                 _standardize(concatenated_timeseries[group], return_parameters=True)
@@ -572,7 +583,13 @@ class CAP(_CAPGetter):
         return concatenated_timeseries
 
     @staticmethod
-    def _get_runs(requested_runs, curr_runs):
+    def _get_runs(
+        requested_runs: Union[list, None], curr_runs: list
+    ) -> tuple[list, Union[list, None]]:
+        """
+        Filters the current runs available for a subject if specific runs are requested.
+        Also returns a list of missing runs that were requested
+        """
         if requested_runs:
             requested_runs = [str(run).removeprefix("run-") for run in requested_runs]
             requested_runs = [f"run-{run}" for run in requested_runs]
@@ -583,8 +600,16 @@ class CAP(_CAPGetter):
         return runs, miss_runs
 
     def _select_optimal_clusters(
-        self, method, configs, show_figs, output_dir, progress_bar, as_pickle, **kwargs
-    ):
+        self,
+        method: str,
+        configs: dict,
+        show_figs: bool,
+        output_dir: Union[str, None],
+        progress_bar: bool,
+        as_pickle: bool,
+        **kwargs,
+    ) -> None:
+        """Selects optimal number of clusters based on the specific ``cluster_selection_method``."""
         self._cluster_scores = {}
         self._optimal_n_clusters = {}
         self._kmeans = {}
@@ -675,8 +700,16 @@ class CAP(_CAPGetter):
         self._cluster_scores.update({"Scores": performance_dict})
 
     def _plot_method(
-        self, method, performance_dict, group, show_figs, output_dir, as_pickle, **kwargs
-    ):
+        self,
+        method: str,
+        performance_dict: dict,
+        group: str,
+        show_figs: bool,
+        output_dir: Union[str, None],
+        as_pickle: bool,
+        **kwargs,
+    ) -> None:
[…]
