### 2025-05-16 #8196 @donishadsmith
Hi, @yibeichan and @robbisg! Just to note, I ended up having to increment the version due to a user sending an email about a feature that they needed (which is directly tested) and due to an issue with a plotly's latest release breaking a functionality and caused a test to fail.
