### 2025-05-16 01fe379
[skip ci] Add version directive to doc
