### 2025-05-18 5b8c7ef
Add doc strings to properties, update version restriction for plotly, add check for parallel_log_config
