+++ .github/workflows/testing.yaml
@@ -2,7 +2,7 @@ name: Run Tests
 
 on:
   push:
-    branches: [main, development]
+    branches: [main]
   pull_request:
     branches:
       - '*'
