### 2025-05-18 cb864e7
Add doc strings to properties, update version restriction for plotly,… (#26)
* Add doc strings to properties, update version restriction for plotly, add check for parallel_log_config

* Update yaml, prevent running test on push for development to prioritize merge

* Update benchmark yaml for pull request
