+++ neurocaps/_utils/logger.py
@@ -13,12 +13,36 @@ _USER_MODULE_HANDLERS = {}
 
 
 class _Flush(logging.StreamHandler):
+    """Flush logs immediately."""
+
     def emit(self, record: logging.LogRecord) -> None:
         super().emit(record)
         self.flush()
 
 
-def _logger(name, flush=False, top_level=True, parallel_log_config=None):
+def _logger(
+    name: str, flush: bool = False, top_level: bool = True, parallel_log_config: dict = None
+):
+    """
+    Generates module specific loggers, defaults to outputting logs at the informational level to
+    the standard output. The ``top_level`` parameter is used to determine if logging configurations
+    are present at the time of package import to prevent logging's from sending logs to standard
+    error by default.
+
+    Notes
+    -----
+    **Parallel Processing:** When parallel processing using joblib's loky backend with certain
+    modules such as "neurocaps._utils.extraction.extract_timeseries", the logger needs to be
+    passed to each child process to ensure to ensure logging configurations are respected. Hence in
+    the "extract_timeseries", this function is also called inside ``_extract_timeseries`` function.
+
+    To ensure that there is no issue with race conditions when intending to log to files,
+    ``parallel_log_config`` is offered to allow for controlled queue-based logging with
+    ``QueueHandler``. This requires a centralized queue with a manager object that is able to share
+    reference objects of the centralized queue with each child process. Then ``QueueListener`` can
+    be set up in the main process to listen to the centralized queue object and write messages to
+    the specified file.
+    """
     global _USER_ROOT_HANDLER, _USER_MODULE_HANDLERS
 
     logger = logging.getLogger(name)
@@ -33,8 +57,13 @@ def _logger(name, flush=False, top_level=True, parallel_log_config=None):
     if logger.name == parallel_module and not top_level and parallel_log_config:
         # Only QueueHandler will be in handler list
         logger.handlers.clear()
-        if "queue" in parallel_log_config:
-            logger.addHandler(QueueHandler(parallel_log_config["queue"]))
+        queue = parallel_log_config.get("queue")
+        # Non-strict check
+        if not queue:
+            ValueError("'queue' is a mandatory key and must contain a queue with a manager object.")
+        else:
+            logger.addHandler(QueueHandler(queue))
+
         if "level" in parallel_log_config:
             logger.setLevel(parallel_log_config["level"])
 
