+++ neurocaps/_utils/analysis/capgetter.py
@@ -18,6 +18,11 @@ class _CAPGetter:
     ### Attributes exist when CAP initialized
     @property
     def parcel_approach(self) -> Union[ParcelApproach, None]:
+        """
+        Parcellation information with "maps" (path to parcellation file), "nodes" (labels), and
+        "regions" (anatomical regions or networks). This property is also settable (accepts a
+        dictionary or pickle file). Returns a deep copy.
+        """
         return self._parcel_approach
 
     @parcel_approach.setter
@@ -26,11 +31,18 @@ class _CAPGetter:
 
     @property
     def groups(self) -> Union[dict[str, list[str]], None]:
+        """Mapping of groups names to lists of subject IDs. Returns a deep copy."""
         return copy.deepcopy(self._groups)
 
     ### Attributes exist when CAP.get_caps() used
     @property
     def subject_table(self) -> Union[dict[str, str], None]:
+        """
+        Lookup table mapping subject IDs to their groups. Derived from ``self.groups`` each time
+        ``self.get_caps()`` is ran. While this property can be modified using its setter, any
+        changes will be overwritten based on ``self.groups`` on the subsequent call to
+        ``self.get_caps()``. Returns a deep copy.
+        """
         return copy.deepcopy(getattr(self, "_subject_table", None))
 
     @subject_table.setter
@@ -45,22 +57,49 @@ class _CAPGetter:
 
     @property
     def n_clusters(self) -> Union[int, list[int], None]:
+        """
+        An integer or list of integers representing the number of clusters used for k-means.
+        Defined after running ``self.get_caps()``.
+        """
         return getattr(self, "_n_clusters", None)
 
     @property
     def n_cores(self) -> Union[int, None]:
+        """
+        Number of cores specified used for multiprocessing with Joblib. Defined after running
+        ``self.get_caps()``.
+        """
         return getattr(self, "_n_cores", None)
 
     @property
     def runs(self) -> Union[list[Union[int, str]], None]:
+        """
+        Run IDs specified in the analysis. Defined after running ``self.get_caps()``.
+        """
         return getattr(self, "_runs", None)
 
     @property
     def standardize(self) -> Union[bool, None]:
+        """
+        Whether region-of-interests (ROIs)/columns were standardized during analysis.
+        Defined after running ``self.get_caps()``.
+        """
         return getattr(self, "_standardize", None)
 
     @property
     def concatenated_timeseries(self) -> Union[dict[str, NDArray[np.floating]], None]:
+        """
+        Group-specific concatenated timeseries data. Can be deleted using
+        ``del self.concatenated_timeseries``. Defined after running ``self.get_caps()``. Returns
+        a reference.
+
+        ::
+
+            {"GroupName": np.array(shape=[(participants x TRs), ROIs])}
+
+        .. note:: For versions >= 0.25.0, subject IDs are sorted lexicographically prior to\
+        concatenation and the order is determined by ``self.groups``.
+        """
         return getattr(self, "_concatenated_timeseries", None)
 
     @concatenated_timeseries.deleter
@@ -69,26 +108,68 @@ class _CAPGetter:
 
     @property
     def means(self) -> Union[dict[str, NDArray[np.floating]], None]:
+        """
+        Group-specific feature means if standardization was applied. Defined after running
+        ``self.get_caps()``. Returns a deep copy.
+
+        ::
+
+            {"GroupName": np.array(shape=[1, ROIs])}
+        """
         return copy.deepcopy(getattr(self, "_mean_vec", None))
 
     @property
     def stdev(self) -> Union[dict[str, NDArray[np.floating]], None]:
+        """
[…]
