+++ .github/workflows/minimum_dependencies.yml
@@ -2,7 +2,7 @@ name: Test Minimum Dependencies
 
 on:
   push:
-    branches: [main, development]
+    branches: [main]
   pull_request:
     branches:
       - '*'
