### 2025-05-18 6fbc0ea
Restrict dependency versions
