### 2025-05-20 6b049e4
Improve typing
