+++ neurocaps/analysis/cap.py
@@ -1,7 +1,7 @@
 """Contains the CAP class for performing co-activation patterns analyses"""
 
 import collections, itertools, os, re, sys, tempfile
-from typing import Callable, Literal, Optional, Union
+from typing import Any, Callable, Literal, Optional, Union
 from typing_extensions import Self
 
 import nibabel as nib, numpy as np, matplotlib.pyplot as plt, pandas as pd, seaborn, surfplot
@@ -11,7 +11,7 @@ from joblib import Parallel, delayed
 from nilearn.plotting.cm import _cmap_d
 from neuromaps.transforms import mni152_to_fslr, fslr_to_fslr
 from neuromaps.datasets import fetch_fslr
-from numpy.typing import NDArray
+from numpy.typing import ArrayLike, NDArray
 from scipy.stats import pearsonr
 from tqdm.auto import tqdm
 
@@ -257,7 +257,7 @@ class CAP(_CAPGetter):
             Literal["elbow", "davies_bouldin", "silhouette", "variance_ratio"]
         ] = None,
         random_state: Optional[int] = None,
-        init: Union[Literal["k-means++", "random"], Callable, np.typing.ArrayLike] = "k-means++",
+        init: Union[Literal["k-means++", "random"], Callable, ArrayLike] = "k-means++",
         n_init: Union[Literal["auto"], int] = "auto",
         max_iter: int = 300,
         tol: float = 0.0001,
@@ -507,7 +507,10 @@ class CAP(_CAPGetter):
                     self._subject_table.update({subj_id: group})
 
     def _concatenate_timeseries(
-        self, subject_timeseries: SubjectTimeseries, runs: Union[list, None], progress_bar: bool
+        self,
+        subject_timeseries: SubjectTimeseries,
+        runs: Union[list[int], list[str], None],
+        progress_bar: bool,
     ) -> dict[str, NDArray]:
         """
         Concatenates the timeseries data of all subjects into a single numpy array if ``groups`` are
@@ -585,8 +588,8 @@ class CAP(_CAPGetter):
 
     @staticmethod
     def _get_runs(
-        requested_runs: Union[list, None], curr_runs: list
-    ) -> tuple[list, Union[list, None]]:
+        requested_runs: Union[list[int], list[str], None], curr_runs: list[str]
+    ) -> tuple[list[str], Union[list[str], None]]:
         """
         Filters the current runs available for a subject if specific runs are requested.
         Also returns a list of missing runs that were requested
@@ -603,7 +606,7 @@ class CAP(_CAPGetter):
     def _select_optimal_clusters(
         self,
         method: str,
-        configs: dict,
+        configs: dict[str, Any],
         show_figs: bool,
         output_dir: Union[str, None],
         progress_bar: bool,
@@ -703,7 +706,7 @@ class CAP(_CAPGetter):
     def _plot_method(
         self,
         method: str,
-        performance_dict: dict,
+        performance_dict: dict[str, dict[str, float]],
         group: str,
         show_figs: bool,
         output_dir: Union[str, None],
@@ -798,7 +801,7 @@ class CAP(_CAPGetter):
                 "Cannot calculate metrics since `self.kmeans` is None. Run `self.get_caps()` first."
             )
 
-    def _check_required_attrs(self, attr_names: list) -> None:
+    def _check_required_attrs(self, attr_names: list[str]) -> None:
         """Checks if certain attributes, needed for a specific function, are present."""
         for attr_name in attr_names:
             if getattr(self, attr_name, None) is None:
@@ -1156,7 +1159,7 @@ class CAP(_CAPGetter):
[…]
