### 2025-05-21 fd46729
[skip ci] Attempt general workflow image
- First attempt to visually communicate the general analysis process to
  derive CAPs
