### 2025-05-22 90416e9
Add new KNN approach
