+++ CHANGELOG.md
@@ -13,6 +13,11 @@ noted in the changelog (e.g., new functions or parameters, changes in parameter 
 improvements/enhancements. All fixes and modifications are backwards compatible.
 - *.postN* : Consists of documentation changes or metadata-related updates, such as modifications to type hints.
 
+## [0.29.5] - 2025-05-22
+### 🚀 New/Added
+- Adds "method" key to `knn_dict`, which defaults to majority vote KNN approach but allows for
+distance weighted KNN to be used.
+
 ## [0.29.4] - 2025-05-20
 ### 🐛 Fixes
 - Fixes and improve type hints for public API and some private functions
