### 2025-05-22 02d06af
Update nifti name in logged messages and docs
