### 2025-05-25 e08bc32
Add spearman correlation (#30)
