### 2025-05-26 f23e2dc
[skip ci] Update type
