### 2025-05-29 c665bb8
[skip ci] Add citation
