+++ paper/paper.bib
@@ -1,3 +1,14 @@
+ @article{Liu2013,
+          title={Decomposition of spontaneous brain activity into distinct fMRI co-activation patterns},
+          volume={7},
+          ISSN={1662-5137},
+          url={http://journal.frontiersin.org/article/10.3389/fnsys.2013.00101/abstract},
+          DOI={10.3389/fnsys.2013.00101},
+          journal={Frontiers in Systems Neuroscience},
+          author={Liu, Xiao and Chang, Catie and Duyn, Jeff H.},
+          year={2013}
+}
+
 @article{Jiang2022,
     title={Time-varying dynamic network model for dynamic resting state functional connectivity in fMRI and MEG imaging},
     volume={254}, ISSN={10538119},
