+++ README.md
@@ -14,7 +14,7 @@
 
 NeuroCAPs (**Neuro**imaging **C**o-**A**ctivation **P**attern**s**) is a Python package for performing Co-Activation
 Patterns (CAPs) analyses on resting-state or task-based fMRI data. CAPs identifies recurring brain states by applying
-k-means clustering on BOLD timeseries data [^1].
+k-means clustering on BOLD timeseries data [^1][^2].
 
 <img src="docs/assets/workflow.png" width=70% height=70%>
 
@@ -146,8 +146,8 @@ node level of the parcellation.
 - **Surface Plot Visualization:** Project CAPs onto a surface plot.
 - **Correlation Matrix Creation:** Create a correlation matrix from CAPs.
 - **CAPs Metrics Calculation:** Calculate several CAPs metrics as described in
-[Liu et al., 2018](https://doi.org/10.1016/j.neuroimage.2018.01.041)[^1] and
-[Yang et al., 2021](https://doi.org/10.1016/j.neuroimage.2021.118193)[^2]:
+[Liu et al., 2018](https://doi.org/10.1016/j.neuroimage.2018.01.041)[^2] and
+[Yang et al., 2021](https://doi.org/10.1016/j.neuroimage.2021.118193)[^3]:
     - *Temporal Fraction (Fraction of Time):* The proportion of total volumes spent in a single CAP
     over all volumes in a run.
     - *Persistence (Dwell Time):* The average time spent in a single CAP before transitioning to another CAP
@@ -157,12 +157,12 @@ node level of the parcellation.
     - *Transition Probability:* The probability of transitioning from one CAP to another CAP (or the same CAP).
     This is calculated as (Number of transitions from A to B)/(Total transitions from A).
 - **Cosine Similarity Radar Plots:** Create radar plots showing the cosine similarity between
-positive and negative activations of each CAP and each a-priori regions in a parcellation [^3] [^4].
+positive and negative activations of each CAP and each a-priori regions in a parcellation [^4] [^5].
 
 Additional functions in the `neurocaps.analysis` module includes:
 
 - `merge_dicts`: Merge the subject timeseries dictionaries for overlapping subjects across tasks to
-identify similar CAPs across different tasks [^5]. The merged dictionary can be saved as a pickle file.
+identify similar CAPs across different tasks [^6]. The merged dictionary can be saved as a pickle file.
 - `standardize`: Standardizes each run independently for all subjects in the subject timeseries.
 - `change_dtype`: Changes the dtype of all subjects in the subject timeseries to help with memory usage.
 - ``transition_matrix``: Uses the subject-level transition probabilities outputted from the ``CAP``
@@ -174,7 +174,7 @@ documentation website for a more extensive demonstration of the features include
 
 **Demonstration**:
 
-Use dataset from OpenNeuro [^6]:
+Use dataset from OpenNeuro [^7]:
 ```python
 # Download Sample Dataset from OpenNeuro, requires the openneuro-py package
 # [Dataset] doi: doi:10.18112/openneuro.ds005381.v1.0.0
@@ -531,29 +531,33 @@ Please refer the [contributing guidelines](https://neurocaps.readthedocs.io/en/s
 on how to contribute to NeuroCAPs.
 
 ## References
[…]
