### 2025-05-30 32fd60e
Add check for confounds with asterisk
