### 2025-05-30 f8cfdcc
Add check for confounds with asterisk (#32)
