### 2025-06-01 39ac4d1
Update plotly version (#33)
