+++ pyproject.toml
@@ -54,7 +54,7 @@ dependencies = [
     "surfplot>=0.2.0",
     "neuromaps>=0.0.5",
     "pybids>=0.16.5; platform_system != 'Windows'",
-    "plotly>=5.19.0, !=6.1.0, <=6.1.1", # TODO: ``fig.write_image`` issue in 6.1.0
+    "plotly>=5.19.0, !=6.1.0, <=6.1.2", # TODO: ``fig.write_image`` issue in 6.1.0
     "nbformat>=5.10.0", # Needed for plotly figures to display in Jupyter Notebooks without errors
     "kaleido==0.1.0.post1; platform_system == 'Windows'", # Used by plotly for saving plots; TODO: Check if versions > 1.0.0 work on Windows
     "kaleido>=0.2.0, <1.0.0; platform_system != 'Windows'",
