### 2025-06-09 #8196 @yibeichan
Hi @donishadsmith , I've reviewed the repository, and the demo notebooks are working well. Here are some notes/issues
- Regarding parcellations, Scheafer is widely used, but could also consider more atlases that include the subcortex as well. For example, xcp-d takes the fmriprep output as input and extracts time series using different parcellations (https://xcp-d.readthedocs.io/en/latest/outputs.html#parcellations-and-atlases). Here is their AtlasPack (https://github.com/PennLINC/AtlasPack)
- Please add a `.gitignore` file
- Please check all the bullet point formatting in this usage section https://neurocaps.readthedocs.io/en/stable/#usage

I have some comments on the method itself, but that's beyond the scope of this JOSS paper review, but it would be related to your main paper? I hope to have a chance to read your main paper!
