### 2025-06-09 #8196 @yibeichan
@donishadsmith thanks for the explanation. The AtlasPack has an integrated subcortex within the Schaefer (4S series), allowing you to use it directly without modifying too much code. Just keep in mind that the integrated Schaefer parcellations have removed midwall, so might need to be careful with parcel names, etc., but it's not mandatory for proceeding. you can update it whenever you can (e.g., your next package release)

thanks for the method paper! I also read the papers you cited in your documentation; I found some interesting results that align with my HMM results (an ongoing project).

Nice work! hope we get chances to chat more about dynamic brain states
