### 2025-06-09 75db1e4
Add gitignore and fix readthedocs (#35)
