+++ docs/introduction.rst
@@ -87,8 +87,8 @@ Main features for ``CAP`` includes:
 - **Surface Plot Visualization:** Project CAPs onto a surface plot.
 - **Correlation Matrix Creation:** Create a correlation matrix from CAPs.
 - **CAPs Metrics Calculation:** Calculate several CAPs metrics as described in
-`Liu et al., 2018 <https://doi.org/10.1016/j.neuroimage.2018.01.041>`_ [2]_ and
-`Yang et al., 2021 <https://doi.org/10.1016/j.neuroimage.2021.118193>`_ [3]_:
+  `Liu et al., 2018 <https://doi.org/10.1016/j.neuroimage.2018.01.041>`_ [2]_ and
+  `Yang et al., 2021 <https://doi.org/10.1016/j.neuroimage.2021.118193>`_ [3]_:
     - *Temporal Fraction (Fraction of Time):* The proportion of total volumes spent in a single CAP
       over all volumes in a run.
     - *Persistence (Dwell Time):* The average time spent in a single CAP before transitioning to
