### 2025-06-09 #8196 @donishadsmith
@yibeichan I was added the `.gitignore` and fixed the readthedocs bullet issue for https://neurocaps.readthedocs.io/en/stable/#usage in donishadsmith/neurocaps#35 and donishadsmith/neurocaps#36!
