+++ docs/user_guide/installation.rst
@@ -1,5 +1,7 @@
 Installation
 ============
+**NeuroCAPs requires Python 3.9-3.12.**
+
 To install NeuroCAPs, follow the instructions below using your preferred terminal.
 
 Standard Installation from PyPi
