### 2025-06-11 7ee1246
Minor code cleaning
