### 2025-06-13 c0dc92b
[ENH] Improve "Custom" parcel approach
Improves the "Custom" parcel approach by decoupling from only allowing parcellations where all nodes have a left and right hemisphere version.
Allows parcellations not split by hemisphere or that contain a mix of nodes (lateralized and non-lateralized) to be used.

- Remove hemisphere_labels kwargs for `CAP.caps2plot`
- Adds add_custom_node_names kwarg for `CAP.caps2plot`
- Update benchmark to only run commit if submodule hasnt changed
- Fix a case when plots are saved and whitespace is not replaced with
  underscores
