+++ neurocaps/_utils/__init__.py
@@ -1,5 +1,9 @@
 from .check_kwargs import _check_kwargs
-from .check_parcel_approach import _check_parcel_approach, _collapse_aal_node_names
+from .check_parcel_approach import (
+    _check_parcel_approach,
+    _collapse_aal_node_names,
+    _extract_custom_region_indices,
+)
 from .logger import _logger
 from .analysis import _CAPGetter, _build_tree, _cap2statmap, _get_target_indices, _run_kmeans
 from .extraction import (
