### 2025-06-13 61f8de8
[ENH] Improve "Custom" parcel approach to support lateralized and non-lateralized regions (#42)
Improves the "Custom" parcel approach by decoupling from only allowing parcellations where all nodes have a left and right hemisphere version. Allows parcellations not split by hemisphere or that contain a mix of nodes (lateralized and non-lateralized) to be used.

- Remove hemisphere_labels kwargs for `CAP.caps2plot`
- Adds add_custom_node_names kwarg for `CAP.caps2plot`
- Update benchmark to only run commit if submodule hasn't changed
- Fixes a case when plots are saved and whitespace is not repla
[…]
