+++ neurocaps/_utils/check_parcel_approach.py
@@ -1,12 +1,14 @@
 """Internal function for checking the validity of parcel_approach."""
 
 import os, re
+from typing import Any
 
 import numpy as np
 from nilearn import datasets
 
 import neurocaps._utils.io as io_utils
 from .logger import _logger
+from ..typing import ParcelApproach
 
 LG = _logger(__name__)
 
@@ -15,13 +17,24 @@ VALID_DICT_STUCTURES = {
     "AAL": {"version": "SPM12"},
     "Custom": {
         "maps": "/location/to/parcellation.nii.gz",
-        "nodes": ["LH_Vis1", "LH_Vis2", "LH_Hippocampus", "RH_Vis1", "RH_Vis2", "RH_Hippocampus"],
-        "regions": {"Vis": {"lh": [0, 1], "rh": [3, 4]}, "Hippocampus": {"lh": [2], "rh": [5]}},
+        "nodes": [
+            "LH_Vis1",
+            "LH_Vis2",
+            "LH_Hippocampus",
+            "RH_Vis1",
+            "RH_Vis2",
+            "RH_Hippocampus",
+            "Cerebellum_1",
+        ],
+        "regions": {
+            "Vis": {"lh": [0, 1], "rh": [3, 4]},
+            "Hippocampus": {"lh": [2], "rh": [5]},
+            "Cerebellum": [6],
+        },
     },
 }
 
 
-# TODO: Incorporate https://github.com/PennLINC/AtlasPack
 def _check_parcel_approach(parcel_approach, call="TimeseriesExtractor"):
     """
     Pipeline to ensure ``parcel_approach`` is valid and process the ``parcel_approach`` if certain
@@ -206,66 +219,98 @@ def _check_custom_structure(custom_parcel, custom_example):
     example_msg = f"Refer to example: {custom_example}"
 
     if "nodes" in custom_parcel:
-        if not (
-            isinstance(custom_parcel.get("nodes"), (list, np.ndarray))
-            and list(custom_parcel.get("nodes"))
-        ):
-            raise TypeError(
-                "The 'nodes' subkey must be a non-empty list or numpy array containing the "
-                f"node labels. {example_msg}"
-            )
-
-        if not all(isinstance(element, str) for element in custom_parcel["nodes"]):
-            raise TypeError(
-                "All elements in the 'nodes' subkey's list or numpy array must be a string. "
-                f"{example_msg}"
-            )
+        _check_custom_nodes(custom_parcel, example_msg)
 
     if "regions" in custom_parcel:
-        if not (custom_parcel.get("regions") and isinstance(custom_parcel.get("regions"), dict)):
-            raise TypeError(
-                "The 'regions' subkey must be a non-empty dictionary containing the "
-                f"regions/networks labels. {example_msg}"
+        _check_custom_regions(custom_parcel, example_msg)
+
+
+def _check_custom_nodes(custom_parcel, example_msg):
+    """Check structure of the "nodes" subkey."""
+    if not (
+        isinstance(custom_parcel.get("nodes"), (list, np.ndarray))
+        and list(custom_parcel.get("nodes"))
+    ):
+        raise TypeError(
+            "The 'nodes' subkey must be a non-empty list or numpy array containing the "
+            f"node labels. {example_msg}"
+        )
+
+    if not all(isinstance(element, str) for element in custom_parcel["nodes"]):
+        raise TypeError(
+            "All elements in the 'nodes' subkey's list or numpy array must be a string. "
+            f"{example_msg}"
+        )
+
+
+def _check_custom_regions(custom_parcel, example_msg):
+    """Check structure of the "regions" subkey."""
+    if not (custom_parcel.get("regions") and isinstance(custom_parcel.get("regions"), dict)):
[…]
