+++ docs/introduction.rst
@@ -62,7 +62,7 @@ Main features for ``TimeseriesExtractor`` includes:
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 - **Timeseries Extraction:** Extract timeseries for resting-state or task data using Schaefer, AAL,
-  or a lateralized Custom parcellations (which can be manually defined) for spatial dimensionality
+  or a Custom parcellation (which can be manually defined) for spatial dimensionality
   reduction.
 - **Parallel Processing:** Parallelize at the subject-level (one subject per CPU core) to speed up
   timeseries extraction.
