### 2025-06-19 donishadsmith/neurocaps#44 @chiarammatti
Clustering method
### New Feature

Hi,
Thank you for the great work on the repo, it's extremely helpful!
I have a question regarding the choice of using eucledian distance as a metric when performing k-means clustering in `get_caps()`. Why are you not using `1 - pearson correlation` as distance metric as it is normally reported in CAPs analysis papers? Is there a way to choose the type of distance to use?

### Use Case

_No response_

### Proposed Solution

_No response_
