### 2025-06-21 2df61c0
Minor internal cleaning (#45)
* Minor internal cleaning

* [skip ci] Add badge
