### 2025-06-21 2c9b87a
Update type hint (#47)
