+++ pyproject.toml
@@ -56,8 +56,8 @@ dependencies = [
     "pybids>=0.16.5; platform_system != 'Windows'",
     "plotly>=5.19.0, !=6.1.0, <=6.1.2", # TODO: ``fig.write_image`` issue in 6.1.0
     "nbformat>=5.10.0", # Needed for plotly figures to display in Jupyter Notebooks without errors
-    "kaleido==0.1.0.post1; platform_system == 'Windows'", # Used by plotly for saving plots; TODO: Check if versions > 1.0.0 work on Windows
-    "kaleido>=0.2.0, <1.0.0; platform_system != 'Windows'",
+    "kaleido==1.0.0; platform_system == 'Windows'", # Used by plotly for saving plots; TODO: Check if versions > 1.0.0 work on Windows
+    "kaleido>=1.0.0, <1.1.0; platform_system != 'Windows'",
     "setuptools>=77.0.1; python_version>='3.12'", # Results in installation errors if not included as dependency
     "typing_extensions>=4.10.0",
     "vtk>=9.2.0, <9.4.0", # Version >= 9.4.0 works for 3.13 but incompatible with brainspace due no  "__vtkname__" in Pointset
