### 2025-06-26 #8196 @robbisg
Hi,

I reviewed the package and I have some question and improvement to be performed.

1. I think that the documentation should be improved. For example, a full example with real data should be included to the Tutorials section.
2. I found also the API documentation description of CAP class and methods can be improved.
3. The guidelines for seeking help and report issues.
4. I think that it could be also useful to allow the users to modify the clustering algorithm, despite this is inline with the methodology used in the original paper. But this may be discussed in the JOSS paper.
5. I would also add references in the documentation that can be clicked to access other linked classes and definitions. For example, the CAP class references ParcelConfig but I cannot reach this page directly.
6. I would add that an example on how to use custom parcellation (a function to easily get the dictionary used in custom parcellation would be appreciated.)
7. I think that caps2plot function is not easily interpretable. This must be improved both in the documentation and in the plot.
8. The same is for caps2radar, what is the neuroscientific interpretation of this plot. How this can be interpreted.
9. I would also explicitly mention that parcels and images must be in the MNI standard space.
10. I am a bit concerned with function naming and signatures. But this may be out of the scope of the JOSS paper.
11. The paper should include a typical workflow or scenario in which to use NeuroCAPS.
