+++ neurocaps/analysis/cap.py
@@ -43,6 +43,11 @@ class CAP(_CAPGetter):
     Performs k-means clustering for CAP identification, computes temporal dynamics metrics, and
     provides visualization tools for analyzing brain activation patterns.
 
+    .. important::
+       Parcellation maps are expected to be in a standard MNI space. This is due to the
+       ``CAP.caps2surf`` assuming that the parcellation map is in MNI standard space when
+       transforming data between volumetric and surface spaces.
+
     Parameters
     ----------
     parcel_approach: :obj:`ParcelConfig`, :obj:`ParcelApproach`, or :obj:`str`, default=None
@@ -380,8 +385,8 @@ class CAP(_CAPGetter):
             Occurs when ``cluster_selection_method`` is set to elbow but kneed's ``KneeLocator``
             does not detect an elbow in the convex curve.
 
-        Notes
-        -----
+        Note
+        ----
         **KMeans Algorithm:** Refer to `scikit-learn's Documentation
         <https://scikit-learn.org/stable/modules/generated/sklearn.cluster.KMeans.html>`_ for
         additional information about the ``KMeans`` algorithm used in this method.
@@ -487,7 +492,7 @@ class CAP(_CAPGetter):
         Sets ``groups`` to None to allow ``self.get_caps`` to create a new "All Subjects" group
         with different subject IDs based on the inputted ``subject_timeseries``.
 
-        .. version_added:: 0.28.5
+        .. versionadded:: 0.28.5
         """
         self._groups = None
 
@@ -832,10 +837,11 @@ class CAP(_CAPGetter):
         progress_bar: bool = False,
     ) -> Union[dict[str, pd.DataFrame], None]:
         """
-        Compute Participant-wise CAP Metrics.
+        Calculate Participant-wise CAP Metrics.
 
-        Computes the following temporal dynamic metrics (as described by Liu et al., 2018 and
-        Yang et al., 2021):
+        Calculates temporal dynamic metrics for each CAP across all participants. The metrics are
+        calculated as described by Liu et al. (2018) and Yang et al. (2021) and include the
+        following:
 
          - "temporal_fraction" (fraction of time): Proportion of total volumes spent in a single CAP
            over all volumes in a run.
@@ -985,9 +991,13 @@ class CAP(_CAPGetter):
         Returns
         -------
         dict[str, pd.DataFrame] or dict[str, dict[str, pd.DataFrame]]
-            Dictionary containing `pandas.DataFrame` - one for each requested metric. In the case of
-            "transition_probability", each group has a separate dataframe which is returned in the
-            from of `dict[str, dict[str, pd.DataFrame]]`. Only returned if ``return_df`` is True.
+            Dictionary containing `pandas.DataFrame`: Only returned if ``return_df`` is True.
+
+            - For "temporal_fraction", "counts", "persistence", and "transition_frequency": one
+            dataframe is returned for each requested metric.
+
+            - For "transition_probability": each group has a separate dataframe which is returned
+            in the from of `dict[str, dict[str, pd.DataFrame]]`.
 
 
         Important
@@ -1599,8 +1609,9 @@ class CAP(_CAPGetter):
         """
         Generate Heatmaps and Outer Product Plots for CAPs.
 
-        Plot CAPs as heatmaps or outer products at the node or region levels. Separate plots are
-        generated for each group.
+        Produces a 2D plot for each CAP, visualized as either a heatmap or an outer product.
+        The visualization can be generated at two levels of granularity (nodes or regions).
+        Separate plots are generated for each group.
 
         Parameters
         ----------
@@ -1614,14 +1625,31 @@ class CAP(_CAPGetter):
[…]
