### 2025-06-27 f46ab03
Improve AAL default regions
