### 2025-06-28 02c4ad5
Improve AAL default regions (#51)
