### 2025-06-30 a9022f7
Update dependency plotly to >=6.2.0, !=6.1.0, <=6.2.0
