### 2025-07-01 donishadsmith/neurocaps#44 @donishadsmith
Hi @chiarammatti, I'm sorry for the late response since I just saw your response!

Funny enough, I was actually working on code that standardizes across the ROIs exactly has you mentioned (row-wise) since the current implementation only allows for column-wise standardization. Yes, as you mentioned this would emphasize the shape of the pattern itself and would better approximate Pearson correlation while still using Euclidean distance. I think this will be a more straightforward solution than a custom KMeans function primarily due to have to consider speed and edge cases from a coding perspective (though it's still on the table for some future release if I can determine a clean way to implement it).

Thank you for the excellent reference paper! I can actually release the latest version (0.31.0) really soon because the code is already in place and is already tested. I would just need to update the documentation and check a few more things. I will add that paper to the reference in the documentation.
