### 2025-07-01 donishadsmith/neurocaps#44 @donishadsmith
Hi @chiarammatti, I'm sorry for the late response!

Yes, as you mentioned I believe that this would emphasize the shape of the pattern itself and would better approximate Pearson correlation while still using Euclidean distance. I am considering whether to add that type of standardization as a feature or go the route of a custom KMeans solution (I believe the former is the most straightforward approach). Currently, I have experimental code that allows for row-wise standardization, while also ensuring this standardization is also done prior to predicting cluster assignment of each TR (`return_cap_labels`) to ensure accuracy for the temporal metric computations  (`calculate_metrics`). It's tested, though I need to update documentation and determine the best way to implement this feature to ensure API consistency and clarity.

Thank you for the excellent reference paper! There is a possibility that I can release this code in a future version. Since, this software is currently under review in JOSS, I will first ask the reviewers for recommendations to see if this would be a good path to take. I want to ensure that this new code will not be too disruptive and the proper deprecations are taken. If they ok the code, I can add it to a future release.
