### 2025-07-01 #8196 @donishadsmith
Hey @robbisg and @yibeichan!

I wanted to ask for your input on a possible enhancement that I can include along with the feedback both of you provided for my package. A user requested an enhancement (donishadsmith/neurocaps#44) to allow k-means with Euclidean distance to [better approximate correlation distance](https://arxiv.org/abs/1601.02213), similar to MATLAB's use of Pearson correlation distance in k-means.

A possible implementation could involve row-wise standardization (normalizing TRs) of the concatenated matrix prior to clustering. This could be incorporated into the package by doing the following:
1. Add a "standardize_trs" parameter to `CAP.get_caps`
2. Apply this transformation before predicting CAP labels and computing metrics to ensure correctness

My primary concern is API clarity. Currently, the "standardize" parameter performs column-wise (ROI) standardization and appears in: `TimeseriesExtractor`, `CAP.get_caps`, and the standalone function `neurocaps.analysis.standardize`. I wanted to ask if having both a "standardize" (columns/ROIs) and "standardize_trs" (rows/TRs) parameter in `CAP.get_caps` be too confusing?

For clarity, my proposed solutions for how to handle this enhancement are the following:
1. Keep the "standardize" parameter and simply add "standardize_trs" to `CAP.get_caps`
2. Replace the "standardize" parameter with "standardize_rois" in `CAP.get_caps`  (with deprecation warning) while keeping the "standardize" parameter in `TimeseriesExtractor` and `neurocaps.analysis.standardize`
3. Change all instances of "standardize" to "standardize_rois" throughout the package

Thanks for your input!
