### 2025-07-03 2e90c75
Initial refactor work
