+++ neurocaps/analysis/cap/_internals/matrix.py
@@ -1,0 +1,433 @@
+def check_cap_length(cap_dict, parcel_approach):
+    cap_dict = cap_dict[list(cap_dict)[0]]
+    cap_1_vector = cap_dict[list(cap_dict)[0]]
+    # Get parcellation name
+    parc_name = get_parc_name(parcel_approach)
+    if cap_1_vector.shape[0] != len(parcel_approach[parc_name]["nodes"]):
+        raise ValueError(
+            "Number of nodes used for CAPs does not equal the number of nodes specified in "
+            "`parcel_approach`."
+        )
+
+
+def _compute_region_means(self, parcellation_name: str) -> None:
+    """
+    Creates an attribute called ``self._region_means``, representing the average values of
+    all nodes in a corresponding region to create region heatmaps or outer product plots.
+    """
+    self._region_means = {group: {} for group in self._groups}
+
+    # List of regions remains list for Schaefer and AAL but converts keys to list for Custom
+    regions = list(self._parcel_approach[parcellation_name]["regions"])
+
+    group_caps = [(group, cap) for group in self._groups for cap in self._caps[group]]
+    for group, cap in group_caps:
+        region_means = None
+        for region in regions:
+            if parcellation_name != "Custom":
+                region_indxs = np.array(
+                    [
+                        index
+                        for index, node in enumerate(
+                            self._parcel_approach[parcellation_name]["nodes"]
+                        )
+                        if region in node
+                    ]
+                )
+            else:
+                region_indxs = np.array(
+                    extract_custom_region_indices(self._parcel_approach, region)
+                )
+
+            if region_means is None:
+                region_means = np.array([np.average(self._caps[group][cap][region_indxs])])
+            else:
+                region_means = np.hstack(
+                    [region_means, np.average(self._caps[group][cap][region_indxs])]
+                )
+
+        # Append regions and their means
+        self._region_means[group].update({"Regions": regions})
+        self._region_means[group].update({cap: region_means})
+
+
+def _extract_scope_information(
+    self, scope: str, parcellation_name: str, add_custom_node_labels: bool
+) -> tuple[dict[str, dict[str, NDArray]], list[str]]:
+    """
+    Extracting region means of each CAP from ``self._region_means`` if scope is "region" else
+    extracts the CAP vectors from ``self._caps``. Also extracts region labels or nodes
+    from the ``parcel_approach``.
+    """
+    if scope == "regions":
+        cap_dict = {
+            group: {k: v for k, v in self._region_means[group].items() if k != "Regions"}
+            for group in self._region_means
+        }
+        labels = list(self._parcel_approach[parcellation_name]["regions"])
+    elif scope == "nodes":
+        cap_dict = self._caps
+        labels = self._extract_node_names(parcellation_name, add_custom_node_labels)
+
+    return cap_dict, labels
+
+
+def _extract_node_names(
+    self, parcellation_name: str, add_custom_node_labels: bool
+) -> tuple[dict[str, dict[str, NDArray]], list[str]]:
+    """
+    Extracts the node names from the ``parcel_approach``. For "Custom", if the region is
+    lateralized, then the label is returned in the form "{Hemisphere} {Region}", else its
+    returned as "{Region}".
+    """
+    if parcellation_name in ["Schaefer", "AAL"]:
+        labels = self._parcel_approach[parcellation_name]["nodes"]
+    else:
[…]
