+++ neurocaps/analysis/cap/_internals/metrics.py
@@ -1,0 +1,307 @@
+"""Internal module for computing temporal dynamic metrics."""
+
+import itertools, os
+from typing import Union
+
+import numpy as np
+import pandas as pd
+from numpy.typing import NDArray
+
+from neurocaps._utils import io as io_utils
+from neurocaps._utils.logging import setup_logger
+
+LG = setup_logger(__name__)
+
+
+def filter_metrics(metrics: Union[list[str], tuple[str], None]) -> list[str]:
+    """
+    Filters metrics to ensure only the supported metrics ("temporal_fraction", "persistence",
+    "counts", "transition_frequency") are in the list. Maintains the order of the original
+    user-specified metrics.
+    """
+    metrics = (
+        ("temporal_fraction", "persistence", "counts", "transition_frequency")
+        if metrics is None
+        else metrics
+    )
+    metrics = [metrics] if isinstance(metrics, str) else metrics
+    metrics_set = set(metrics)
+
+    valid_metrics = {
+        "temporal_fraction",
+        "persistence",
+        "counts",
+        "transition_frequency",
+        "transition_probability",
+    }
+    set_diff = metrics_set - valid_metrics
+    metrics_set = metrics_set.intersection(valid_metrics)
+    # Ensure original order maintained
+    ordered_metrics = [metric for metric in metrics if metric in metrics_set]
+
+    if set_diff:
+        formatted_string = ", ".join(["'{a}'".format(a=x) for x in set_diff])
+        LG.warning(f"The following invalid metrics will be ignored: {formatted_string}.")
+
+    if not ordered_metrics:
+        formatted_string = ", ".join(["'{a}'".format(a=x) for x in valid_metrics])
+        raise ValueError(
+            f"No valid metrics in `metrics` list. Valid metrics are: {formatted_string}."
+        )
+
+    return ordered_metrics
+
+
+def extract_caps_info(
+    cap_dict: dict[str, dict[str, NDArray]],
+) -> tuple[list[str], int, dict[str, int]]:
+    """
+    Extracts CAP-related information, specifically the names of the CAPs (i.e. CAP-1, CAP-2,
+    etc), the maximum number of CAPs found across all groups (e.g if Group A has 4 CAPs and
+    group B has 5 CAPs then 5 is returned), and the number of CAPs for each group.
+    """
+    group_cap_counts = {}
+
+    for group_name in cap_dict:
+        # Store the length of caps in each group
+        group_cap_counts.update({group_name: len(cap_dict[group_name])})
+
+    # CAP names based on groups with the most CAPs
+    cap_names = list(cap_dict[max(group_cap_counts, key=group_cap_counts.get)])
+    max_cap = max([int(name.split("-")[-1]) for name in cap_names])
+
+    return cap_names, max_cap, group_cap_counts
+
+
+def create_transition_pairs(
+    cap_dict: dict[str, dict[str, NDArray]],
+) -> dict[str, list[tuple[int, int]]]:
+    """Obtains all possible transition pairs."""
+    group_caps = {}
+    all_pairs = {}
+
+    for group_name in cap_dict:
+        group_caps.update({group_name: [int(name.split("-")[-1]) for name in cap_dict[group_name]]})
+        all_pairs.update(
+            {group_name: list(itertools.product(group_caps[group_name], group_caps[group_name]))}
+        )
+
+    return all_pairs
+
+
+def build_df(
+    metrics: list[str],
+    group_names: list[str],
+    cap_names: list[str],
+    pairs: dict[str, list[tuple[int, int]]],
+) -> dict[str, pd.DataFrame]:
[…]
