### 2025-07-03 e745c40
Additional refactoring
