+++ neurocaps/analysis/cap/_internals/matrix.py
@@ -1,3 +1,24 @@
+"""Internal module containing helper functions related to ``CAP.caps2plot``."""
+
+import collections, re
+from typing import Any, Union
+
+import numpy as np
+import matplotlib.pyplot as plt
+import pandas as pd
+import seaborn
+from numpy.typing import NDArray
+
+from neurocaps.typing import ParcelApproach
+from neurocaps._utils import io as io_utils
+from neurocaps._utils.parcellation import (
+    collapse_aal_node_names,
+    extract_custom_region_indices,
+    get_parc_name,
+)
+from neurocaps._utils.plotting_utils import PlotFuncs
+
+
 def check_cap_length(cap_dict, parcel_approach):
     cap_dict = cap_dict[list(cap_dict)[0]]
     cap_1_vector = cap_dict[list(cap_dict)[0]]
@@ -10,91 +31,100 @@ def check_cap_length(cap_dict, parcel_approach):
         )
 
 
-def _compute_region_means(self, parcellation_name: str) -> None:
+def compute_region_means(
+    parcel_approach: ParcelApproach, group_names: list[str], cap_dict: dict[str, dict[str, NDArray]]
+) -> dict[str, dict[str, NDArray]]:
     """
     Creates an attribute called ``self._region_means``, representing the average values of
     all nodes in a corresponding region to create region heatmaps or outer product plots.
     """
-    self._region_means = {group: {} for group in self._groups}
+    region_means_dict = {group_name: {} for group_name in group_names}
 
     # List of regions remains list for Schaefer and AAL but converts keys to list for Custom
-    regions = list(self._parcel_approach[parcellation_name]["regions"])
+    parc_name = get_parc_name(parcel_approach)
+    regions = list(parcel_approach[parc_name]["regions"])
 
-    group_caps = [(group, cap) for group in self._groups for cap in self._caps[group]]
-    for group, cap in group_caps:
+    group_caps_names = [
+        (group_name, cap_name) for group_name in group_names for cap_name in cap_dict[group_name]
+    ]
+    for group_name, cap_name in group_caps_names:
         region_means = None
         for region in regions:
-            if parcellation_name != "Custom":
+            if parc_name != "Custom":
                 region_indxs = np.array(
                     [
                         index
-                        for index, node in enumerate(
-                            self._parcel_approach[parcellation_name]["nodes"]
-                        )
+                        for index, node in enumerate(parcel_approach[parc_name]["nodes"])
                         if region in node
                     ]
                 )
             else:
-                region_indxs = np.array(
-                    extract_custom_region_indices(self._parcel_approach, region)
-                )
+                region_indxs = np.array(extract_custom_region_indices(parcel_approach, region))
 
             if region_means is None:
-                region_means = np.array([np.average(self._caps[group][cap][region_indxs])])
+                region_means = np.array([np.average(cap_dict[group_name][cap_name][region_indxs])])
             else:
                 region_means = np.hstack(
-                    [region_means, np.average(self._caps[group][cap][region_indxs])]
+                    [region_means, np.average(cap_dict[group_name][cap_name][region_indxs])]
                 )
 
         # Append regions and their means
-        self._region_means[group].update({"Regions": regions})
-        self._region_means[group].update({cap: region_means})
+        region_means_dict[group_name].update({"Regions": regions})
+        region_means_dict[group_name].update({cap_name: region_means})
+
[…]
