### 2025-07-03 5b6829f
Refactor timeseries_extractor
