+++ neurocaps/analysis/cap/_internals/correlation.py
@@ -1,13 +1,14 @@
 """Internal module containing helper functions for ``CAP.caps2corr``."""
 
 import numpy as np
+from pandas import DataFrame
 from scipy.stats import pearsonr, spearmanr
 
 # Dictionary of scipy correlation functions
 CORR_FUNC = {"pearson": pearsonr, "spearman": spearmanr}
 
 
-def add_significance_values(df, corr_df, method, fmt):
+def add_significance_values(df: DataFrame, corr_df: DataFrame, method: str, fmt: str) -> DataFrame:
     """Add p-values to each correlation value in the dataframe."""
     # Get p-values; use np.eye to make main diagonals equal zero; implementation of
     # tozCSS from:
