### 2025-07-04 0fb8a39
Remove staticmethod from functions
- Interesting note that only the macOS runner failed due to having the
  static method decorator on a non-class function
