### 2025-07-04 04eda4c
Internal code refactoring (#53)
- Significant internal refactoring to clean the code in the public API, changes include:
    - Moving files in `neurocaps._utils.extraction` and `neurocaps._utils.analysis` to internal
    folders in `neurocaps.extraction` and `neurocaps.analysis`
    - Changing internal function names to be more descriptive
    - Moving certain internal functions in `TimeseriesExtractor` and `CAP` into separate internal
    modules
    - Changing internal file names
