+++ neurocaps/extraction/timeseries_extractor.py
@@ -1,36 +1,29 @@
 """Contains the TimeseriesExtractor class for extracting timeseries"""
 
-import json, os, re
+import os
 from functools import lru_cache
 from multiprocessing.queues import Queue
 from typing import Any, Literal, Optional, Union
 from typing_extensions import Self
 
-import matplotlib.pyplot as plt, numpy as np
 from joblib import Parallel, delayed
-from numpy.typing import NDArray
 from pandas import DataFrame
 from tqdm.auto import tqdm
 
 import neurocaps._utils.io as io_utils
-from ..exceptions import BIDSQueryError
-from ..typing import ParcelConfig, ParcelApproach, SubjectTimeseries
-from .._utils import (
-    _TimeseriesExtractorGetter,
-    _PlotDefaults,
-    _PlotFuncs,
-    _check_kwargs,
-    _check_confound_names,
-    _check_parcel_approach,
-    _extract_custom_region_indices,
-    _extract_timeseries,
-    _logger,
-)
+from ._internals import check_confound_names, process_subject_runs, TimeseriesExtractorGetter
+from ._internals import bids_query, vizualization
+from neurocaps.exceptions import BIDSQueryError
+from neurocaps.typing import ParcelConfig, ParcelApproach, SubjectTimeseries
+from neurocaps._utils.helpers import resolve_kwargs
+from neurocaps._utils.logging import setup_logger
+from neurocaps._utils.parcellation import check_parcel_approach
+from neurocaps._utils.plotting_utils import PlotDefaults, PlotFuncs
 
-LG = _logger(__name__)
+LG = setup_logger(__name__)
 
 
-class TimeseriesExtractor(_TimeseriesExtractorGetter):
+class TimeseriesExtractor(TimeseriesExtractorGetter):
     """
     Timeseries Extractor Class.
 
@@ -281,12 +274,12 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
 
         self._space = space
         # Check parcel_approach
-        self._parcel_approach = _check_parcel_approach(parcel_approach=parcel_approach)
+        self._parcel_approach = check_parcel_approach(parcel_approach=parcel_approach)
 
         if use_confounds:
             if confound_names:
                 # Replace confounds if not None
-                confound_names = _check_confound_names(
+                confound_names = check_confound_names(
                     high_pass, confound_names, n_acompcor_separate
                 )
 
@@ -642,10 +635,8 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
         self._validate_get_bold_params()
 
         # Initialize new attributes
-        self._subject_ids = []
         self._subject_timeseries = {}
         self._qc = {}
-        self._subject_info = {}
         self._n_cores = n_cores
 
         layout = self._call_layout(bids_dir, pipeline_name)
@@ -687,7 +678,15 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
             subj_ids = sorted([subj_id for subj_id in subj_ids if subj_id in run_subjects])
 
         # Setup extraction
-        self._setup_extraction(layout, subj_ids, exclude_niftis, verbose)
+        self._subject_ids, self._subject_info = bids_query.setup_extraction(
+            layout,
+            subj_ids,
+            self._space,
+            exclude_niftis,
+            self._signal_clean_info,
+            self._task_info,
+            verbose,
+        )
 
         if self._n_cores:
             # Generate list of tuples for each subject
@@ -709,7 +708,7 @@ class TimeseriesExtractor(_TimeseriesExtractorGetter):
 
             parallel = Parallel(return_as="generator", n_jobs=self._n_cores, backend="loky")
             outputs = tqdm(
[…]
