### 2025-07-04 8e2895c
Update badge [skip ci]
