### 2025-07-04 634d910
Improve dataframe creation when computing metrics
