+++ neurocaps/analysis/cap/_internals/metrics.py
@@ -68,7 +68,7 @@ def extract_caps_info(
 
     # CAP names based on groups with the most CAPs
     cap_names = list(cap_dict[max(group_cap_counts, key=group_cap_counts.get)])
-    max_cap = max([int(name.split("-")[-1]) for name in cap_names])
+    max_cap = max(group_cap_counts.values())
 
     return cap_names, max_cap, group_cap_counts
 
@@ -77,40 +77,67 @@ def create_transition_pairs(
     cap_dict: dict[str, dict[str, NDArray]],
 ) -> dict[str, list[tuple[int, int]]]:
     """Obtains all possible transition pairs."""
-    group_caps = {}
-    all_pairs = {}
-
-    for group_name in cap_dict:
-        group_caps.update({group_name: [int(name.split("-")[-1]) for name in cap_dict[group_name]]})
-        all_pairs.update(
-            {group_name: list(itertools.product(group_caps[group_name], group_caps[group_name]))}
-        )
-
-    return all_pairs
+    group_caps = {
+        group_name: [get_cap_id(cap_name) for cap_name in cap_dict[group_name]]
+        for group_name in cap_dict
+    }
+    return {
+        group_name: list(itertools.product(cap_id, cap_id))
+        for group_name, cap_id in group_caps.items()
+    }
 
 
-def build_df(
+def get_cap_id(cap_name):
+    """
+    Extracts the CAP name, assumed to be in the form `CAP-{n}` (i.e. CAP-1) and returns the int.
+    """
+    return int(cap_name.split("-")[-1])
+
+
+def create_columns_names(
     metrics: list[str],
     group_names: list[str],
     cap_names: list[str],
     pairs: dict[str, list[tuple[int, int]]],
-) -> dict[str, pd.DataFrame]:
-    """Initializes the output dataframes with column names for each requested metric."""
-    df_dict = {}
+) -> dict[str, Union[list[str], dict[str, list[str]]]]:
+    """
+    Creates the column names for each requested metric. Used downstream has the column
+    names for each metrics dataframe.
+    """
+    columns_names_dict = {}
     base_cols = ["Subject_ID", "Group", "Run"]
 
     for metric in metrics:
         if metric not in ["transition_frequency", "transition_probability"]:
-            df_dict.update({metric: pd.DataFrame(columns=base_cols + list(cap_names))})
+            columns_names_dict.update({metric: base_cols + list(cap_names)})
         elif metric == "transition_probability":
-            df_dict[metric] = {}
-            for group in group_names:
-                col_names = base_cols + [f"{x}.{y}" for x, y in pairs[group]]
-                df_dict[metric].update({group: pd.DataFrame(columns=col_names)})
+            columns_names_dict[metric] = {}
+            for group_name in group_names:
+                col_names = base_cols + [f"{x}.{y}" for x, y in pairs[group_name]]
+                columns_names_dict[metric].update({group_name: col_names})
         else:
-            df_dict.update({metric: pd.DataFrame(columns=base_cols + ["Transition_Frequency"])})
+            columns_names_dict.update({metric: base_cols + ["Transition_Frequency"]})
 
-    return df_dict
+    return columns_names_dict
+
+
+def initialize_all_metrics_dict(
+    metrics: list[str], group_names: list[str]
+) -> dict[str, Union[list, dict[str, list]]]:
+    """
+    Initializes a dictionary intended to store all computations for a metric across all
+    subjects. The dictionary will be converted to a dataframe downstream.
+    """
+    all_metrics_dict = {}
+    for metric in metrics:
+        if metric != "transition_probability":
[…]
