### 2025-07-05 fe40c9c
Add whitespace and type hint [skip ci]
