### 2025-07-05 1adc681
Minor refactor
