+++ neurocaps/__init__.py
@@ -1,7 +1,7 @@
-from . import analysis, extraction, exceptions
+from . import analysis, extraction, exceptions, utils
 
 # Don't include "typing"
-__all__ = ["analysis", "extraction", "exceptions"]
+__all__ = ["analysis", "extraction", "exceptions", "utils"]
 
 # Version in single place
-__version__ = "0.31.1"
+__version__ = "0.32.0"
