### 2025-07-06 e0eeeeb
Add helper function for custom parcellation & change "_utils" to "utils" (#54)
- Clickable links to type definitions in docs
- New function to create the custom parcel approach (tests added)
- Additional folder name and file name changes (exclusively to contents in utils)
- Updated submodule
- Other docs related updates
