+++ neurocaps/utils/__init__.py
@@ -1,0 +1,3 @@
+from .parcellations import generate_custom_parcel_approach
+
+__all__ = ["generate_custom_parcel_approach"]
