+++ tests/test_utils.py
@@ -1,0 +1,114 @@
+import os, shutil
+
+import pandas as pd
+import pytest
+
+from neurocaps.utils import generate_custom_parcel_approach
+
+
+@pytest.fixture(autouse=False, scope="function")
+def create_empty_nifti_file(tmp_dir):
+    """Creates an empty NifTI file"""
+    filename = os.path.join(tmp_dir.name, "schaefer4s.nii.gz")
+    with open(filename, "w") as f:
+        pass
+
+    yield filename
+
+
+@pytest.fixture(autouse=False, scope="function")
+def copy_test_data(tmp_dir):
+    """Copies dataframe to temporary directory."""
+    tsv_file = os.path.join(tmp_dir.name, "schaefer4s.tsv")
+    shutil.copyfile(os.path.join(os.path.dirname(__file__), "data", "schaefer4s.tsv"), tsv_file)
+
+    tsv_error_file = os.path.join(tmp_dir.name, "schaefer4s_error.tsv")
+    shutil.copyfile(
+        os.path.join(os.path.dirname(__file__), "data", "schaefer4s_error.tsv"), tsv_error_file
+    )
+
+    yield tsv_file, tsv_error_file
+
+
+def test_generate_custom_parcel_approach(copy_test_data, create_empty_nifti_file):
+    """Tests the ``generate_custom_parcel_approach`` with and without hemispheres."""
+    tsv_file, _ = copy_test_data
+    nifti_file = create_empty_nifti_file
+    df = pd.read_csv(tsv_file, sep="\t")
+
+    parcel_approach_no_hemi = generate_custom_parcel_approach(
+        filepath_or_df=tsv_file,
+        maps_path=nifti_file,
+        column_map={"nodes": "label", "regions": "network_label"},
+        background_label="Background",
+    )
+
+    # Check that node labels match the dataframe (minus the background)
+    expected_nodes = df["label"][1:].tolist()
+    assert parcel_approach_no_hemi["Custom"]["nodes"] == expected_nodes
+
+    # Check the regions mapping for non-lateralized case
+    regions = parcel_approach_no_hemi["Custom"]["regions"]
+    assert isinstance(regions["Vis"], list)
+    assert regions["Vis"] == [0, 1, 2, 3, 4, 5, 6, 7, 8, 50, 51, 52, 53, 54, 55, 56, 57]
+
+    assert isinstance(regions["SomMot"], list)
+    assert regions["SomMot"] == [9, 10, 11, 12, 13, 14, 58, 59, 60, 61, 62, 63, 64, 65]
+
+    parcel_approach_with_hemi = generate_custom_parcel_approach(
+        filepath_or_df=tsv_file,
+        maps_path=nifti_file,
+        column_map={
+            "nodes": "label",
+            "regions": "network_label",
+            "hemispheres": "hemisphere_labels",
+        },
+        hemisphere_map={"lh": ["LH"], "rh": ["RH"]},
+        background_label="Background",
+    )
+
+    assert parcel_approach_with_hemi["Custom"]["nodes"] == expected_nodes
+
+    # Check the regions mapping for lateralized case
+    regions_hemi = parcel_approach_with_hemi["Custom"]["regions"]
+    assert isinstance(regions_hemi["Vis"], dict)
+    assert "lh" in regions_hemi["Vis"] and "rh" in regions_hemi["Vis"]
+    assert regions_hemi["Vis"]["lh"] == [0, 1, 2, 3, 4, 5, 6, 7, 8]
+    assert regions_hemi["Vis"]["rh"] == [50, 51, 52, 53, 54, 55, 56, 57]
+
+    assert isinstance(regions_hemi["SomMot"], dict)
[…]
