### 2025-07-06 464572d
Fix return type hint
