+++ neurocaps/__init__.py
@@ -4,4 +4,4 @@ from . import analysis, extraction, exceptions, utils
 __all__ = ["analysis", "extraction", "exceptions", "utils"]
 
 # Version in single place
-__version__ = "0.32.0"
+__version__ = "0.32.1"
