+++ CHANGELOG.md
@@ -13,9 +13,13 @@ noted in the changelog (e.g., new functions or parameters, changes in parameter 
 improvements/enhancements. All fixes and modifications are backwards compatible.
 - *.postN* : Consists of documentation changes or metadata-related updates, such as modifications to type hints.
 
+## [0.32.1] - 2025-07-06
+### 🐛 Fixes
+- Return type hint for `generate_custom_parcel_approach`.
+
 ## [0.32.0] - 2025-07-06
 ### 🚀 New/Added
-- Created "generate_custom_parcel_approach" to help create a "Custom" `parcel_approach`
+- Created `generate_custom_parcel_approach` to help create a "Custom" `parcel_approach`
 for a parcellations metadata file.
 ### 🔧 Refactoring
 - "_utils" is now "utils". Private file in the "utils" folder now start with "underscore".
