### 2025-07-06 a3f83e4
Fix module docstring [skip ci]
