+++ paper/paper.md
@@ -1,12 +1,13 @@
 ---
-title: 'NeuroCAPs: A Python Package for Performing Co-Activation Patterns Analyses on Resting-State and Task-Based fMRI Data'
+title: 'NeuroCAPs: A Python Package for Performing Co-Activation Patterns Analyses on Resting-State
+        and Task-Based fMRI Data'
 tags:
     - Python
     - neuroimaging
     - fMRI
     - dynamic functional connectivity
     - co-activation patterns
-date: 4 April 2025
+date: 7 July 2025
 output: pdf_document
 authors:
     - name: Donisha Smith
@@ -21,114 +22,200 @@ bibliography: paper.bib
 ---
 
 # Summary
-Co-Activation Patterns (CAPs) is a dynamic functional connectivity technique that clusters similar spatial distributions
-of brain activity. To make this analytical technique more accessible to neuroimaging researchers, NeuroCAPs,
-an open source Python package, was developed. This package performs end-to-end CAPs analyses on preprocessed resting-state
-or task-based functional magnetic resonance imaging (fMRI) data, and is most optimized for data preprocessed with
-fMRIPrep, a robust preprocessing pipeline designed to minimize manual user input and enhance reproducibility
+Co-Activation Patterns (CAPs) is a dynamic functional connectivity technique that clusters similar
+spatial distributions of brain activity. To make this analytical technique more accessible to
+neuroimaging researchers, NeuroCAPs, an open source Python package, was developed. This package
+performs end-to-end CAPs analyses on preprocessed resting-state or task-based functional magnetic
+resonance imaging (fMRI) data, and is most optimized for data preprocessed with fMRIPrep, a robust
+preprocessing pipeline designed to minimize manual user input and enhance reproducibility
 [@Esteban2019].
 
 # Background
-Numerous fMRI studies employ static functional connectivity (sFC) techniques to analyze correlative activity within and
-between brain regions. However, these approaches operate under the assumption that functional connectivity patterns,
-which change within seconds [@Jiang2022], remain stationary throughout the entire data acquisition period
-[@Hutchison2013].
+Numerous fMRI studies employ static functional connectivity (sFC) techniques to analyze correlative
+activity within and between brain regions. However, these approaches operate under the assumption
+that functional connectivity patterns, which change within seconds [@Jiang2022], remain stationary
+throughout the entire data acquisition period [@Hutchison2013].
 
-Unlike sFC approaches, dynamic functional connectivity (dFC) methods enable the analysis of dynamic functional states,
-which are characterized by consistent, replicable, and distinct periods of time-varying brain connectivity patterns
-[@Rabany2019]. Among these techniques, CAPs analysis aggregates similar spatial distributions of brain activity using
-clustering techniques, typically the k-means algorithm, to capture the dynamic nature of brain activity [@Liu2013; @Liu2018].
+Unlike sFC approaches, dynamic functional connectivity (dFC) methods enable the analysis of dynamic
+functional states, which are characterized by consistent, replicable, and distinct periods of
+time-varying brain connectivity patterns [@Rabany2019]. Among these techniques, CAPs analysis
+aggregates similar spatial distributions of brain activity using clustering techniques, typically
+the k-means algorithm, to capture the dynamic nature of brain activity [@Liu2013; @Liu2018].
 
 # Statement of Need
-Currently, performing an end-to-end CAPs analysis presents challenges due to the numerous steps required. Researchers must:
+The typical CAPs workflow can be programmatically time-consuming to manually orchestrate as it
+generally entails several steps:
 
[…]
