### 2025-07-06 b0dc2c1
Clean grid table [skip ci]
