+++ neurocaps/utils/parcellations.py
@@ -105,7 +105,7 @@ def generate_custom_parcel_approach(
 
     +------------------------+----------+------------+-----+
     | labels                 | networks | hemisphere | ... |
-    +============+===========+==========+============+=====+
+    +========================+==========+============+=====+
     | SalVentAttn_TempOccPar | VAN      | L          | ... |
     +------------------------+----------+------------+-----+
     | SalVentAttn_FrOperIns  | VAN      | R          | ... |
