+++ docs/introduction.rst
@@ -41,15 +41,11 @@
 
 NeuroCAPs (**Neuro**\ imaging **C**\ o-**A**\ ctivation **P**\ attern\ **s**) is a Python package for performing
 Co-Activation Patterns (CAPs) analyses on resting-state or task-based fMRI data. CAPs identifies recurring brain states
-by applying k-means clustering on BOLD timeseries data [1]_ [2]_.
+by applying k-means clustering on BOLD timeseries data [1]_.
 
 .. image:: assets/workflow.png
    :width: 70%
    :align: center
-
-**Note:** NeuroCAPs is most optimized for fMRI data preprocessed with `fMRIPrep <https://fmriprep.org/en/stable/>`_ and
-assumes the data is BIDs compliant. Refer to `NeuroCAPs' BIDS Structure and Entities Documentation <https://neurocaps.readthedocs.io/en/stable/bids.html>`_
-for additional information.
 
 Citing
 ------
@@ -59,68 +55,46 @@ Citing
 
 Usage
 -----
-This package contains two main classes: ``TimeseriesExtractor`` for extracting the timeseries, and
-``CAP`` for performing the CAPs analysis.
+NeuroCAPs is built around two main classes (``TimeseriesExtractor`` and ``CAP``) and includes several
+features to perform the complete CAPs workflow from postprocessing to visualizations.
+Notable features includes:
 
-Main features for ``TimeseriesExtractor`` includes:
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+- Timeseries Extraction (``TimeseriesExtractor``):
 
-- **Timeseries Extraction:** Extract timeseries for resting-state or task data using Schaefer, AAL,
-  or a manually defined deterministic "Custom" parcellation for spatial dimensionality
-  reduction.
-- **Parallel Processing:** Parallelize at the subject-level (one subject per CPU core) to speed up
-  timeseries extraction.
-- **Saving Timeseries:** Save the nested dictionary containing timeseries (mapping subject id
-  -> run id -> timeseries data) as a pickle file.
-- **Reporting Quality Control:** Reports statistics related to framewise displacement and dummy
-  volumes per-subject.
-- **Visualization:** Visualize the timeseries at the region or node level of the parcellation for a
-  given subject and run.
+  - Extracts BOLD timeseries from resting-state or task-based fMRI data.
+  - Supports deterministic parcellations such as the Schaefer and AAL, in addition to custom-defined deterministic parcellations.
+  - Performs nuisance regression, motion scrubbing, and additional features
+  - Reports quality control information based on framewise displacement
 
-Main features for ``CAP`` includes:
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+  .. important::
+      NeuroCAPs is most optimized for fMRI data preprocessed with
+      [fMRIPrep](https://fmriprep.org/en/stable/) and assumes the data is BIDs compliant.
+      Refer to [NeuroCAPs' BIDS Structure and Entities Documentation](https://neurocaps.readthedocs.io/en/stable/bids.html)
+      for additional information.
 
-- **Grouping:** Perform CAPs analysis for entire sample or groups of subject IDs
-- **Optimal Cluster Size Identification:** Perform the Davies Bouldin, Silhouette, Elbow, or
-  Variance Ratio criterions to identify the optimal cluster size and automatically save the optimal
-  model as an attribute.
-- **Parallel Processing:** Use parallel processing to speed up optimal cluster size identification.
-- **CAPs Visualization:** Visualize the CAPs as outer products or heatmaps at either the region or
-  node level of the parcellation.
-- **Save CAPs as NifTIs:** Convert the atlas used for parcellation to a statistical NifTI image.
[…]
