### 2025-07-07 4e0d6ee
Update dependency vtk to >=9.5.0, <9.6.0
