+++ neurocaps/utils/_parcellation_validation.py
@@ -87,18 +87,18 @@ def check_keys(parcel_dict):
     return all(key in parcel_dict[list(parcel_dict)[0]] for key in required_keys)
 
 
-def process_aal(parcel_dict):
+def process_aal(config_dict):
     """
     Converts a ``parcel_approach`` containing initialization keys for "AAL" the the final processed
     version. Uses nilearn to fetch the parcellation map and labels.
     """
-    if "version" not in parcel_dict["AAL"]:
+    if "version" not in config_dict["AAL"]:
         LG.warning("'version' not specified in `parcel_approach`. Defaulting to 'SPM12'.")
-        parcel_dict["AAL"].update({"version": "SPM12"})
+        config_dict["AAL"].update({"version": "SPM12"})
 
     # Get atlas map; named "maps" to match nilearn and only contains a single file
-    fetched_aal = datasets.fetch_atlas_aal(version=parcel_dict["AAL"]["version"], verbose=0)
-    parcel_dict["AAL"].update({"maps": fetched_aal.maps})
+    fetched_aal = datasets.fetch_atlas_aal(version=config_dict["AAL"]["version"], verbose=0)
+    parcel_dict = {"AAL": {"maps": fetched_aal.maps}}
 
     # Get nodes
     parcel_dict["AAL"].update({"nodes": list(fetched_aal.labels)})
@@ -110,9 +110,18 @@ def process_aal(parcel_dict):
     regions = collapse_aal_node_names(parcel_dict["AAL"]["nodes"])
     parcel_dict["AAL"].update({"regions": regions})
 
-    # Clean initialization keys
-    for key in VALID_DICT_STUCTURES["AAL"]:
-        parcel_dict["AAL"].pop(key, None)
+    # Copy configuration and add information; no resolution information due to not wanting to
+    # load in the parcellation file and checking header/diagonal of affine matrix
+    # This is called be an __init__
+    # Added name key for consistency purposes
+    metadata_dict = dict(
+        name="AAL",
+        n_nodes=len(parcel_dict["AAL"]["nodes"]),
+        n_regions=len(parcel_dict["AAL"]["regions"]),
+        space="MNI",
+        **config_dict["AAL"],
+    )
+    parcel_dict["AAL"]["metadata"] = metadata_dict
 
     return parcel_dict
 
@@ -159,26 +168,26 @@ def process_custom(parcel_dict, call):
     check_custom_structure(parcel_dict["Custom"], custom_example)
 
 
-def process_schaefer(parcel_dict):
+def process_schaefer(config_dict):
     """
     Converts a ``parcel_approach`` containing initialization keys for "Schafer" the the final
     processed version. Uses nilearn to fetch the parcellation map and labels.
     """
-    if "n_rois" not in parcel_dict["Schaefer"]:
+    if "n_rois" not in config_dict["Schaefer"]:
         LG.warning("'n_rois' not specified in `parcel_approach`. Defaulting to 400 ROIs.")
-        parcel_dict["Schaefer"].update({"n_rois": 400})
+        config_dict["Schaefer"].update({"n_rois": 400})
 
-    if "yeo_networks" not in parcel_dict["Schaefer"]:
+    if "yeo_networks" not in config_dict["Schaefer"]:
         LG.warning("'yeo_networks' not specified in `parcel_approach`. Defaulting to 7 networks.")
-        parcel_dict["Schaefer"].update({"yeo_networks": 7})
+        config_dict["Schaefer"].update({"yeo_networks": 7})
 
-    if "resolution_mm" not in parcel_dict["Schaefer"]:
[…]
