### 2025-07-08 6ede16c
Add fetching function
