+++ neurocaps/utils/parcellations.py
@@ -1,6 +1,6 @@
 """Public utility functions for parcellations."""
 
-from typing import Literal, Optional, Union
+from typing import Any, Literal, Optional, Union
 
 import pandas as pd
 
@@ -8,8 +8,59 @@ from . import _io as io_utils
 from ._helpers import list_to_str
 from ._logging import setup_logger
 from neurocaps.typing import CustomParcelApproach
+from neurocaps.utils.datasets import _fetch as fetch_utils
 
 LG = setup_logger(__name__)
+
+
+def fetch_preset_parcel_approach(
+    name: str, n_nodes: Optional[int] = None
+) -> dict[Literal["Custom"], CustomParcelApproach]:
+    """
+    Fetches a Preset "Custom" Parcellation Approach.
+
+    Creates a directory in the user home directory named "neurocaps_data" and downloads
+    the ``parcel_approach`` (a JSON file) and the associated NifTI image from the
+    Open Science Framework (OSF) if the corresponding files are not present in the
+    directory.
+
+    .. versionadded:: 0.32.2
+
+    Parameters
+    ----------
+    name: :obj:`str`
+        The name of the preset ``parcel_approach`` to fetch. Options are "HCPex" and "4S".
+    n_nodes: :obj:`int` or :obj:`None`, default=None
+        Currently only relevant to "4S". Options for the "4S" are: 156, 256, 356,
+        456, 556, 656, 757, 956, 1056. Defaults to 456 if None.
+
+        .. note::
+           The 856 node version of "4S" is currently unavailable.
+
+    Returns
+    -------
+    dict[Literal["Custom"], CustomParcelApproach]
+        A dictionary representing the "Custom" ``parcel_approach``.
+
+    Note
+    ----
+    **Region Mapping**: The mapping of regions/networks corresponds to the indices in the "nodes"
+    list not its label ID. So, the first non-background index will be 0.
+
+    See Also
+    --------
+    CustomParcelApproach
+       The type definition for the Custom parcellation approach.
+       (See `CustomParcelApproach Documentation
+       <https://neurocaps.readthedocs.io/en/stable/api/generated/neurocaps.typing.CustomParcelApproach.html#neurocaps.typing.CustomParcelApproach>`_)
+
+    ParcelApproach
+        Type definition representing the structure of the Schaefer, AAL, and Custom parcellation
+        approaches.
+        (See `ParcelApproach Documentation
+        <https://neurocaps.readthedocs.io/en/stable/api/generated/neurocaps.typing.ParcelApproach.html#neurocaps.typing.ParcelApproach>`_)
+    """
+    return fetch_utils.fetch_custom_parcel_approach(name, n_nodes)
 
 
 def generate_custom_parcel_approach(
@@ -18,9 +69,10 @@ def generate_custom_parcel_approach(
     column_map: dict[Literal["nodes", "regions", "hemispheres"], str],
     hemisphere_map: Optional[dict[Literal["lh", "rh"], list[str]]] = None,
     background_label: Optional[str] = "Background",
+    metadata: Optional[dict[str, Any]] = None,
 ) -> dict[Literal["Custom"], CustomParcelApproach]:
     """
-    Generates the "Custom" parcellation approach dictionary from a metadata file.
+    Generate a "Custom" Parcellation Approach From a Tabular Metadata File.
 
     Constructs the nested dictionary for the "Custom" ``parcel_approach`` by reading a tabular
     file containing metadata about the parcellation.
@@ -54,7 +106,7 @@ def generate_custom_parcel_approach(
            ``column_map``. If the "regions" key is not provided, then the output custom
[…]
