+++ neurocaps/utils/datasets/_fetch.py
@@ -1,0 +1,156 @@
+"""Module containing functions related to fetching the preset parcellation approaches."""
+
+import os
+from typing import Literal, Union
+
+import numpy as np
+
+try:
+    from nilearn.datasets._utils import fetch_files as _fetch_files
+except ImportError:
+    from nilearn.datasets.utils import _fetch_files
+
+from ._presets import (
+    ATLAS_N_NODES,
+    PRESET_ATLAS_NAME,
+    PRESET_JSON_NAME,
+    PRESET_METADATA,
+    OSF_FILE_URLS,
+    VALID_PRESETS,
+)
+from neurocaps.typing import CustomParcelApproach
+from neurocaps.utils import _io as io_utils
+from neurocaps.utils._helpers import list_to_str
+from neurocaps.utils._logging import setup_logger
+
+LG = setup_logger(__name__)
+
+
+def is_valid_preset(name: str) -> None:
+    """Checks if the requested preset name is valid."""
+    if not name in VALID_PRESETS:
+        # No period
+        raise ValueError(
+            f"'{name}' is not a supported parcellation. Only the following are supported: "
+            f"{list_to_str(VALID_PRESETS)}"
+        )
+
+
+def get_data_dir() -> str:
+    """
+    Gets the full path for 'neurocaps_data' in the users home directory. If it does not
+    exist, then the directory and the subfolders are created.
+    """
+    data_dir = os.path.expanduser(os.path.join("~", "neurocaps_data"))
+    # Create manually instead of using `fetch_files` to notify user that a directory was created
+    io_utils.makedir(data_dir)
+
+    return data_dir
+
+
+def get_preset_subdir_paths(root_dir) -> tuple[str, str]:
+    """Returns the names of the subdirs."""
+    return os.path.join(root_dir, "presets", "jsons"), os.path.join(root_dir, "presets", "niftis")
+
+
+def check_n_nodes(name: str, n_nodes: Union[int, None]) -> Union[int, None]:
+    """Checks if the number of nodes for the parcellation is valid."""
+    has_node_info = name in ATLAS_N_NODES
+    if not has_node_info:
+        if n_nodes:
+            LG.warning(
+                f"`n_nodes` parameter is not valid for the {name} parcellation and will be ignored."
+            )
+        return None
+
+    if not n_nodes:
+        n_nodes = ATLAS_N_NODES[name]["default_n"]
+        LG.warning(f"Defaulting to {n_nodes} nodes for the {name} parcellation.")
+    elif n_nodes not in ATLAS_N_NODES[name]["valid_n"]:
+        if not isinstance(n_nodes, int):
+            raise ValueError("`n_nodes` must be an integer.")
+        else:
+            raise ValueError(
+                f"{n_nodes} is not valid for the {name} parcellation. "
+                f"Only the following are supported: {list_to_str(list(ATLAS_N_NODES.get(name)))}."
+            )
+
+    return n_nodes
+
+
+def get_preset_filenames(name: str, n_nodes: int = None) -> tuple[str, str]:
+    """Retrieve the preset filenames."""
+    json_filename, nifti_filename = PRESET_JSON_NAME[name], PRESET_ATLAS_NAME[name]
+    if name in ATLAS_N_NODES and n_nodes:
+        json_filename = json_filename.format(n_nodes)
+        nifti_filename = nifti_filename.format(n_nodes)
+
+    return json_filename, nifti_filename
+
+
+def get_osf_file_url(filename: str) -> str:
+    """Retrieves the url for a specific file."""
+    base_filename = os.path.basename(filename)
+
+    return r"https://osf.io/{}/download".format(OSF_FILE_URLS[base_filename])
+
+
[…]
