### 2025-07-08 0dabaaf
Add fetch function (#56)
* Add metadata key

* Add fetching function

* Fix blank plot when saving
