+++ neurocaps/utils/__init__.py
@@ -1,3 +1,3 @@
-from .parcellations import generate_custom_parcel_approach
+from .parcellations import fetch_preset_parcel_approach, generate_custom_parcel_approach
 
-__all__ = ["generate_custom_parcel_approach"]
+__all__ = ["fetch_preset_parcel_approach", "generate_custom_parcel_approach"]
