### 2025-07-08 e04906b
Fix version directive, add whitespace, doc updates [skip ci]
