+++ README.md
@@ -31,11 +31,10 @@ pip install neurocaps
 
 ```
 
-**Windows Users**
-
-To avoid installation errors related to long paths not being enabled, PyBIDS will not be installed by default.
-Refer to official [Microsoft documentation](https://learn.microsoft.com/en-us/windows/win32/fileio/maximum-file-path-limitation?tabs=powershell)
-to enable long paths.
+#### Windows Users
+PyBIDS will not be installed by default due to installation errors that may occur if long paths
+aren't enabled (Refer to official [Microsoft documentation](https://learn.microsoft.com/en-us/windows/win32/fileio/maximum-file-path-limitation?tabs=powershell)
+to enable this feature).
 
 To include PyBIDS in your installation, use:
 
@@ -73,8 +72,8 @@ pip install -e .
 git submodule update --init
 
 ```
-**Windows Users**
 
+#### Windows Users
 To include PyBIDS when installing the development version on Windows, use:
 
 ```bash
@@ -87,7 +86,6 @@ git submodule update --init
 ```
 
 ## Docker
-
 If [Docker](https://docs.docker.com/) is available on your system, you can use the NeuroCAPs Docker
 image, which includes the demos and configures a headless display for VTK.
 
@@ -113,7 +111,7 @@ docker run -it neurocaps
 docker run -it -p 9999:9999 neurocaps notebook
 ```
 
-## Usage
+## Features
 NeuroCAPs is built around two main classes (``TimeseriesExtractor`` and ``CAP``) and includes several
 features to perform the complete CAPs workflow from postprocessing to visualizations.
 Notable features includes:
@@ -153,366 +151,107 @@ Notable features includes:
 Full details for every function and parameter are available in the
 [API Documentation](https://neurocaps.readthedocs.io/en/stable/api.html).
 
-## Demonstration
+## Workflow
+The following code demonstrates a high-level workflow overview using NeuroCAPs to perform the CAPs
+analysis. An interactive variant of this workflow is available on the
+[readthedocs](https://neurocaps.readthedocs.io/en/stable/examples/workflow.html). Additional
+[tutorials]([demos](https://github.com/donishadsmith/neurocaps/tree/main/demos)) and
+[interactive demonstrations](https://github.com/donishadsmith/neurocaps/tree/main/demos) are
+also provided.
 
-Use dataset from OpenNeuro [^7]:
-```python
-# Download Sample Dataset from OpenNeuro, requires the openneuro-py package
-# [Dataset] doi: doi:10.18112/openneuro.ds005381.v1.0.0
-import os
-from openneuro import download
-
-demo_dir = "neurocaps_demo"
-os.makedirs(demo_dir)
-
-# Include the run-1 and run-2 data from two tasks for two subjects
-include = [
-    "dataset_description.json",
-    "sub-0004/ses-2/func/*run-[12]*events*",
-    "sub-0006/ses-2/func/*run-[12]*events*",
-    "derivatives/fmriprep/sub-0004/fmriprep/sub-0004/ses-2/func/*run-[12]*confounds_timeseries*",
-    "derivatives/fmriprep/sub-0004/fmriprep/sub-0004/ses-2/func/*run-[12]_space-MNI152NLin*preproc_bold*",
-    "derivatives/fmriprep/sub-0006/fmriprep/sub-0006/ses-2/func/*run-[12]*confounds_timeseries*",
[…]
