### 2025-07-09 e1c17b1
Update util for testing
