+++ tests/utils.py
@@ -422,6 +422,9 @@ def check_outputs(tmp_dir, values_dict, plot_type="map", plot_name=None):
     Check if the proper number of images are being generated by a CAP class function. Also checks
     if the naming scheme is correct.
     """
+    min_plot_byte_size = 50000
+    min_csv_byte_size = 50
+
     if plot_type == "map":
         heatmap_files = glob.glob(os.path.join(tmp_dir.name, "*heatmap*.png"))
         assert any(["nodes" in x for x in heatmap_files])
@@ -438,31 +441,34 @@ def check_outputs(tmp_dir, values_dict, plot_type="map", plot_name=None):
         [os.remove(file) for file in heatmap_files + outer_files]
     elif plot_type == "radar":
         if "html" in values_dict:
-            radar_html = glob.glob(os.path.join(tmp_dir.name, "*radar*.html"))
-            assert len(radar_html) == values_dict["html"]
-            [os.remove(file) for file in radar_html]
+            radar_htmls = glob.glob(os.path.join(tmp_dir.name, "*radar*.html"))
+            assert len(radar_htmls) == values_dict["html"]
+            assert all(os.path.getsize(x) >= min_plot_byte_size for x in radar_htmls)
+            [os.remove(file) for file in radar_htmls]
         elif "json" in values_dict:
-            radar_json = glob.glob(os.path.join(tmp_dir.name, "*radar*.json"))
-            assert len(radar_json) == values_dict["json"]
+            radar_jsons = glob.glob(os.path.join(tmp_dir.name, "*radar*.json"))
+            assert len(radar_jsons) == values_dict["json"]
 
-            with open(radar_json[0], "r") as f:
+            with open(radar_jsons[0], "r") as f:
                 fig = json.load(f)
                 fig = plotly.graph_objects.Figure(fig)
 
                 assert isinstance(fig, plotly.graph_objs._figure.Figure)
 
-            [os.remove(file) for file in radar_json]
+            [os.remove(file) for file in radar_jsons]
         else:
-            radar_png = glob.glob(os.path.join(tmp_dir.name, "*radar*.png"))
-            assert len(radar_png) == values_dict["png"]
-            [os.remove(file) for file in radar_png]
+            radar_pngs = glob.glob(os.path.join(tmp_dir.name, "*radar*.png"))
+            assert len(radar_pngs) == values_dict["png"]
+            assert all(os.path.getsize(x) >= min_plot_byte_size for x in radar_pngs)
+            [os.remove(file) for file in radar_pngs]
     elif plot_type in ["corr", "trans"]:
         name = "correlation_matrix" if plot_type == "corr" else "transition_probability"
         png_files = glob.glob(os.path.join(tmp_dir.name, f"*{name}*.png"))
         csv_files = glob.glob(os.path.join(tmp_dir.name, f"*{name}*.csv"))
         assert len(png_files) == values_dict["png"]
         assert len(csv_files) == values_dict["csv"]
-        assert os.path.getsize(csv_files[0]) > 0
+        assert all(os.path.getsize(x) >= min_plot_byte_size for x in png_files)
+        assert all(os.path.getsize(x) >= min_csv_byte_size for x in csv_files)
         [os.remove(file) for file in png_files + csv_files]
     elif plot_type == "nifti":
         nii_files = glob.glob(os.path.join(tmp_dir.name, "*.nii.gz"))
@@ -484,6 +490,7 @@ def check_outputs(tmp_dir, values_dict, plot_type="map", plot_name=None):
 
         [os.remove(file) for file in pickle_files]
     else:
-        surface_png = glob.glob(os.path.join(tmp_dir.name, "*surface*.png"))
[…]
