### 2025-07-10 c03200a
Change internal variable name
- Add warning
- Update type hint
