### 2025-07-12 498be82
Add gordon to preset
