+++ neurocaps/utils/datasets/_presets.py
@@ -1,15 +1,18 @@
 """Module containing preset constants (kind of since some information is added at runtime)."""
 
-VALID_PRESETS = ["HCPex", "4S"]
+# Consider adding to json if file gets too large
+VALID_PRESETS = ["HCPex", "4S", "Gordon"]
 
 PRESET_ATLAS_NAME = {
     "HCPex": "tpl-MNI152NLin2009cAsym_atlas-HCPex_2mm.nii.gz",
     "4S": "tpl-MNI152NLin2009cAsym_atlas-4S{}Parcels_res-01_dseg.nii.gz",
+    "Gordon": "tpl-MNI_atlas-Gordon.nii.gz",
 }
 
 PRESET_JSON_NAME = {
     "HCPex": "atlas-HCPex_desc-CustomParcelApproach.json",
     "4S": "atlas-4S{}Parcels_desc-CustomParcelApproach.json",
+    "Gordon": "atlas-Gordon_desc-CustomParcelApproach.json",
 }
 
 PRESET_METADATA = {
@@ -28,9 +31,17 @@ PRESET_METADATA = {
         "space": "MNI152NLin2009cAsym",
         "source": "https://github.com/PennLINC/AtlasPack",
     },
+    "Gordon": {
+        "name": "Gordon",
+        "n_nodes": 333,
+        "n_regions": 13,
+        "space": "MNI",
+        "doi": "10.1093/cercor/bhu239",
+        "source": "https://www.mir.wustl.edu/research/research-centers/neuroimaging/labs/egordon-lab/resources/",
+    },
 }
 
-# TODO: Check 856 and don't provide for now, may not exist on OSF based on git annex
+# TODO: Don't provide 856 for now, may not exist on OSF based on git annex
 ATLAS_N_NODES = {
     "4S": {"valid_n": [156, 256, 356, 456, 556, 656, 756, 956, 1056], "default_n": 456}
 }
@@ -56,4 +67,6 @@ OSF_FILE_URLS = {
     "atlas-4S956Parcels_desc-CustomParcelApproach.json": "9kmgb",
     "tpl-MNI152NLin2009cAsym_atlas-4S1056Parcels_res-01_dseg.nii.gz": "zsah9",
     "atlas-4S1056Parcels_desc-CustomParcelApproach.json": "a87c2",
+    "tpl-MNI_atlas-Gordon.nii.gz": "ynpcu",
+    "atlas-Gordon_desc-CustomParcelApproach.json": "dygfz",
 }
