### 2025-07-12 5b036ac
Add gordon to preset (#57)
