+++ neurocaps/analysis/cap/_internals/spatial.py
@@ -26,7 +26,7 @@ def cap_to_img(atlas_file, cap_vector, fwhm, knn_dict):
     # not 1
     target_array = np.unique(atlas_fdata)
 
-    # Start at 1 to avoid assigment to the background label
+    # Start at 1 to avoid assignment to the background label
     for indx, value in enumerate(cap_vector, start=1):
         atlas_array[atlas_fdata == target_array[indx]] = value
 
@@ -34,7 +34,7 @@ def cap_to_img(atlas_file, cap_vector, fwhm, knn_dict):
 
     # Knn implementation to aid in coverage issues
     if knn_dict:
-        stat_map = perform_knn(atlas_file, knn_dict, stat_map)
+        stat_map = perform_knn(atlas, knn_dict, stat_map)
 
     # Add smoothing to stat map to help mitigate potential coverage issues
     if fwhm is not None:
@@ -44,24 +44,22 @@ def cap_to_img(atlas_file, cap_vector, fwhm, knn_dict):
 
 
 @lru_cache(maxsize=2)
-def build_tree(atlas_file):
+def build_tree(atlas):
     """Uses scipy's ``KDTree`` to optimize k-nearest neighbors interpolation."""
-    atlas = nib.load(atlas_file)
     non_zero_indices = np.array(np.where(atlas.get_fdata() != 0)).T
     kdtree = KDTree(non_zero_indices)
 
     return kdtree, non_zero_indices
 
 
-def get_remove_indices(atlas_file, remove_labels):
+def get_remove_indices(atlas, remove_labels):
     """If requested, gets the coordinates containing a specified label to be removed."""
-    atlas = nib.load(atlas_file)
     remove_indxs = np.where(np.isin(atlas.get_fdata(), remove_labels))
 
     return remove_indxs
 
 
-def perform_knn(atlas_file, knn_dict, stat_map):
+def perform_knn(atlas, knn_dict, stat_map):
     """
     Perform KNN to assist with coverage issues prior to plotting.
 
@@ -71,11 +69,11 @@ def perform_knn(atlas_file, knn_dict, stat_map):
 
     # Get target indices
     target_indices = get_target_indices(
-        atlas_file, knn_dict["reference_atlas"], knn_dict["resolution_mm"], remove_labels
+        atlas, knn_dict["reference_atlas"], knn_dict["resolution_mm"], remove_labels
     )
 
     # Get non-zero indices; Build kdtree for nearest neighbors
-    kdtree, non_zero_indices = build_tree(atlas_file)
+    kdtree, non_zero_indices = build_tree(atlas)
 
     # Get k
     k = knn_dict["k"]
@@ -105,13 +103,11 @@ def perform_knn(atlas_file, knn_dict, stat_map):
 
 
 @lru_cache(maxsize=4)
-def get_target_indices(atlas_file, reference_atlas, resolution_mm, remove_labels):
+def get_target_indices(atlas, reference_atlas, resolution_mm, remove_labels):
     """
     Uses a reference atlas ("Schaefer" or "AAL") as a mask to identify non-background coordinates
     to use for k-nearest neighbors interpolation.
     """
-    atlas = nib.load(atlas_file)
-
     if reference_atlas == "Schaefer":
         reference_atlas_map = datasets.fetch_atlas_schaefer_2018(
             resolution_mm=resolution_mm, verbose=0
@@ -119,20 +115,10 @@ def get_target_indices(atlas_file, reference_atlas, resolution_mm, remove_labels
     else:
         reference_atlas_map = datasets.fetch_atlas_aal(verbose=0)["maps"]
 
-    # Resample schaefer to atlas file using nearest interpolation to retain labels
-    kwargs = {
[…]
