### 2025-07-12 be6942b
Get correct density
