+++ CHANGELOG.md
@@ -13,6 +13,10 @@ noted in the changelog (e.g., new functions or parameters, changes in parameter 
 improvements/enhancements. All fixes and modifications are backwards compatible.
 - *.postN* : Consists of documentation changes or metadata-related updates, such as modifications to type hints.
 
+## [0.32.4] - 2025-07-14
+### 🐛 Fixes
+- Get fslr atlas of the correct density to prevent error.
+
 ## [0.32.3] - 2025-07-14
 ### 🚀 New/Added
 - Add "Gordon" to preset
