### 2025-07-15 c865d48
Update function signatures
