+++ neurocaps/extraction/_internals/postprocess.py
@@ -151,7 +151,6 @@ def process_subject_runs(
     task_info,
     tr,
     verbose,
-    flush,
     parallel_log_config=None,
 ):
     """
@@ -167,9 +166,7 @@ def process_subject_runs(
     A default id of "run-0" is given to data where there is no run ID.
     """
     # Logger inside function to give logger to each child process if parallel processing is done
-    LG = setup_logger(
-        __name__, flush=flush, top_level=False, parallel_log_config=parallel_log_config
-    )
+    LG = setup_logger(__name__, top_level=False, parallel_log_config=parallel_log_config)
 
     # Initialize subject dictionary and quality control dictionary
     subject_timeseries = {subj_id: {}}
