### 2025-07-15 804e0fd
Update function signatures (#59)
