### 2025-07-15 182238a
Update type hint and doc string
