### 2025-07-17 ae0c172
clean code - update placeholder values
