### 2025-07-19 378124c
fix `knn_dict` default
