### 2025-07-19 1ac53fc
Make plot defaults class public
