+++ neurocaps/analysis/transition_matrix.py
@@ -4,10 +4,10 @@ from typing import Optional, Union
 
 import pandas as pd
 
+from neurocaps.utils import PlotDefaults
 from neurocaps.utils import _io as io_utils
 from neurocaps.utils._helpers import resolve_kwargs
-
-from neurocaps.utils._plotting_utils import MatrixVisualizer, PlotDefaults, PlotFuncs
+from neurocaps.utils._plot_utils import MatrixVisualizer, PlotFuncs
 
 
 def transition_matrix(
@@ -67,35 +67,10 @@ def transition_matrix(
         If True, returns a dictionary with a transition probability matrix for each group.
 
     **kwargs
-        Keyword arguments used when modifying figures. Valid keywords include:
-
-        - dpi: :obj:`int`, default=300 -- Dots per inch for the figure.
-        - figsize: :obj:`tuple`, default=(8, 6) -- Size of the figure in inches.
-        - fontsize: :obj:`int`, default=14 -- Font size for the plot title, x-axis title, and
-          y-axis title of each plot.
-        - xticklabels_size: :obj:`int`, default=8 -- Font size for x-axis tick labels.
-        - yticklabels_size: :obj:`int`, default=8 -- Font size for y-axis tick labels.
-        - shrink: :obj:`float`, default=0.8 -- Fraction by which to shrink the colorbar.
-        - cbarlabels_size: :obj:`int`, default=8 -- Font size for the colorbar labels.
-        - xlabel_rotation: :obj:`int`, default=0 -- Rotation angle for x-axis labels.
-        - ylabel_rotation: :obj:`int`, default=0 -- Rotation angle for y-axis labels.
-        - annot: :obj:`bool`, default=False -- Add values to each cell.
-        - annot_kws: :obj:`dict`, default=None, -- Customize the annotations.
-        - fmt: :obj:`str`, default=".2g" -- Modify how the annotated vales are presented.
-        - linewidths: :obj:`float`, default=0 -- Padding between each cell in the plot.
-        - borderwidths: :obj:`float`, default=0 -- Width of the border around the plot.
-        - linecolor: :obj:`str`, default="black" -- Color of the line that separates each cell.
-        - edgecolors: :obj:`str` or :obj:`None`, default=None -- Color of the edges.
-        - alpha: :obj:`float` or :obj:`None`, default=None -- Controls transparency and ranges from
-          0 (transparent) to 1 (opaque).
-        - bbox_inches: :obj:`str` or :obj:`None`, default="tight" -- Alters size of the whitespace
-          in the saved image.
-        - cmap: :obj:`str`, :obj:`callable` default="coolwarm" -- Color map for the plot cells.
-          Options include strings to call seaborn's pre-made palettes, ``seaborn.diverging_palette``
-          function to generate custom palettes, and ``matplotlib.color.LinearSegmentedColormap`` to
-          generate custom palettes.
-        - vmin: :obj:`float` or :obj:`None`, default=None -- The minimum value to display in colormap.
-        - vmax: :obj:`float` or :obj:`None`, default=None -- The maximum value to display in colormap.
+        Additional keyword arguments for customizing plots.
+        See :meth:`neurocaps.utils.PlotDefaults.transition_matrix` for all available options and
+        their default values (See `PlotDefaults Documentation for transition_matrix\
+        <https://neurocaps.readthedocs.io/en/stable/api/generated/neurocaps.utils.PlotDefaults.transition_matrix.html#neurocaps.utils.PlotDefaults.transition_matrix>`_)
 
     Returns
     -------
