+++ neurocaps/utils/__init__.py
@@ -1,3 +1,4 @@
 from .parcellations import fetch_preset_parcel_approach, generate_custom_parcel_approach
+from .plot_defaults import PlotDefaults
 
 __all__ = ["fetch_preset_parcel_approach", "generate_custom_parcel_approach"]

[…]
