### 2025-07-20 b9ab5f4
Use nbsphinx to render notebooks [skip ci]
- Easier workflow download
