+++ docs/examples/custom.rst
@@ -1,276 +1,0 @@
-Tutorial 7: Creating Custom Parcellation Approaches
-===================================================
-
-.. |colab| image:: https://colab.research.google.com/assets/colab-badge.svg
-   :target: https://colab.research.google.com/github/donishadsmith/neurocaps/blob/stable/docs/examples/notebooks/custom.ipynb
-
-|colab|
-
-While NeuroCAPs leverages Nilearn's fetch functions for the `Schaefer <https://nilearn.github.io/stable/modules/generated/nilearn.datasets.fetch_atlas_schaefer_2018.html>`_
-and `AAL <https://nilearn.github.io/stable/modules/generated/nilearn.datasets.fetch_atlas_aal.html>`_, additional
-deterministic parcellations (lateralized and non-lateralized) can be manually defined. For custom parcellation approaches, three subkeys are
-recognized: "maps", "nodes", and "regions". For additional details on these subkeys, refer to the
-`"Custom Parcellations" sub-section <https://neurocaps.readthedocs.io/en/stable/user_guide/parcellations.html#custom-parcellations>`_.
-
-**Note:** Non-lateralized parcellations are supported in versions >= 0.30.0.
-
-There are three methods to create the "Custom" `parcel_approach`.
-
-1. Manual Creation
-------------------
-
-.. code-block:: python
-
-    # Fetching atlas NiFTI image and labels from Github
-    import os, subprocess, sys
-
-    demo_dir = "neurocaps_demo"
-    os.makedirs(demo_dir, exist_ok=True)
-
-    if sys.platform != "win32":
-        cmd = [
-            [
-                "wget",
-                "-q",
-                "-P",
-                demo_dir,
-                "https://github.com/wayalan/HCPex/raw/main/HCPex_v1.1/HCPex_LookUpTable.txt",
-            ],
-            [
-                "wget",
-                "-q",
-                "-P",
-                demo_dir,
-                "https://github.com/wayalan/HCPex/raw/main/HCPex_v1.1/HCPex_2mm.nii",
-            ],
-        ]
-    else:
-        cmd = [
-            [
-                "curl",
-                "-L",
-                "-o",
-                f"{demo_dir}\\HCPex_LookUpTable.txt",
-                "https://github.com/wayalan/HCPex/raw/main/HCPex_v1.1/HCPex_LookUpTable.txt",
-            ],
-            [
-                "curl",
-                "-L",
-                "-o",
-                f"{demo_dir}\\HCPex.nii.gz",
-                "https://github.com/wayalan/HCPex/raw/main/HCPex_v1.1/HCPex_2mm.nii",
-            ],
-        ]
-
-    for command in cmd:
-        subprocess.run(command, check=True)
-
-.. code-block:: python
-
-    import pandas as pd
-
-    parcel_approach = {"Custom": {}}
-
-    # Set path to parcellation NifTI image
-    parcel_approach["Custom"]["maps"] = os.path.join(demo_dir, "HCPex.nii.gz")
-
-    # Get the nodes and ensure that the first node (index 0) is the first non-background node
-    parcel_approach["Custom"]["nodes"] = pd.read_csv(
-        os.path.join(demo_dir, "HCPex_LookUpTable.txt"),
-        sep=None,
-        engine="python",
-    )["Label"].values[1:]
-
-    # Setting the region names and their corresponding indices in the "nodes" list
-    # in this case it is just the label id - 1
-    parcel_approach["Custom"]["regions"] = {
[…]
