### 2025-07-22 cfc3676
Add new utility functions to make tutorials easier
