+++ neurocaps/analysis/standardize.py
@@ -29,7 +29,7 @@ def standardize(
     subject_timeseries_list: :obj:`list[SubjectTimeseries]` or :obj:`list[str]`
         A list where each element consist of a dictionary mapping subject IDs to their run IDs and
         associated timeseries (TRs x ROIs) as a NumPy array. Can also be a list consisting of paths
-        to pickle files containing this same structure. Refer to documentation for
+        to serialized files containing this same structure. Refer to documentation for
         ``SubjectTimeseries`` in the "See Also" section for an example structure.
 
     output_dir: :obj:`str` or :obj:`None`, default=None
