+++ neurocaps/utils/__init__.py
@@ -1,4 +1,18 @@
 from .parcellations import fetch_preset_parcel_approach, generate_custom_parcel_approach
 from .plot_defaults import PlotDefaults
+from .samples_generators import (
+    simulate_bids_dataset,
+    create_dataset_description,
+    save_dataset_description,
+    simulate_subject_timeseries,
+)
 
-__all__ = ["fetch_preset_parcel_approach", "generate_custom_parcel_approach"]
+__all__ = [
+    "fetch_preset_parcel_approach",
+    "generate_custom_parcel_approach",
+    "PlotDefaults",
+    "simulate_bids_dataset",
+    "create_dataset_description",
+    "save_dataset_description",
+    "simulate_subject_timeseries",
+]

[…]
