### 2025-07-22 e69697b
Update rst and tutorial [skip ci]
