### 2025-07-22 ec407f5
Type hint fixes
