+++ CHANGELOG.md
@@ -15,6 +15,10 @@ noted in the changelog (e.g., new functions or parameters, changes in parameter 
 improvements/enhancements. All fixes and modifications are backwards compatible.
 - *.postN* : Consists of documentation changes or metadata-related updates, such as modifications to type hints.
 
+## [0.34.2] - 2025-07-22
+### 🐛 Fixes
+- Type hint fixes and fix undocumented parameter.
+
 ## [0.34.1] - 2025-07-21
 ### 🚀 New/Added
 - Added new utility functions to make tutorials easier.
