+++ docs/introduction.rst
@@ -92,7 +92,7 @@ Notable features includes:
   - plot transition matrices
   - merge timeseries data across tasks or session [6]_
   - generate the custom parcellation dictionary structure from the parcellation's metadata file
-  - fetche preset custom parcellation approaches
+  - fetch preset custom parcellation approaches
 
 Refer to the demos to the `demos <https://github.com/donishadsmith/neurocaps/tree/main/demos>`_ or
 `tutorials <https://neurocaps.readthedocs.io/en/latest/examples/examples.html>`_ for an
