### 2025-07-26 1c59e07
Add new parameters to simulate bids function
