### 2025-07-26 568ae3e
Add new parameters to `simulate_bids_dataset`
