+++ tests/test_TimeseriesExtractor.py
@@ -2449,7 +2449,12 @@ def test_parallel_and_sequential_preprocessing_equivalence(setup_environment_3, 
 
     # Sequential
     extractor.get_bold(
-        bids_dir=bids_dir, session="002", task="rest", pipeline_name=pipeline_name, tr=1.2
+        bids_dir=bids_dir,
+        session="002",
+        task="rest",
+        pipeline_name=pipeline_name,
+        tr=1.2,
+        n_cores=None,
     )
 
     for sub in extractor.subject_timeseries:
