+++ tests/test_exceptions.py
@@ -12,17 +12,7 @@ from .utils import Parcellation
 # Activate fixture to copy dataset to temporary directory
 pytestmark = pytest.mark.usefixtures("data_dir")
 
-BIDSMSG = (
-    "No subject IDs found - potential reasons:\n"
-    "1. Incorrect template space (default: 'MNI152NLin2009cAsym'). "
-    "Fix: Set correct template space using `self.space = 'TEMPLATE_SPACE'` (e.g. 'MNI152NLin6Asym')\n"
-    "2. File names do not contain specific entities required for querying such as 'sub-', 'space-', "
-    "'task-', or 'desc-' (e.g 'sub-01_ses-1_task-rest_space-MNI152NLin2009cAsym_desc-preproc-bold.nii.gz')\n"
-    "3. Incorrect task name specified in `task` parameter.\n"
-    "4. The cache may need to be cleared using ``TimeseriesExtractor._call_layout.cache_clear()`` if the "
-    "directory has been changed (e.g. new files added, file names changed, etc) during the current Python "
-    "session."
-)
+BIDSMSG = r"No subject IDs found - potential reasons:*"
 
 ELBOWMSG = (
     f"[GROUP: All Subjects] - No elbow detected. Try adjusting the sensitivity parameter "
@@ -57,7 +47,7 @@ def test_wrong_task_with_entities(get_vars):
 
     extractor = TimeseriesExtractor(parcel_approach=Parcellation.get_custom("parcellation"))
 
-    with pytest.raises(BIDSQueryError, match=re.escape(BIDSMSG)):
+    with pytest.raises(BIDSQueryError, match=BIDSMSG):
         extractor.get_bold(bids_dir=bids_dir, task="placeholder")
 
 
@@ -67,7 +57,7 @@ def test_no_entities(get_vars, remove_task_entity):
 
     extractor = TimeseriesExtractor(parcel_approach=Parcellation.get_custom("parcellation"))
 
-    with pytest.raises(BIDSQueryError, match=re.escape(BIDSMSG)):
+    with pytest.raises(BIDSQueryError, match=BIDSMSG):
         extractor.get_bold(bids_dir=bids_dir, task="rest")
 
 
