### 2025-07-26 71cfd0e
Update type hints + minor code cleaning
