### 2025-07-26 1dbff69
Update type hints + minor code cleaning (#60)
