+++ neurocaps/extraction/_internals/bids_query.py
@@ -2,7 +2,7 @@
 
 import json, os, re
 
-from typing import Any, Union
+from typing import Any, Optional, Union
 
 from neurocaps.utils._logging import setup_logger
 
@@ -12,7 +12,7 @@ LG = setup_logger(__name__)
 def setup_extraction(
     layout: Any,
     subj_ids: list[str],
-    space: Union[str, None],
+    space: str,
     exclude_niftis: Union[list[str], None],
     signal_clean_info: dict[str, Any],
     task_info: dict[str, Any],
@@ -113,7 +113,7 @@ def query_files(
     suffix: Union[str, None] = None,
     desc: Union[str, None] = None,
     event: bool = False,
-    space: str = None,
+    space: Optional[str] = None,
 ):
     """
     Queries specific files (sorted lexicographically) using ``BidsLayout``.
@@ -183,9 +183,10 @@ def build_dict(
     return files
 
 
-def exclude_nifti_files(niftis: list[str], exclude_niftis: list[str]) -> list[str]:
+def exclude_nifti_files(niftis: list[str], exclude_niftis: Union[str, list[str]]) -> list[str]:
     """Excludes certain NIfTI files based on ``exclude_niftis``."""
     exclude_niftis = exclude_niftis if isinstance(exclude_niftis, list) else [exclude_niftis]
+
     return [nifti for nifti in niftis if os.path.basename(nifti) not in exclude_niftis]
 
 
@@ -202,7 +203,9 @@ def create_header(subj_id: str, task_info: dict[str, Any]) -> str:
 
 
 def check_files(
-    files: dict[str, list[str]], signal_clean_info: dict[str, Any], task_info: dict[str, Any]
+    files: dict[str, Union[list[str], None]],
+    signal_clean_info: dict[str, Any],
+    task_info: dict[str, Any],
 ) -> tuple[Union[bool, None], Union[str, None]]:
     """
     Simple initial check to ensure the required files are needed based on certain
