### 2025-07-27 8d00a6d
Update md, rst, and bib [skip ci]
- Update line length
- Fix workflow example
- Switch from images for tables to md table to prevent sizing issue
