+++ paper/paper.bib
@@ -184,6 +184,19 @@
     pages={763–778},
     language={en} }
 
+ @article{Gordon2016,
+    title={Generation and Evaluation of a Cortical Area Parcellation from Resting-State Correlations},
+    volume={26},
+    ISSN={1047-3211, 1460-2199},
+    DOI={10.1093/cercor/bhu239},
+    number={1}, journal={Cerebral Cortex},
+    publisher={Oxford University Press (OUP)},
+    author={Gordon, Evan M. and Laumann, Timothy O. and Adeyemo, Babatunde and Huckins, Jeremy F. and Kelley, William M. and Petersen, Steven E.},
+    year={2016},
+    month=jan,
+    pages={288–303},
+    language={en} }
+
 @software{Nilearn,
     author = {Nilearn contributors},
     license = {BSD-4-Clause},
