+++ README.md
@@ -151,17 +151,29 @@ Notable features includes:
 Full details for every function and parameter are available in the
 [API Documentation](https://neurocaps.readthedocs.io/en/stable/api.html).
 
-## Workflow
-The following code demonstrates a high-level workflow overview using NeuroCAPs to perform the CAPs
-analysis. An interactive variant of this workflow is available on the
-[readthedocs](https://neurocaps.readthedocs.io/en/stable/tutorials/tutorial-8.html). Additional
-[tutorials]([demos](https://neurocaps.readthedocs.io/en/stable/tutorials/)) and
+## Quick Start
+The following code demonstrates a high-level example using NeuroCAPs (with simulated data) to
+perform the CAPs analysis. A variant of this example using real data from [OpenNeuro](https://openneuro.org/)
+is available on the [readthedocs](https://neurocaps.readthedocs.io/en/stable/tutorials/tutorial-8.html).
+Additional [tutorials]([demos](https://neurocaps.readthedocs.io/en/stable/tutorials/)) and
 [interactive demonstrations](https://github.com/donishadsmith/neurocaps/tree/main/demos) are
-also provided.
+also available.
 
 1. Extract timeseries data
 ```python
+import numpy as np
+
 from neurocaps.extraction import TimeseriesExtractor
+from neurocaps.utils import simulate_bids_dataset
+
+# Set seed
+np.random.seed(0)
+
+# Generate a simulated BIDS directory with fMRIPrep derivatives
+# or replace with a real BIDs dataset with fMRIPrep derivatives
+bids_root = simulate_bids_dataset(
+    n_subs=3, n_runs=1, n_volumes=100, task_name="rest"
+)
 
 # Using Schaefer, one of the default parcellation approaches
 parcel_approach = {"Schaefer": {"n_rois": 100, "yeo_networks": 7}}
@@ -169,18 +181,8 @@ parcel_approach = {"Schaefer": {"n_rois": 100, "yeo_networks": 7}}
 # List of fMRIPrep-derived confounds for nuisance regression
 confound_names = [
     "cosine*",
-    "trans_x",
-    "trans_x_derivative1",
-    "trans_y",
-    "trans_y_derivative1",
-    "trans_z",
-    "trans_z_derivative1",
-    "rot_x",
-    "rot_x_derivative1",
-    "rot_y",
-    "rot_y_derivative1",
-    "rot_z",
-    "rot_z_derivative1",
+    "trans*",
+    "rot*",
     "a_comp_cor_00",
     "a_comp_cor_01",
     "a_comp_cor_02",
@@ -195,7 +197,7 @@ extractor = TimeseriesExtractor(
     confound_names=confound_names,
     standardize=False,
     fd_threshold={
-        "threshold": 0.50,
+        "threshold": 0.90,
         "outlier_percentage": 0.30,
     },
 )
@@ -205,12 +207,7 @@ extractor = TimeseriesExtractor(
 # within the BIDS root directory
 # The extracted timeseries data is automatically stored
 extractor.get_bold(
-    bids_dir="path/to/bids/root",
-    pipeline_name="fmriprep",
-    session="1",
-    task="rest",
-    tr=2,
-    verbose=False,
+    bids_dir=bids_root, task="rest", tr=2, n_cores=1, verbose=False
 )
 
 # Retrieve the dataframe containing QC information for each subject
@@ -218,6 +215,12 @@ extractor.get_bold(
 qc_df = extractor.report_qc()
 print(qc_df)
 ```
+
[…]
