### 2025-07-29 #8196 @donishadsmith
> > Hey [@robbisg](https://github.com/robbisg)! I wanted to know if you had any feedback related to improving my current submission, whether related to documentation or features. Thank you so much!
> > And thank you again [@yibeichan](https://github.com/yibeichan) for all of your feedback! I hope I was able to address everything in the 0.30.0 update. As I mentioned before, the major pressing issue was the limitation on using atlases with non-lateralized nodes such as the one from AtlasPack. While I know it wasn't mandatory, I managed to find a way to make the Schaefer 4S atlas (and additional atlases) compatible with my package in that 0.30.0 update so that NeuroCAPs could be published without this limitation. I updated the associated paper accordingly to drop the requirement that atlases must be lateralized.
>
> Hi [@donishadsmith](https://github.com/donishadsmith) , thanks for improving the documentation and the software. You also added a workflow section in your paper, however, I thought of a more descriptive paragraph, for example by adding what are the conclusions you can infer from the radar plot or what are the comments you can discuss on the transition probability, for non-expert and general audience, I'm sorry for being unclear during the revision. If you can update this part, then I think that the work sounds to me!

No problem, @robbisg! Thank you for the additional feedback to make the paper more accessible to a general audience. I updated the paper to include what inferences for the radar image in the paper and additional comments for temporal fraction and the transition matrix. Let me know if the additions I made works and thank you!
