+++ .pre-commit-config.yaml
@@ -24,7 +24,7 @@ repos:
     rev: 1.19.1
     hooks:
     -   id: blacken-docs
-        args: [--line-length=79]
+        args: [--line-length=90]
         additional_dependencies:
         - black
 

[…]
