+++ paper/paper.md
@@ -46,7 +46,7 @@ the k-means algorithm, to capture the dynamic nature of brain activity [@Liu2013
 The typical CAPs workflow can be programmatically time-consuming to manually orchestrate as it
 generally entails several steps:
 
-1. implement spatial dimensionality reduction of timeseries data using a parcellation
+1. implement spatial dimensionality reduction of timeseries data
 2. perform nuisance regression and scrub high-motion volumes (excessive head motion)
 3. concatenate the timeseries data from multiple subjects into a single matrix
 4. apply k-means clustering to the concatenated data and select the optimal number of
@@ -69,9 +69,43 @@ outputs (e.g. NiBabies [@Goncalves2025]). Furthermore, NeuroCAPs only supports t
 algorithm for clustering, which is the clustering algorithm that was originally used and is often
 employed when performing the CAPs analysis [@Liu2013].
 
+# Modules
+The core functionalities of NeuroCAPs are concentrated in three modules:
+
+1. `neurocaps.extraction` contains the `TimeseriesExtractor` class, which:
+
+- collects preprocessed BOLD data from an BIDS-compliant dataset [@Yarkoni2019]
+- leverages Nilearn's [@Nilearn] `NiftiLabelsMasker` to perform nuisance regression and spatial
+  dimensionality reduction using deterministic parcellations (e.g., Schaefer [@Schaefer2018],
+  AAL [@Tzourio-Mazoyer2002])
+- scrubs high-motion volumes using fMRIPrep-derived framewise displacement (FD) values
+- reports quality control information related to high-motion or non-steady state volumes
+
+2. `neurocaps.analysis` contains the CAP class for performing the main analysis, as well as several
+standalone utility functions.
+
+- The `CAP` class:
+  - performs k-means clustering [@scikit-learn] to identify CAPs, supporting both single and
+    optimized cluster selection with heuristics such as the silhouette and elbow method [@Arvai2023]
+  - computes subject-level temporal dynamics metrics (e.g., temporal fraction, transition
+    probabilities, etc) for statistical analysis
+  - converts identified CAPs back into NIfTI statistical maps for spatial interpretation
+  - integrates multiple plotting libraries [@Hunter:2007; @Waskom2021; @plotly; @Gale2021] to
+    provide a diverse range of visualization options
+
+- Standalone functions:
+Provide tools for data standardization [@harris2020array], merging timeseries data across sessions
+or tasks, and creating group-averaged transition matrices.
+
+3. `neurocaps.utils` contains utility functions for:
+
+- fetching preset parcellation approaches (i.e. 4S, HCPex [@Huang2022], and Gordon [@Gordon2016])
+- generating custom parcellation approaches from tabular metadata
+- customizing plots and simulating data
+
 # Usage
-The following code demonstrates basic usage of NeuroCAPs (with simulated data) to perform CAPs analysis
-. A version of this example using real data is available on
+The following code demonstrates basic usage of NeuroCAPs (with simulated data) to perform CAPs analysis.
+A version of this example using real data is available on
[…]
