+++ paper/paper.bib
@@ -119,3 +119,169 @@
     year={2025},
     month=feb
     }
+
+ @article{Yarkoni2019,
+    title={PyBIDS: Python tools for BIDS datasets},
+    volume={4},
+    rights={http://creativecommons.org/licenses/by/4.0/},
+    ISSN={2475-9066},
+    DOI={10.21105/joss.01294},
+    number={40},
+    journal={Journal of Open Source Software},
+    author={Yarkoni, Tal and Markiewicz, Christopher and De La Vega, Alejandro and Gorgolewski, Krzysztof and Salo, Taylor and Halchenko, Yaroslav and McNamara, Quinten and DeStasio, Krista and Poline, Jean-Baptiste and Petrov, Dmitry and Hayot-Sasson, Valérie and Nielson, Dylan and Carlin, Johan and Kiar, Gregory and Whitaker, Kirstie and DuPre, Elizabeth and Wagner, Adina and Tirrell, Lee and Jas, Mainak and Hanke, Michael and Poldrack, Russell and Esteban, Oscar and Appelhoff, Stefan and Holdgraf, Chris and Staden, Isla and Thirion, Bertrand and Kleinschmidt, Dave and Lee, John and Di Castello, Matteo and Notter, Michael and Blair, Ross},
+    year={2019},
+    month=aug,
+    pages={1294} }
+
+ @misc{Arvai2023,
+    title={kneed},
+    rights={Open Access},
+    url={https://zenodo.org/record/8127224},
+    DOI={10.5281/ZENODO.8127224},
+    publisher={Zenodo},
+    author={Arvai, Kevin},
+    year={2023},
+    month=jul }
+
+ @article{Schaefer2018,
+    title={Local-Global Parcellation of the Human Cerebral Cortex from Intrinsic Functional Connectivity MRI},
+    volume={28},
+    rights={https://academic.oup.com/journals/pages/open_access/funder_policies/chorus/standard_publication_model},
+    ISSN={1047-3211, 1460-2199},
+    DOI={10.1093/cercor/bhx179},
+    number={9},
+    journal={Cerebral Cortex},
+    author={Schaefer, Alexander and Kong, Ru and Gordon, Evan M and Laumann, Timothy O and Zuo, Xi-Nian and Holmes, Avram J and Eickhoff, Simon B and Yeo, B T Thomas},
+    year={2018},
+    month=sep,
+    pages={3095–3114},
+    language={en} }
+
+ @article{Tzourio-Mazoyer2002,
+    title={Automated Anatomical Labeling of Activations in SPM Using a Macroscopic Anatomical Parcellation of the MNI MRI Single-Subject Brain},
+    volume={15},
+    rights={https://www.elsevier.com/tdm/userlicense/1.0/},
+    ISSN={10538119},
+    DOI={10.1006/nimg.2001.0978},
+    number={1},
+    journal={NeuroImage},
+    author={Tzourio-Mazoyer, N. and Landeau, B. and Papathanassiou, D. and Crivello, F. and Etard, O. and Delcroix, N. and Mazoyer, B. and Joliot, M.},
+    year={2002},
+    month=jan,
+    pages={273–289},
+    language={en} }
+
+ @article{Huang2022,
+    title={An extended Human Connectome Project multimodal parcellation atlas of the human cortex and subcortical areas},
+    volume={227},
+    ISSN={1863-2653, 1863-2661},
+    DOI={10.1007/s00429-021-02421-6},
+    number={3},
+    journal={Brain Structure and Function},
+    author={Huang, Chu-Chung and Rolls, Edmund T. and Feng, Jianfeng and Lin, Ching-Po},
+    year={2022},
+    month=apr,
+    pages={763–778},
+    language={en} }
+
+ @article{Gordon2016,
[…]

[…]
