+++ pyproject.toml
@@ -65,7 +65,7 @@ dependencies = [
     ]
 
 [project.optional-dependencies]
-benchmark = ["asv", "virtualenv<20.32.1"]
+benchmark = ["asv", "virtualenv<20.33.1"]
 demo = ["ipywidgets", "openneuro-py"]
 development = ["pre-commit"]
 test = ["pytest", "pytest-cov", "pytest-rerunfailures"]
