### 2025-08-04 #8196 @donishadsmith
Hey @yibeichan , quick question regarding your review, I saw the "Automated tests" box wasn't checked, and I wanted to verify if you were able to run the tests in the repo without any issues in your environment? There have been some very specific platform and dependency quirks that I've had to work around and wanted to ensure that there wasn't additional issues that need to resolve. Thanks!
