+++ paper/paper.md
@@ -23,79 +23,71 @@ bibliography: paper.bib
 
 # Summary
 Co-Activation Patterns (CAPs) is a dynamic functional connectivity technique that clusters similar
-spatial distributions of brain activity. To make this analytical technique more accessible to
-neuroimaging researchers, NeuroCAPs, an open source Python package, was developed. This package
-performs end-to-end CAPs analyses on preprocessed resting-state or task-based functional magnetic
-resonance imaging (fMRI) data, and is most optimized for data preprocessed with fMRIPrep, a robust
-preprocessing pipeline designed to minimize manual user input and enhance reproducibility
-[@Esteban2019].
+spatial distributions of brain activity. NeuroCAPs is an open-source Python package that makes CAPs
+analysis accessible to neuroimaging researchers for preprocessed resting-state or task-based fMRI
+data. The package is optimized for fMRIPrep-preprocessed data [@Esteban2019], leveraging fMRIPrep's
+robust, BIDS-compliant outputs.
 
 # Background
 Numerous fMRI studies employ static functional connectivity (sFC) techniques to analyze correlative
-activity within and between brain regions. However, these approaches operate under the assumption
-that functional connectivity patterns, which change within seconds [@Jiang2022], remain stationary
-throughout the entire data acquisition period [@Hutchison2013].
+activity within and between brain regions. However, these approaches assume functional connectivity
+patterns, which change within seconds [@Jiang2022], remain stationary throughout the entire data
+acquisition period [@Hutchison2013].
 
 Unlike sFC approaches, dynamic functional connectivity (dFC) methods enable the analysis of dynamic
 functional states, which are characterized by consistent, replicable, and distinct periods of
 time-varying brain connectivity patterns [@Rabany2019]. Among these techniques, CAPs analysis
-aggregates similar spatial distributions of brain activity using clustering techniques, typically
-the k-means algorithm, to capture the dynamic nature of brain activity [@Liu2013; @Liu2018].
+aggregates similar spatial distributions of brain activity using a clustering algorithm
+(i.e. k-means) to capture the dynamic nature of brain activity [@Liu2013; @Liu2018].
 
 # Statement of Need
 The typical CAPs workflow can be programmatically time-consuming to manually orchestrate as it
 generally entails several steps:
 
-1. implement spatial dimensionality reduction of timeseries data
-2. perform nuisance regression and scrub high-motion volumes (excessive head motion)
-3. concatenate the timeseries data from multiple subjects into a single matrix
-4. apply k-means clustering to the concatenated data and select the optimal number of
-   clusters (CAPs) using heuristics such as the elbow or silhouette methods
-5. generate different visualizations to enhance the interpretability of the CAP
+1. apply spatial dimensionality reduction to timeseries data
+2. perform nuisance regression and remove high-motion volumes
+3. concatenate timeseries data from multiple subjects into a single matrix
+4. implement k-means clustering with optimal cluster selection
+5. generate visualizations for CAP interpretation
 
-While other excellent CAPs toolboxes exist, they are often implemented in proprietary languages such
-as MATLAB (which is the case for TbCAPs [@Bolton2020]), lack comprehensive end-to-end analytical
+While excellent CAPs toolboxes exist, they are often implemented in proprietary languages such
+as MATLAB (TbCAPs [@Bolton2020]), lack comprehensive end-to-end analytical
[…]
