### 2025-08-04 45bf377
Fix pytest command in md [skip ci]
