### 2025-08-04 4906a05
correct pytest installation on mac [skip ci] (#63)
