### 2025-08-04 #8196 @donishadsmith
@juliaferraioli thank you for feedback. I went through the paper and refined several areas to reduce verbosity. The word count is now 1022!

@yibeichan thank you for confirming that the tests ran on your Mac and for the pull request!
