+++ paper/paper.md
@@ -38,11 +38,10 @@ Unlike sFC approaches, dynamic functional connectivity (dFC) methods enable the 
 functional states, which are characterized by consistent, replicable, and distinct periods of
 time-varying brain connectivity patterns [@Rabany2019]. Among these techniques, CAPs analysis
 aggregates similar spatial distributions of brain activity using a clustering algorithm
-(i.e. k-means) to capture the dynamic nature of brain activity [@Liu2013; @Liu2018].
+(i.e., k-means) to capture the dynamic nature of brain activity [@Liu2013; @Liu2018].
 
 # Statement of Need
-The typical CAPs workflow can be programmatically time-consuming to manually orchestrate as it
-generally entails several steps:
+The CAPs workflow can be programmatically time-consuming as researchers must:
 
 1. apply spatial dimensionality reduction to timeseries data
 2. perform nuisance regression and remove high-motion volumes
@@ -86,19 +85,19 @@ The core functionalities of NeuroCAPs are concentrated in three modules:
   - integrates multiple plotting libraries [@Hunter:2007; @Waskom2021; @plotly; @Gale2021] for
     diverse visualizations
 
-- Standalone functions: provides tools for merging timeseries across sessions/tasks and creating
-  group-averaged transition matrices.
+- Standalone functions: provides tools for within-run ROI standardization [@harris2020array],
+merging timeseries across sessions/tasks and creating group-averaged transition matrices.
 
 3. `neurocaps.utils` contains utility functions for:
 
-- fetching preset parcellation approaches (i.e. 4S, HCPex [@Huang2022], and Gordon [@Gordon2016])
+- fetching non-default (i.e., Schaefer and AAL) preset parcellation approaches
+  (i.e., 4S, HCPex [@Huang2022], and Gordon [@Gordon2016])
 - generating custom parcellation approaches from tabular metadata
 - customizing plots and simulating data
 
 # Workflow
-The following code demonstrates basic usage of NeuroCAPs (with simulated data) to perform CAPs
-analysis. A version of this example using real data is available on
-[NeuroCAPs' readthedocs](https://neurocaps.readthedocs.io/en/stable/tutorials/tutorial-8.html).
+The following code demonstrates basic usage of NeuroCAPs (with simulated data) to perform CAPs analysis,
+[a version using real data is also available](https://neurocaps.readthedocs.io/en/stable/tutorials/tutorial-8.html).
 
 1. Extract timeseries data
 ```python
@@ -112,7 +111,7 @@ np.random.seed(0)
 # Generate a BIDS directory with fMRIPrep derivatives
 bids_root = simulate_bids_dataset(n_subs=3, n_runs=1, n_volumes=100, task_name="rest")
 
-# Using Schaefer, one of the default parcellation approaches
+# Set the parcel approach
 parcel_approach = {"Schaefer": {"n_rois": 100, "yeo_networks": 7}}
 
 # List of fMRIPrep-derived confounds for nuisance regression
@@ -125,11 +124,11 @@ extractor = TimeseriesExtractor(
     parcel_approach=parcel_approach,
     confound_names=confound_names,
     standardize=False,
-    # Run discarded if more than 30% of volumes exceed FD threshold
+    # Run discarded if >30% of volumes exceed FD threshold
     fd_threshold={"threshold": 0.90, "outlier_percentage": 0.30},
 )
 
-# Extract preprocessed BOLD data
+# Extract timeseries
 extractor.get_bold(bids_dir=bids_root, task="rest", tr=2, n_cores=1, verbose=False)
 
[…]
