### 2025-08-06 bf1de3c
Update api rst [skip ci]
