+++ README.md
@@ -3,14 +3,11 @@
 [![Latest Version](https://img.shields.io/pypi/v/neurocaps.svg)](https://pypi.python.org/pypi/neurocaps/)
 [![Python Versions](https://img.shields.io/pypi/pyversions/neurocaps.svg)](https://pypi.python.org/pypi/neurocaps/)
 [![DOI](https://img.shields.io/badge/DOI-10.5281%2Fzenodo.11642615-teal)](https://doi.org/10.5281/zenodo.16430050)
-[![Github Repository](https://img.shields.io/badge/Source%20Code-neurocaps-purple)](https://github.com/donishadsmith/neurocaps)
 [![Test Status](https://github.com/donishadsmith/neurocaps/actions/workflows/testing.yaml/badge.svg)](https://github.com/donishadsmith/neurocaps/actions/workflows/testing.yaml)
 [![Documentation Status](https://readthedocs.org/projects/neurocaps/badge/?version=stable)](http://neurocaps.readthedocs.io/en/stable/?badge=stable)
 [![Codecov](https://codecov.io/github/donishadsmith/neurocaps/graph/badge.svg?token=WS2V7I16WF)](https://codecov.io/github/donishadsmith/neurocaps)
-[![Code Style: Black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![License](https://img.shields.io/badge/License-MIT-blue.svg)](https://opensource.org/licenses/MIT)
-![Platform Support](https://img.shields.io/badge/OS-Ubuntu%20|%20macOS%20|%20Windows-blue)
-[![Docker](https://img.shields.io/badge/docker-donishadsmith/neurocaps-darkblue.svg?logo=docker)](https://hub.docker.com/r/donishadsmith/neurocaps/tags/)
+[![Docker](https://img.shields.io/badge/Docker-donishadsmith/neurocaps-darkblue.svg?logo=docker)](https://hub.docker.com/r/donishadsmith/neurocaps/tags/)
 [![JOSS](https://joss.theoj.org/papers/0e5c44d5d82402fa0f28e6a8833428f0/status.svg)](https://joss.theoj.org/papers/0e5c44d5d82402fa0f28e6a8833428f0)
 
 NeuroCAPs (**Neuro**imaging **C**o-**A**ctivation **P**attern**s**) is a Python package for performing Co-Activation
@@ -63,41 +60,15 @@ docker run -it -p 9999:9999 neurocaps notebook
 ```
 
 ## Features
-NeuroCAPs is built around two main classes (``TimeseriesExtractor`` and ``CAP``) and includes several
+NeuroCAPs is built around two main classes (`TimeseriesExtractor` and `CAP`) and includes several
 features to perform the complete CAPs workflow from postprocessing to visualizations.
 Notable features includes:
 
-- Timeseries Extraction (``TimeseriesExtractor``):
-    - extracts BOLD timeseries from resting-state or task-based fMRI data
-    - supports deterministic parcellations such as the Schaefer and AAL, in addition to custom-defined deterministic parcellations
-    - performs nuisance regression, motion scrubbing, and additional features
-    - reports quality control information based on framewise displacement
-
-    **Important**:
-       NeuroCAPs is most optimized for fMRI data preprocessed with
-       [fMRIPrep](https://fmriprep.org/en/stable/) and assumes the data is BIDs compliant.
-       Refer to [NeuroCAPs' BIDS Structure and Entities Documentation](https://neurocaps.readthedocs.io/en/stable/bids.html)
-       for additional information.
-
-- CAPs Analysis (``CAP``):
-    - performs k-means clustering on individuals or groups
[…]
