+++ docs/user_guide/user_guide.rst
@@ -6,8 +6,8 @@ dimensionality reduction, logging configurations for sequential and parallel pro
 of NeuroCAPs.
 
 .. toctree::
-   user_guide/installation
-   user_guide/bids
-   user_guide/parcellations
-   user_guide/logging
-   user_guide/outputs
+   installation
+   bids
+   parcellations
+   logging
+   outputs
