### 2025-08-06 cabf974
Fix colab link and add source link [skip ci]
