### 2025-08-07 a768881
account for case when impulse event
- Also change arg position in counts for segment no effect to computation since equality is the same in segments anyway and test confirms
  the temporal fraction = persistence * counts computation. Fix for
  readability

- add some unit tests
