+++ CHANGELOG.md
@@ -15,6 +15,11 @@ noted in the changelog (e.g., new functions or parameters, changes in parameter 
 improvements/enhancements. All fixes and modifications are backwards compatible.
 - *.postN* : Consists of documentation changes or metadata-related updates, such as modifications to type hints.
 
+## [0.35.0] - Unreleased
+### 🐛 Fixes
+- Accounts for case when duration is coded as 0 in event timing files as 0 in BIDS can indicate
+impulse.
+
 ## [0.34.3] - 2025-07-25
 ### 🚀 New/Added
 - Added `n_cores` and `progress_bar` parameters to `simulate_bids_dataset`.
