### 2025-08-08 e6d90ac
Remove redundant badge [skip ci]
