### 2025-08-08 b1d4381
Switch to deepcopy from .copy for dict and list to not bypass validation
- Add missing test assert
