### 2025-08-08 7e76b46
Increment version
- Remove typing from test since it is likely brittle and just check if
  values are positive. Windows minimum numpy typing issue
