### 2025-08-08 #8196 @donishadsmith
@juliaferraioli Thank you for the information!

Here is the Zenodo archive: https://zenodo.org/records/16782539
Tagged version: 0.35.0
