+++ CHANGELOG.md
@@ -15,6 +15,11 @@ noted in the changelog (e.g., new functions or parameters, changes in parameter 
 improvements/enhancements. All fixes and modifications are backwards compatible.
 - *.postN* : Consists of documentation changes or metadata-related updates, such as modifications to type hints.
 
+## [0.35.1] - 2025-08-08
+### 🐛 Fixes
+- Properly set data type in nifti image header to address floating point issues for certain atlases
+when converting CAP vectors to niftis
+
 ## [0.35.0] - 2025-08-08
 - JOSS Version
 ### 🐛 Fixes
