### 2025-08-11 96d9d27
Update dependency virtualenv to <20.33.2 (#64)
* Update doc [skip ci]

* Update dependency virtualenv to <20.33.2

---------

Co-authored-by: Donisha Smith <112973674+donishadsmith@users.noreply.github.com>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
