+++ pyproject.toml
@@ -66,7 +66,7 @@ dependencies = [
 
 [project.optional-dependencies]
 all = ["neurocaps[benchmark, demo, development, test, windows]"]
-benchmark = ["asv", "virtualenv<20.33.1"]
+benchmark = ["asv", "virtualenv<20.33.2"]
 demo = ["ipywidgets", "openneuro-py"]
 development = ["pre-commit"]
 test = ["pytest", "pytest-cov", "pytest-rerunfailures"]
