### 2025-08-13 #8196 @juliaferraioli
@donishadsmith looks like there's an issue with the URL on line 129: it resolves to https://plotly.com/python/radar-chart/;/. For that citation, I would check the [bibtex](https://github.com/donishadsmith/neurocaps/blob/ec0cdb43b2cd4fcf18e564c0a08b4cdaf4db6a65/paper/paper.bib#L248) to ensure you're citing it properly. Are you citing a specific chart within the page, or the page itself?
