+++ paper/paper.bib
@@ -246,21 +246,22 @@
  }
 
 @online{plotly,
-    author = {Plotly Technologies Inc.},
-    title = {Chart title},
-    publisher = {Plotly Technologies Inc.},
-    howpublished = {https://plotly.com/python/radar-chart/}
-}
+author = {Plotly Technologies Inc.},
+title = {Collaborative data science},
+publisher = {Plotly Technologies Inc.},
+address = {Montreal, QC},
+year = {2015},
+url = {https://plot.ly} }
 
- @misc{Gale2021,
-    title={Surfplot: Publication-ready brain surface figures},
-    rights={Open Access},
-    url={https://zenodo.org/record/5567926},
-    DOI={10.5281/ZENODO.5567926},
-    publisher={Zenodo},
-    author={Gale, Daniel J. and Vos de Wael., Reinder and Benkarim, Oualid and Bernhardt, Boris},
-    year={2021},
-    month=oct
+@misc{Gale2021,
+title={Surfplot: Publication-ready brain surface figures},
+rights={Open Access},
+url={https://zenodo.org/record/5567926},
+DOI={10.5281/ZENODO.5567926},
+publisher={Zenodo},
+author={Gale, Daniel J. and Vos de Wael., Reinder and Benkarim, Oualid and Bernhardt, Boris},
+year={2021},
+month=oct
 }
 
 @Article{harris2020array,
