### 2025-08-14 #8196 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
